Select sampling clusters using systematic sampling

Select sampling clusters using systematic sampling

Usage

get_sampling_clusters(N_clusters, n_clusters)

get_sampling_list(cluster_list, n_clusters)

Arguments

N_clusters

Total number of clusters in survey area.

n_clusters

Number of sampling clusters to be selected.

cluster_list

A data.frame containing at least the name or any other identifier for the entire set of clusters to sample from.

Value

An integer vector for get_sampling_clusters() giving the row index for selected clusters. A data.frame for [get_sampling_list()] which is a subset of cluster_list.

Examples

get_sampling_clusters(N_clusters = 211, n_clusters = 35)
#>  [1]   5  11  17  23  29  35  41  47  53  59  65  71  77  83  89  95 101 107 113
#> [20] 119 125 131 137 143 149 155 161 167 173 179 185 191 197 203 209
get_sampling_list(cluster_list = village_list, n_clusters = 70)
#> # A tibble: 72 × 4
#>       id chiefdom section    village  
#>    <dbl> <chr>    <chr>      <chr>    
#>  1     7 Badjia   Fallay     Pelewahun
#>  2    21 Badjia   Njargbahun Yorvuhun 
#>  3    35 Bagbe    Jongo      Mokpendeh
#>  4    49 Bagbe    Kemoh      Niahun   
#>  5    63 Bagbe    Nyallay    Fuinda   
#>  6    77 Bagbo    Bum        Tokpombu 
#>  7    91 Bagbo    Jimmi      Garinga  
#>  8   105 Bagbo    Jimmi      Jimm     
#>  9   119 Bagbo    Mano       Tokpombu 
#> 10   133 Bagbo    Tissawa    Ngolawoma
#> # ℹ 62 more rows