Select sampling clusters using systematic sampling

Select sampling clusters using systematic sampling

Usage

get_sampling_interval(N_clusters, n_clusters)

select_random_start(interval)

select_sampling_clusters(N_clusters, n_clusters)

create_sampling_list(cluster_list, n_clusters)

Arguments

N_clusters

Total number of clusters in survey area

n_clusters

Number of sampling clusters to be selected

interval

Sampling interval usually calculated using get_sampling_interval()

cluster_list

A data.frame containing at least the name or any other identifier for the entire set of clusters to sample from.

Value

A numeric value for get_sampling_interval() and select_random_start(); An integer vector for select_sampling_clusters() giving the row index for selected clusters; A data.frame for create_sampling_list() which is a subset of cluster_list

Examples

get_sampling_interval(N_clusters = 211, n_clusters = 35)
#> [1] 6
interval <- get_sampling_interval(N_clusters = 211, n_clusters = 35)
select_random_start(interval)
#> [1] 1
select_sampling_clusters(N_clusters = 211, n_clusters = 35)
#>  [1]   2   8  14  20  26  32  38  44  50  56  62  68  74  80  86  92  98 104 110
#> [20] 116 122 128 134 140 146 152 158 164 170 176 182 188 194 200 206
create_sampling_list(cluster_list = village_list, n_clusters = 70)
#> # A tibble: 72 × 4
#>       id chiefdom section    village   
#>    <dbl> <chr>    <chr>      <chr>     
#>  1     1 Badjia   Damia      Ngelehun  
#>  2    15 Badjia   Kpallay    Gumahun   
#>  3    29 Badjia   Sei        Korjehun  
#>  4    43 Bagbe    Jongo      Kondiama  
#>  5    57 Bagbe    Niawa      Bonkema   
#>  6    71 Bagbe    Samawa     Palima    
#>  7    85 Bagbo    Gorapon    Mohapewa  
#>  8    99 Bagbo    Jimmi      Bandajuma 
#>  9   113 Bagbo    Kpangbalia Kpangbalia
#> 10   127 Bagbo    Tissawa    Sembehun  
#> # ℹ 62 more rows