Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    1       1          1
#> 12    0       1          1
#> 13    0       1          1
#> 14    0       1          1
#> 15    1       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    1       1          1
#> 20    0       1          1
#> 21    1       1          2
#> 22    1       1          2
#> 23    0       1          2
#> 24    0       1          2
#> 25    1       1          2
#> 26    0       1          2
#> 27    0       1          2
#> 28    2       1          2
#> 29    2       1          2
#> 30    4       1          2
#> 31    1       1          3
#> 32    4       1          3
#> 33    0       1          3
#> 34    1       1          3
#> 35    2       1          3
#> 36    0       1          3
#> 37    2       1          3
#> 38    1       1          3
#> 39    1       1          3
#> 40    0       1          3
#> 41    0       1          4
#> 42    1       1          4
#> 43    2       1          4
#> 44    3       1          4
#> 45    1       1          4
#> 46    3       1          4
#> 47    4       1          4
#> 48    2       1          4
#> 49    1       1          4
#> 50    3       1          4
#> 51    1       1          5
#> 52    0       1          5
#> 53    3       1          5
#> 54    2       1          5
#> 55    4       1          5
#> 56    4       1          5
#> 57    2       1          5
#> 58    2       1          5
#> 59    2       1          5
#> 60    3       1          5
#> 61    0       1          6
#> 62    3       1          6
#> 63    3       1          6
#> 64    1       1          6
#> 65    4       1          6
#> 66    3       1          6
#> 67    0       1          6
#> 68    7       1          6
#> 69    2       1          6
#> 70    5       1          6
#> 71    5       1          7
#> 72    3       1          7
#> 73    4       1          7
#> 74    1       1          7
#> 75    2       1          7
#> 76    4       1          7
#> 77    6       1          7
#> 78    2       1          7
#> 79    5       1          7
#> 80    1       1          7
#> 81    1       1          8
#> 82    3       1          8
#> 83    3       1          8
#> 84    4       1          8
#> 85    4       1          8
#> 86    4       1          8
#> 87    0       1          8
#> 88    5       1          8
#> 89    1       1          8
#> 90    2       1          8
#> 91    3       1          9
#> 92    7       1          9
#> 93    4       1          9
#> 94    5       1          9
#> 95    5       1          9
#> 96    4       1          9
#> 97    1       1          9
#> 98    4       1          9
#> 99    2       1          9
#> 100   3       1          9
#> 101   5       1         10
#> 102   3       1         10
#> 103   7       1         10
#> 104   3       1         10
#> 105   1       1         10
#> 106   5       1         10
#> 107   5       1         10
#> 108   2       1         10
#> 109   4       1         10
#> 110   5       1         10
#> 111   2       1         11
#> 112   3       1         11
#> 113   3       1         11
#> 114   6       1         11
#> 115   7       1         11
#> 116   3       1         11
#> 117   7       1         11
#> 118   5       1         11
#> 119   4       1         11
#> 120   6       1         11
#> 121   8       1         12
#> 122   7       1         12
#> 123   5       1         12
#> 124   7       1         12
#> 125   6       1         12
#> 126   5       1         12
#> 127   2       1         12
#> 128   2       1         12
#> 129   4       1         12
#> 130   4       1         12
#> 131   4       1         13
#> 132   1       1         13
#> 133   1       1         13
#> 134   5       1         13
#> 135   4       1         13
#> 136   4       1         13
#> 137   1       1         13
#> 138  10       1         13
#> 139   7       1         13
#> 140   4       1         13
#> 141   5       1         14
#> 142   6       1         14
#> 143   5       1         14
#> 144   4       1         14
#> 145   6       1         14
#> 146   9       1         14
#> 147   7       1         14
#> 148   2       1         14
#> 149   6       1         14
#> 150   4       1         14
#> 151   5       1         15
#> 152   5       1         15
#> 153   6       1         15
#> 154   5       1         15
#> 155   5       1         15
#> 156   6       1         15
#> 157   6       1         15
#> 158   8       1         15
#> 159   5       1         15
#> 160   8       1         15
#> 161   3       1         16
#> 162   7       1         16
#> 163   7       1         16
#> 164   7       1         16
#> 165   7       1         16
#> 166  11       1         16
#> 167   4       1         16
#> 168   8       1         16
#> 169   4       1         16
#> 170   8       1         16
#> 171   6       1         17
#> 172   5       1         17
#> 173   6       1         17
#> 174   3       1         17
#> 175   6       1         17
#> 176   6       1         17
#> 177   6       1         17
#> 178  10       1         17
#> 179   9       1         17
#> 180   8       1         17
#> 181   5       1         18
#> 182  11       1         18
#> 183   6       1         18
#> 184   7       1         18
#> 185  10       1         18
#> 186  12       1         18
#> 187   7       1         18
#> 188   8       1         18
#> 189   4       1         18
#> 190   7       1         18
#> 191   4       1         19
#> 192   7       1         19
#> 193   7       1         19
#> 194   5       1         19
#> 195   5       1         19
#> 196   3       1         19
#> 197  11       1         19
#> 198   3       1         19
#> 199   7       1         19
#> 200   8       1         19
#> 201   8       1         20
#> 202  11       1         20
#> 203   8       1         20
#> 204   9       1         20
#> 205   9       1         20
#> 206   9       1         20
#> 207  11       1         20
#> 208   8       1         20
#> 209   5       1         20
#> 210   4       1         20
#> 211  11       1         21
#> 212   9       1         21
#> 213   8       1         21
#> 214   6       1         21
#> 215   6       1         21
#> 216   5       1         21
#> 217   9       1         21
#> 218   6       1         21
#> 219   6       1         21
#> 220   8       1         21
#> 221  12       1         22
#> 222   8       1         22
#> 223   5       1         22
#> 224  10       1         22
#> 225  10       1         22
#> 226   8       1         22
#> 227   7       1         22
#> 228  10       1         22
#> 229  10       1         22
#> 230  10       1         22
#> 231   9       1         23
#> 232   6       1         23
#> 233  12       1         23
#> 234  10       1         23
#> 235  10       1         23
#> 236   7       1         23
#> 237   7       1         23
#> 238   9       1         23
#> 239   5       1         23
#> 240  12       1         23
#> 241   9       1         24
#> 242  10       1         24
#> 243  10       1         24
#> 244   9       1         24
#> 245   7       1         24
#> 246   3       1         24
#> 247  12       1         24
#> 248   6       1         24
#> 249  10       1         24
#> 250  14       1         24
#> 251   6       1         25
#> 252  12       1         25
#> 253   5       1         25
#> 254  10       1         25
#> 255  11       1         25
#> 256  10       1         25
#> 257  14       1         25
#> 258   8       1         25
#> 259   6       1         25
#> 260   5       1         25
#> 261  12       1         26
#> 262   9       1         26
#> 263   8       1         26
#> 264  11       1         26
#> 265  11       1         26
#> 266   7       1         26
#> 267  10       1         26
#> 268  13       1         26
#> 269  11       1         26
#> 270   8       1         26
#> 271  17       1         27
#> 272  10       1         27
#> 273  11       1         27
#> 274   9       1         27
#> 275  12       1         27
#> 276   8       1         27
#> 277   6       1         27
#> 278  11       1         27
#> 279  11       1         27
#> 280   8       1         27
#> 281   5       1         28
#> 282  13       1         28
#> 283  14       1         28
#> 284   9       1         28
#> 285  11       1         28
#> 286  16       1         28
#> 287  11       1         28
#> 288  14       1         28
#> 289  14       1         28
#> 290  12       1         28
#> 291  10       1         29
#> 292  15       1         29
#> 293  13       1         29
#> 294  10       1         29
#> 295   9       1         29
#> 296   6       1         29
#> 297  13       1         29
#> 298  10       1         29
#> 299  15       1         29
#> 300   8       1         29
#> 301   8       1         30
#> 302   8       1         30
#> 303  10       1         30
#> 304  13       1         30
#> 305   6       1         30
#> 306  17       1         30
#> 307  15       1         30
#> 308   9       1         30
#> 309  12       1         30
#> 310  17       1         30
#> 311  17       1         31
#> 312  14       1         31
#> 313  13       1         31
#> 314  11       1         31
#> 315  13       1         31
#> 316  12       1         31
#> 317  13       1         31
#> 318  11       1         31
#> 319  11       1         31
#> 320  12       1         31
#> 321  13       1         32
#> 322  10       1         32
#> 323  10       1         32
#> 324  11       1         32
#> 325  19       1         32
#> 326  11       1         32
#> 327  13       1         32
#> 328   9       1         32
#> 329  11       1         32
#> 330  12       1         32
#> 331  11       1         33
#> 332  15       1         33
#> 333   9       1         33
#> 334  15       1         33
#> 335  16       1         33
#> 336  13       1         33
#> 337  15       1         33
#> 338  14       1         33
#> 339  14       1         33
#> 340  13       1         33
#> 341  18       1         34
#> 342  11       1         34
#> 343  12       1         34
#> 344  15       1         34
#> 345  12       1         34
#> 346  15       1         34
#> 347  15       1         34
#> 348  10       1         34
#> 349  13       1         34
#> 350   8       1         34
#> 351  15       1         35
#> 352  11       1         35
#> 353  16       1         35
#> 354  11       1         35
#> 355  18       1         35
#> 356  15       1         35
#> 357  10       1         35
#> 358  21       1         35
#> 359  17       1         35
#> 360  19       1         35
#> 361  14       1         36
#> 362  13       1         36
#> 363  15       1         36
#> 364  17       1         36
#> 365  11       1         36
#> 366  20       1         36
#> 367  16       1         36
#> 368  15       1         36
#> 369  17       1         36
#> 370  13       1         36
#> 371  17       1         37
#> 372  14       1         37
#> 373  10       1         37
#> 374  13       1         37
#> 375   9       1         37
#> 376  16       1         37
#> 377  11       1         37
#> 378  15       1         37
#> 379  20       1         37
#> 380  13       1         37
#> 381   9       1         38
#> 382  14       1         38
#> 383  13       1         38
#> 384  16       1         38
#> 385  13       1         38
#> 386  14       1         38
#> 387  16       1         38
#> 388  17       1         38
#> 389  13       1         38
#> 390  13       1         38
#> 391  20       1         39
#> 392  12       1         39
#> 393  20       1         39
#> 394  15       1         39
#> 395  16       1         39
#> 396  11       1         39
#> 397  16       1         39
#> 398  14       1         39
#> 399  22       1         39
#> 400  15       1         39
#> 401  14       1         40
#> 402  13       1         40
#> 403  22       1         40
#> 404  16       1         40
#> 405  16       1         40
#> 406  18       1         40
#> 407  14       1         40
#> 408  14       1         40
#> 409  22       1         40
#> 410  13       1         40
#> 411  16       1         41
#> 412  20       1         41
#> 413  23       1         41
#> 414  16       1         41
#> 415  16       1         41
#> 416  20       1         41
#> 417  13       1         41
#> 418  19       1         41
#> 419  21       1         41
#> 420  16       1         41
#> 421  10       1         42
#> 422  13       1         42
#> 423  19       1         42
#> 424  18       1         42
#> 425  20       1         42
#> 426  24       1         42
#> 427  21       1         42
#> 428  16       1         42
#> 429  19       1         42
#> 430  10       1         42
#> 431  19       1         43
#> 432  15       1         43
#> 433  18       1         43
#> 434  13       1         43
#> 435  16       1         43
#> 436  17       1         43
#> 437  14       1         43
#> 438  14       1         43
#> 439  18       1         43
#> 440  13       1         43
#> 441  20       1         44
#> 442  14       1         44
#> 443  18       1         44
#> 444  11       1         44
#> 445  19       1         44
#> 446  17       1         44
#> 447  18       1         44
#> 448  18       1         44
#> 449  16       1         44
#> 450  20       1         44
#> 451  19       1         45
#> 452  22       1         45
#> 453  18       1         45
#> 454  18       1         45
#> 455  16       1         45
#> 456  17       1         45
#> 457  20       1         45
#> 458  21       1         45
#> 459  19       1         45
#> 460  16       1         45
#> 461  17       1         46
#> 462  14       1         46
#> 463  21       1         46
#> 464  18       1         46
#> 465  19       1         46
#> 466  22       1         46
#> 467  15       1         46
#> 468  20       1         46
#> 469  18       1         46
#> 470  20       1         46
#> 471  13       1         47
#> 472  20       1         47
#> 473  21       1         47
#> 474  18       1         47
#> 475  14       1         47
#> 476  24       1         47
#> 477  23       1         47
#> 478  14       1         47
#> 479  15       1         47
#> 480  23       1         47
#> 481  23       1         48
#> 482  20       1         48
#> 483  16       1         48
#> 484  14       1         48
#> 485  13       1         48
#> 486  21       1         48
#> 487  24       1         48
#> 488  22       1         48
#> 489  14       1         48
#> 490  20       1         48
#> 491  23       1         49
#> 492  20       1         49
#> 493  23       1         49
#> 494  18       1         49
#> 495  16       1         49
#> 496  22       1         49
#> 497  19       1         49
#> 498  23       1         49
#> 499  20       1         49
#> 500  22       1         49
#> 501  15       1         50
#> 502  17       1         50
#> 503  19       1         50
#> 504  18       1         50
#> 505  21       1         50
#> 506  23       1         50
#> 507  21       1         50
#> 508  17       1         50
#> 509  25       2         50
#> 510  15       1         50
#> 511  25       2         51
#> 512  17       1         51
#> 513  19       1         51
#> 514  25       2         51
#> 515  19       1         51
#> 516  22       1         51
#> 517  20       1         51
#> 518  13       1         51
#> 519  17       1         51
#> 520  16       1         51
#> 521  17       1         52
#> 522  24       1         52
#> 523  19       1         52
#> 524  23       1         52
#> 525  22       1         52
#> 526  18       1         52
#> 527  21       1         52
#> 528  22       1         52
#> 529  22       1         52
#> 530  18       1         52
#> 531  23       1         53
#> 532  20       1         53
#> 533  23       1         53
#> 534  25       2         53
#> 535  25       2         53
#> 536  25       2         53
#> 537  21       1         53
#> 538  21       1         53
#> 539  21       1         53
#> 540  21       1         53
#> 541  22       1         54
#> 542  22       1         54
#> 543  24       1         54
#> 544  22       1         54
#> 545  18       1         54
#> 546  15       1         54
#> 547  23       1         54
#> 548  20       1         54
#> 549  26       2         54
#> 550  16       1         54
#> 551  21       1         55
#> 552  22       1         55
#> 553  24       1         55
#> 554  27       2         55
#> 555  19       1         55
#> 556  25       2         55
#> 557  22       1         55
#> 558  22       1         55
#> 559  21       1         55
#> 560  19       1         55
#> 561  23       1         56
#> 562  22       1         56
#> 563  29       2         56
#> 564  23       1         56
#> 565  24       1         56
#> 566  19       1         56
#> 567  34       2         56
#> 568  17       1         56
#> 569  25       2         56
#> 570  20       1         56
#> 571  25       2         57
#> 572  21       1         57
#> 573  18       1         57
#> 574  18       1         57
#> 575  31       2         57
#> 576  19       1         57
#> 577  21       1         57
#> 578  27       2         57
#> 579  20       1         57
#> 580  19       1         57
#> 581  22       1         58
#> 582  26       2         58
#> 583  24       1         58
#> 584  27       2         58
#> 585  25       2         58
#> 586  27       2         58
#> 587  24       1         58
#> 588  24       1         58
#> 589  25       2         58
#> 590  22       1         58
#> 591  18       1         59
#> 592  20       1         59
#> 593  19       1         59
#> 594  22       1         59
#> 595  27       2         59
#> 596  23       1         59
#> 597  27       2         59
#> 598  26       2         59
#> 599  19       1         59
#> 600  20       1         59
#> 601  27       2         60
#> 602  24       1         60
#> 603  25       2         60
#> 604  28       2         60
#> 605  20       1         60
#> 606  16       1         60
#> 607  23       1         60
#> 608  26       2         60
#> 609  28       2         60
#> 610  21       1         60
#> 611  30       2         61
#> 612  20       1         61
#> 613  24       1         61
#> 614  26       2         61
#> 615  21       1         61
#> 616  24       1         61
#> 617  19       1         61
#> 618  26       2         61
#> 619  22       1         61
#> 620  26       2         61
#> 621  25       2         62
#> 622  23       1         62
#> 623  26       2         62
#> 624  19       1         62
#> 625  19       1         62
#> 626  27       2         62
#> 627  25       2         62
#> 628  23       1         62
#> 629  20       1         62
#> 630  21       1         62
#> 631  26       2         63
#> 632  32       2         63
#> 633  28       2         63
#> 634  30       2         63
#> 635  28       2         63
#> 636  26       2         63
#> 637  28       2         63
#> 638  23       1         63
#> 639  30       2         63
#> 640  28       2         63
#> 641  28       2         64
#> 642  26       2         64
#> 643  19       1         64
#> 644  27       2         64
#> 645  25       2         64
#> 646  26       2         64
#> 647  25       2         64
#> 648  29       2         64
#> 649  23       1         64
#> 650  29       2         64
#> 651  25       2         65
#> 652  24       1         65
#> 653  29       2         65
#> 654  26       2         65
#> 655  26       2         65
#> 656  28       2         65
#> 657  27       2         65
#> 658  25       2         65
#> 659  22       1         65
#> 660  23       1         65
#> 661  28       2         66
#> 662  27       2         66
#> 663  30       2         66
#> 664  23       1         66
#> 665  22       1         66
#> 666  30       2         66
#> 667  19       1         66
#> 668  28       2         66
#> 669  26       2         66
#> 670  31       2         66
#> 671  20       1         67
#> 672  31       2         67
#> 673  29       2         67
#> 674  27       2         67
#> 675  28       2         67
#> 676  31       2         67
#> 677  23       1         67
#> 678  27       2         67
#> 679  28       2         67
#> 680  30       2         67
#> 681  32       2         68
#> 682  28       2         68
#> 683  28       2         68
#> 684  26       2         68
#> 685  27       2         68
#> 686  27       2         68
#> 687  30       2         68
#> 688  26       2         68
#> 689  28       2         68
#> 690  29       2         68
#> 691  27       2         69
#> 692  32       2         69
#> 693  28       2         69
#> 694  31       2         69
#> 695  29       2         69
#> 696  26       2         69
#> 697  27       2         69
#> 698  26       2         69
#> 699  32       2         69
#> 700  32       2         69
#> 701  28       2         70
#> 702  28       2         70
#> 703  32       2         70
#> 704  31       2         70
#> 705  31       2         70
#> 706  24       1         70
#> 707  24       1         70
#> 708  31       2         70
#> 709  28       2         70
#> 710  27       2         70
#> 711  34       2         71
#> 712  30       2         71
#> 713  27       2         71
#> 714  35       2         71
#> 715  27       2         71
#> 716  26       2         71
#> 717  27       2         71
#> 718  29       2         71
#> 719  32       2         71
#> 720  31       2         71
#> 721  29       2         72
#> 722  28       2         72
#> 723  21       1         72
#> 724  32       2         72
#> 725  26       2         72
#> 726  30       2         72
#> 727  22       1         72
#> 728  31       2         72
#> 729  26       2         72
#> 730  28       2         72
#> 731  26       2         73
#> 732  25       2         73
#> 733  31       2         73
#> 734  31       2         73
#> 735  26       2         73
#> 736  32       2         73
#> 737  30       2         73
#> 738  30       2         73
#> 739  33       2         73
#> 740  32       2         73
#> 741  30       2         74
#> 742  28       2         74
#> 743  27       2         74
#> 744  29       2         74
#> 745  29       2         74
#> 746  32       2         74
#> 747  27       2         74
#> 748  32       2         74
#> 749  26       2         74
#> 750  27       2         74
#> 751  26       2         75
#> 752  31       2         75
#> 753  28       2         75
#> 754  28       2         75
#> 755  29       2         75
#> 756  31       2         75
#> 757  29       2         75
#> 758  31       2         75
#> 759  28       2         75
#> 760  26       2         75
#> 761  31       2         76
#> 762  30       2         76
#> 763  30       2         76
#> 764  35       2         76
#> 765  33       2         76
#> 766  27       2         76
#> 767  30       2         76
#> 768  27       2         76
#> 769  33       2         76
#> 770  35       2         76
#> 771  39       3         77
#> 772  31       2         77
#> 773  25       2         77
#> 774  30       2         77
#> 775  26       2         77
#> 776  30       2         77
#> 777  33       2         77
#> 778  24       1         77
#> 779  31       2         77
#> 780  35       2         77
#> 781  33       2         78
#> 782  34       2         78
#> 783  34       2         78
#> 784  32       2         78
#> 785  31       2         78
#> 786  29       2         78
#> 787  31       2         78
#> 788  34       2         78
#> 789  29       2         78
#> 790  34       2         78
#> 791  30       2         79
#> 792  32       2         79
#> 793  32       2         79
#> 794  36       2         79
#> 795  30       2         79
#> 796  31       2         79
#> 797  28       2         79
#> 798  30       2         79
#> 799  30       2         79
#> 800  31       2         79
#> 801  31       2         80
#> 802  33       2         80
#> 803  35       2         80
#> 804  32       2         80
#> 805  37       3         80
#> 806  30       2         80
#> 807  28       2         80
#> 808  32       2         80
#> 809  32       2         80
#> 810  32       2         80
#> 811  35       2         81
#> 812  34       2         81
#> 813  33       2         81
#> 814  35       2         81
#> 815  35       2         81
#> 816  31       2         81
#> 817  33       2         81
#> 818  35       2         81
#> 819  34       2         81
#> 820  35       2         81
#> 821  37       3         82
#> 822  34       2         82
#> 823  33       2         82
#> 824  33       2         82
#> 825  33       2         82
#> 826  35       2         82
#> 827  33       2         82
#> 828  34       2         82
#> 829  29       2         82
#> 830  32       2         82
#> 831  29       2         83
#> 832  36       2         83
#> 833  31       2         83
#> 834  36       2         83
#> 835  32       2         83
#> 836  33       2         83
#> 837  34       2         83
#> 838  33       2         83
#> 839  36       2         83
#> 840  29       2         83
#> 841  34       2         84
#> 842  34       2         84
#> 843  31       2         84
#> 844  37       3         84
#> 845  34       2         84
#> 846  35       2         84
#> 847  34       2         84
#> 848  34       2         84
#> 849  33       2         84
#> 850  32       2         84
#> 851  37       3         85
#> 852  31       2         85
#> 853  33       2         85
#> 854  38       3         85
#> 855  37       3         85
#> 856  30       2         85
#> 857  36       2         85
#> 858  33       2         85
#> 859  33       2         85
#> 860  36       2         85
#> 861  36       2         86
#> 862  32       2         86
#> 863  34       2         86
#> 864  35       2         86
#> 865  33       2         86
#> 866  36       2         86
#> 867  37       3         86
#> 868  34       2         86
#> 869  35       2         86
#> 870  33       2         86
#> 871  35       2         87
#> 872  30       2         87
#> 873  37       3         87
#> 874  38       3         87
#> 875  36       2         87
#> 876  36       2         87
#> 877  34       2         87
#> 878  35       2         87
#> 879  34       2         87
#> 880  33       2         87
#> 881  34       2         88
#> 882  38       3         88
#> 883  33       2         88
#> 884  35       2         88
#> 885  31       2         88
#> 886  35       2         88
#> 887  33       2         88
#> 888  33       2         88
#> 889  38       3         88
#> 890  35       2         88
#> 891  35       2         89
#> 892  34       2         89
#> 893  34       2         89
#> 894  34       2         89
#> 895  37       3         89
#> 896  39       3         89
#> 897  37       3         89
#> 898  31       2         89
#> 899  35       2         89
#> 900  35       2         89
#> 901  34       2         90
#> 902  38       3         90
#> 903  32       2         90
#> 904  33       2         90
#> 905  38       3         90
#> 906  35       2         90
#> 907  32       2         90
#> 908  38       3         90
#> 909  36       2         90
#> 910  33       2         90
#> 911  37       3         91
#> 912  36       2         91
#> 913  38       3         91
#> 914  37       3         91
#> 915  35       2         91
#> 916  37       3         91
#> 917  37       3         91
#> 918  39       3         91
#> 919  36       2         91
#> 920  36       2         91
#> 921  36       2         92
#> 922  37       3         92
#> 923  39       3         92
#> 924  35       2         92
#> 925  37       3         92
#> 926  35       2         92
#> 927  38       3         92
#> 928  37       3         92
#> 929  36       2         92
#> 930  39       3         92
#> 931  37       3         93
#> 932  34       2         93
#> 933  38       3         93
#> 934  40       3         93
#> 935  37       3         93
#> 936  36       2         93
#> 937  37       3         93
#> 938  33       2         93
#> 939  37       3         93
#> 940  37       3         93
#> 941  39       3         94
#> 942  37       3         94
#> 943  35       2         94
#> 944  38       3         94
#> 945  38       3         94
#> 946  40       3         94
#> 947  39       3         94
#> 948  38       3         94
#> 949  39       3         94
#> 950  37       3         94
#> 951  40       3         95
#> 952  39       3         95
#> 953  39       3         95
#> 954  37       3         95
#> 955  36       2         95
#> 956  39       3         95
#> 957  39       3         95
#> 958  39       3         95
#> 959  34       2         95
#> 960  39       3         95
#> 961  36       2         96
#> 962  36       2         96
#> 963  37       3         96
#> 964  39       3         96
#> 965  39       3         96
#> 966  39       3         96
#> 967  40       3         96
#> 968  38       3         96
#> 969  38       3         96
#> 970  38       3         96
#> 971  38       3         97
#> 972  38       3         97
#> 973  37       3         97
#> 974  38       3         97
#> 975  39       3         97
#> 976  39       3         97
#> 977  39       3         97
#> 978  39       3         97
#> 979  39       3         97
#> 980  40       3         97
#> 981  40       3         98
#> 982  40       3         98
#> 983  39       3         98
#> 984  40       3         98
#> 985  40       3         98
#> 986  39       3         98
#> 987  40       3         98
#> 988  40       3         98
#> 989  39       3         98
#> 990  39       3         98
#> 991  40       3         99
#> 992  39       3         99
#> 993  39       3         99
#> 994  39       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  40       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100