Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simulate_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simulate_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    1       1          1
#> 13    1       1          1
#> 14    1       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    3       1          2
#> 22    1       1          2
#> 23    0       1          2
#> 24    0       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    0       1          2
#> 28    2       1          2
#> 29    2       1          2
#> 30    1       1          2
#> 31    1       1          3
#> 32    2       1          3
#> 33    2       1          3
#> 34    2       1          3
#> 35    2       1          3
#> 36    2       1          3
#> 37    1       1          3
#> 38    0       1          3
#> 39    0       1          3
#> 40    0       1          3
#> 41    3       1          4
#> 42    4       1          4
#> 43    3       1          4
#> 44    2       1          4
#> 45    2       1          4
#> 46    0       1          4
#> 47    2       1          4
#> 48    3       1          4
#> 49    4       1          4
#> 50    1       1          4
#> 51    4       1          5
#> 52    3       1          5
#> 53    3       1          5
#> 54    2       1          5
#> 55    1       1          5
#> 56    1       1          5
#> 57    2       1          5
#> 58    3       1          5
#> 59    1       1          5
#> 60    0       1          5
#> 61    5       1          6
#> 62    2       1          6
#> 63    1       1          6
#> 64    4       1          6
#> 65    2       1          6
#> 66    2       1          6
#> 67    2       1          6
#> 68    3       1          6
#> 69    1       1          6
#> 70    1       1          6
#> 71    1       1          7
#> 72    1       1          7
#> 73    3       1          7
#> 74    1       1          7
#> 75    7       1          7
#> 76    2       1          7
#> 77    2       1          7
#> 78    3       1          7
#> 79    2       1          7
#> 80    4       1          7
#> 81    2       1          8
#> 82    4       1          8
#> 83    4       1          8
#> 84    6       1          8
#> 85    4       1          8
#> 86    4       1          8
#> 87    3       1          8
#> 88    2       1          8
#> 89    2       1          8
#> 90    6       1          8
#> 91    5       1          9
#> 92    7       1          9
#> 93    5       1          9
#> 94    4       1          9
#> 95    4       1          9
#> 96    3       1          9
#> 97    1       1          9
#> 98    4       1          9
#> 99    6       1          9
#> 100   4       1          9
#> 101   4       1         10
#> 102   1       1         10
#> 103   9       1         10
#> 104   6       1         10
#> 105   1       1         10
#> 106   4       1         10
#> 107   2       1         10
#> 108   7       1         10
#> 109   4       1         10
#> 110   4       1         10
#> 111   3       1         11
#> 112   2       1         11
#> 113   7       1         11
#> 114   4       1         11
#> 115   3       1         11
#> 116   3       1         11
#> 117   2       1         11
#> 118   4       1         11
#> 119   8       1         11
#> 120   5       1         11
#> 121   7       1         12
#> 122   2       1         12
#> 123   6       1         12
#> 124   3       1         12
#> 125   3       1         12
#> 126   6       1         12
#> 127   6       1         12
#> 128   5       1         12
#> 129   5       1         12
#> 130   5       1         12
#> 131   2       1         13
#> 132   6       1         13
#> 133   7       1         13
#> 134   4       1         13
#> 135   6       1         13
#> 136   5       1         13
#> 137   8       1         13
#> 138   7       1         13
#> 139   4       1         13
#> 140   9       1         13
#> 141   6       1         14
#> 142   6       1         14
#> 143   4       1         14
#> 144   3       1         14
#> 145   8       1         14
#> 146   7       1         14
#> 147   8       1         14
#> 148   5       1         14
#> 149   5       1         14
#> 150   6       1         14
#> 151   2       1         15
#> 152   7       1         15
#> 153   8       1         15
#> 154   7       1         15
#> 155   5       1         15
#> 156  11       1         15
#> 157   9       1         15
#> 158   7       1         15
#> 159   7       1         15
#> 160   7       1         15
#> 161   6       1         16
#> 162   8       1         16
#> 163  10       1         16
#> 164   8       1         16
#> 165   4       1         16
#> 166  11       1         16
#> 167   4       1         16
#> 168   7       1         16
#> 169   4       1         16
#> 170   4       1         16
#> 171  12       1         17
#> 172   8       1         17
#> 173  13       1         17
#> 174   6       1         17
#> 175   4       1         17
#> 176  10       1         17
#> 177  10       1         17
#> 178   8       1         17
#> 179  12       1         17
#> 180   6       1         17
#> 181   5       1         18
#> 182  10       1         18
#> 183   8       1         18
#> 184   2       1         18
#> 185   8       1         18
#> 186   6       1         18
#> 187   6       1         18
#> 188   6       1         18
#> 189   5       1         18
#> 190  11       1         18
#> 191  12       1         19
#> 192  10       1         19
#> 193  11       1         19
#> 194   5       1         19
#> 195   8       1         19
#> 196   8       1         19
#> 197  12       1         19
#> 198   9       1         19
#> 199   8       1         19
#> 200   9       1         19
#> 201   6       1         20
#> 202   6       1         20
#> 203  12       1         20
#> 204   2       1         20
#> 205   8       1         20
#> 206  11       1         20
#> 207   9       1         20
#> 208  11       1         20
#> 209   9       1         20
#> 210  15       1         20
#> 211  11       1         21
#> 212  10       1         21
#> 213   6       1         21
#> 214  12       1         21
#> 215   6       1         21
#> 216   7       1         21
#> 217   6       1         21
#> 218   7       1         21
#> 219   9       1         21
#> 220   5       1         21
#> 221  10       1         22
#> 222   6       1         22
#> 223   8       1         22
#> 224  13       1         22
#> 225   5       1         22
#> 226  12       1         22
#> 227  11       1         22
#> 228  10       1         22
#> 229  10       1         22
#> 230  11       1         22
#> 231   9       1         23
#> 232  15       1         23
#> 233  11       1         23
#> 234   8       1         23
#> 235  12       1         23
#> 236   9       1         23
#> 237   4       1         23
#> 238   3       1         23
#> 239   9       1         23
#> 240  10       1         23
#> 241   8       1         24
#> 242   9       1         24
#> 243  10       1         24
#> 244  10       1         24
#> 245   7       1         24
#> 246   6       1         24
#> 247   9       1         24
#> 248   9       1         24
#> 249  15       1         24
#> 250   8       1         24
#> 251   5       1         25
#> 252  14       1         25
#> 253  15       1         25
#> 254   9       1         25
#> 255  14       1         25
#> 256  10       1         25
#> 257  13       1         25
#> 258   8       1         25
#> 259   8       1         25
#> 260   9       1         25
#> 261  15       1         26
#> 262  11       1         26
#> 263  13       1         26
#> 264   9       1         26
#> 265   8       1         26
#> 266   4       1         26
#> 267  12       1         26
#> 268  11       1         26
#> 269   7       1         26
#> 270  13       1         26
#> 271  12       1         27
#> 272   6       1         27
#> 273  12       1         27
#> 274  10       1         27
#> 275   7       1         27
#> 276  11       1         27
#> 277   9       1         27
#> 278  13       1         27
#> 279  12       1         27
#> 280  10       1         27
#> 281   6       1         28
#> 282   9       1         28
#> 283   9       1         28
#> 284  12       1         28
#> 285   7       1         28
#> 286  13       1         28
#> 287   8       1         28
#> 288  14       1         28
#> 289  17       1         28
#> 290  10       1         28
#> 291  12       1         29
#> 292  17       1         29
#> 293  10       1         29
#> 294  13       1         29
#> 295  12       1         29
#> 296  10       1         29
#> 297  10       1         29
#> 298  11       1         29
#> 299  15       1         29
#> 300  16       1         29
#> 301  10       1         30
#> 302   8       1         30
#> 303  10       1         30
#> 304  12       1         30
#> 305  14       1         30
#> 306  14       1         30
#> 307  10       1         30
#> 308  13       1         30
#> 309  12       1         30
#> 310   8       1         30
#> 311  15       1         31
#> 312  17       1         31
#> 313  15       1         31
#> 314  14       1         31
#> 315  15       1         31
#> 316  11       1         31
#> 317  12       1         31
#> 318   8       1         31
#> 319  14       1         31
#> 320  15       1         31
#> 321  10       1         32
#> 322  18       1         32
#> 323  19       1         32
#> 324  14       1         32
#> 325  13       1         32
#> 326  13       1         32
#> 327  18       1         32
#> 328  15       1         32
#> 329  11       1         32
#> 330  11       1         32
#> 331  11       1         33
#> 332  12       1         33
#> 333  18       1         33
#> 334  12       1         33
#> 335  14       1         33
#> 336  16       1         33
#> 337  14       1         33
#> 338  13       1         33
#> 339  17       1         33
#> 340  15       1         33
#> 341  16       1         34
#> 342  13       1         34
#> 343  15       1         34
#> 344  13       1         34
#> 345  11       1         34
#> 346  10       1         34
#> 347  13       1         34
#> 348  11       1         34
#> 349  14       1         34
#> 350  13       1         34
#> 351  13       1         35
#> 352  15       1         35
#> 353  13       1         35
#> 354  20       1         35
#> 355  17       1         35
#> 356  15       1         35
#> 357  16       1         35
#> 358  10       1         35
#> 359   8       1         35
#> 360  14       1         35
#> 361  16       1         36
#> 362  19       1         36
#> 363  17       1         36
#> 364  13       1         36
#> 365  14       1         36
#> 366  15       1         36
#> 367  19       1         36
#> 368  11       1         36
#> 369  16       1         36
#> 370  12       1         36
#> 371  19       1         37
#> 372  11       1         37
#> 373  18       1         37
#> 374  15       1         37
#> 375  15       1         37
#> 376  13       1         37
#> 377  14       1         37
#> 378  13       1         37
#> 379  13       1         37
#> 380  16       1         37
#> 381  17       1         38
#> 382  16       1         38
#> 383  16       1         38
#> 384  17       1         38
#> 385  18       1         38
#> 386  19       1         38
#> 387  17       1         38
#> 388  14       1         38
#> 389  14       1         38
#> 390  14       1         38
#> 391  16       1         39
#> 392  24       1         39
#> 393  15       1         39
#> 394  17       1         39
#> 395  16       1         39
#> 396  17       1         39
#> 397  16       1         39
#> 398  18       1         39
#> 399  15       1         39
#> 400  17       1         39
#> 401  18       1         40
#> 402  21       1         40
#> 403  15       1         40
#> 404  18       1         40
#> 405   9       1         40
#> 406  10       1         40
#> 407  14       1         40
#> 408  15       1         40
#> 409  17       1         40
#> 410  17       1         40
#> 411  17       1         41
#> 412  17       1         41
#> 413  13       1         41
#> 414  14       1         41
#> 415  13       1         41
#> 416  13       1         41
#> 417   8       1         41
#> 418  13       1         41
#> 419  15       1         41
#> 420  18       1         41
#> 421  16       1         42
#> 422  15       1         42
#> 423  18       1         42
#> 424  15       1         42
#> 425  20       1         42
#> 426  18       1         42
#> 427  24       1         42
#> 428  19       1         42
#> 429  17       1         42
#> 430  16       1         42
#> 431  14       1         43
#> 432  19       1         43
#> 433  14       1         43
#> 434  16       1         43
#> 435  15       1         43
#> 436  12       1         43
#> 437  17       1         43
#> 438  22       1         43
#> 439  20       1         43
#> 440  21       1         43
#> 441  13       1         44
#> 442  22       1         44
#> 443  16       1         44
#> 444  16       1         44
#> 445  20       1         44
#> 446  17       1         44
#> 447  17       1         44
#> 448  17       1         44
#> 449  19       1         44
#> 450  21       1         44
#> 451  16       1         45
#> 452  20       1         45
#> 453  12       1         45
#> 454  14       1         45
#> 455  16       1         45
#> 456  20       1         45
#> 457  15       1         45
#> 458  27       2         45
#> 459  20       1         45
#> 460  18       1         45
#> 461  22       1         46
#> 462  21       1         46
#> 463  17       1         46
#> 464  18       1         46
#> 465  16       1         46
#> 466  20       1         46
#> 467  21       1         46
#> 468  16       1         46
#> 469  26       2         46
#> 470  18       1         46
#> 471  19       1         47
#> 472  16       1         47
#> 473  20       1         47
#> 474  24       1         47
#> 475  15       1         47
#> 476  17       1         47
#> 477  16       1         47
#> 478  20       1         47
#> 479  15       1         47
#> 480  20       1         47
#> 481  17       1         48
#> 482  18       1         48
#> 483  18       1         48
#> 484  16       1         48
#> 485  24       1         48
#> 486  19       1         48
#> 487  15       1         48
#> 488  24       1         48
#> 489  15       1         48
#> 490  22       1         48
#> 491  24       1         49
#> 492  18       1         49
#> 493  21       1         49
#> 494  24       1         49
#> 495  25       2         49
#> 496  19       1         49
#> 497  19       1         49
#> 498  18       1         49
#> 499  18       1         49
#> 500  23       1         49
#> 501  22       1         50
#> 502  26       2         50
#> 503  16       1         50
#> 504  17       1         50
#> 505  20       1         50
#> 506  20       1         50
#> 507  21       1         50
#> 508  18       1         50
#> 509  21       1         50
#> 510  20       1         50
#> 511  16       1         51
#> 512  20       1         51
#> 513  21       1         51
#> 514  18       1         51
#> 515  23       1         51
#> 516  17       1         51
#> 517  25       2         51
#> 518  19       1         51
#> 519  18       1         51
#> 520  17       1         51
#> 521  19       1         52
#> 522  23       1         52
#> 523  23       1         52
#> 524  22       1         52
#> 525  22       1         52
#> 526  20       1         52
#> 527  21       1         52
#> 528  20       1         52
#> 529  22       1         52
#> 530  23       1         52
#> 531  24       1         53
#> 532  23       1         53
#> 533  18       1         53
#> 534  17       1         53
#> 535  25       2         53
#> 536  24       1         53
#> 537  16       1         53
#> 538  22       1         53
#> 539  21       1         53
#> 540  17       1         53
#> 541  16       1         54
#> 542  16       1         54
#> 543  27       2         54
#> 544  16       1         54
#> 545  22       1         54
#> 546  22       1         54
#> 547  27       2         54
#> 548  15       1         54
#> 549  20       1         54
#> 550  17       1         54
#> 551  22       1         55
#> 552  23       1         55
#> 553  19       1         55
#> 554  29       2         55
#> 555  23       1         55
#> 556  21       1         55
#> 557  21       1         55
#> 558  23       1         55
#> 559  22       1         55
#> 560  21       1         55
#> 561  26       2         56
#> 562  19       1         56
#> 563  19       1         56
#> 564  25       2         56
#> 565  26       2         56
#> 566  20       1         56
#> 567  29       2         56
#> 568  29       2         56
#> 569  20       1         56
#> 570  23       1         56
#> 571  27       2         57
#> 572  25       2         57
#> 573  25       2         57
#> 574  23       1         57
#> 575  25       2         57
#> 576  25       2         57
#> 577  21       1         57
#> 578  21       1         57
#> 579  23       1         57
#> 580  26       2         57
#> 581  25       2         58
#> 582  22       1         58
#> 583  21       1         58
#> 584  25       2         58
#> 585  27       2         58
#> 586  23       1         58
#> 587  19       1         58
#> 588  21       1         58
#> 589  22       1         58
#> 590  23       1         58
#> 591  26       2         59
#> 592  24       1         59
#> 593  20       1         59
#> 594  26       2         59
#> 595  26       2         59
#> 596  21       1         59
#> 597  25       2         59
#> 598  20       1         59
#> 599  23       1         59
#> 600  25       2         59
#> 601  28       2         60
#> 602  21       1         60
#> 603  27       2         60
#> 604  26       2         60
#> 605  21       1         60
#> 606  22       1         60
#> 607  26       2         60
#> 608  23       1         60
#> 609  28       2         60
#> 610  23       1         60
#> 611  31       2         61
#> 612  28       2         61
#> 613  23       1         61
#> 614  23       1         61
#> 615  25       2         61
#> 616  25       2         61
#> 617  23       1         61
#> 618  26       2         61
#> 619  25       2         61
#> 620  24       1         61
#> 621  26       2         62
#> 622  26       2         62
#> 623  26       2         62
#> 624  22       1         62
#> 625  23       1         62
#> 626  22       1         62
#> 627  28       2         62
#> 628  24       1         62
#> 629  25       2         62
#> 630  25       2         62
#> 631  22       1         63
#> 632  29       2         63
#> 633  25       2         63
#> 634  28       2         63
#> 635  27       2         63
#> 636  26       2         63
#> 637  24       1         63
#> 638  24       1         63
#> 639  22       1         63
#> 640  25       2         63
#> 641  26       2         64
#> 642  25       2         64
#> 643  27       2         64
#> 644  28       2         64
#> 645  23       1         64
#> 646  24       1         64
#> 647  28       2         64
#> 648  27       2         64
#> 649  26       2         64
#> 650  27       2         64
#> 651  27       2         65
#> 652  22       1         65
#> 653  23       1         65
#> 654  31       2         65
#> 655  29       2         65
#> 656  27       2         65
#> 657  27       2         65
#> 658  27       2         65
#> 659  28       2         65
#> 660  26       2         65
#> 661  25       2         66
#> 662  25       2         66
#> 663  28       2         66
#> 664  26       2         66
#> 665  28       2         66
#> 666  26       2         66
#> 667  27       2         66
#> 668  20       1         66
#> 669  29       2         66
#> 670  19       1         66
#> 671  21       1         67
#> 672  26       2         67
#> 673  25       2         67
#> 674  27       2         67
#> 675  25       2         67
#> 676  29       2         67
#> 677  27       2         67
#> 678  30       2         67
#> 679  27       2         67
#> 680  28       2         67
#> 681  24       1         68
#> 682  31       2         68
#> 683  27       2         68
#> 684  28       2         68
#> 685  28       2         68
#> 686  28       2         68
#> 687  21       1         68
#> 688  30       2         68
#> 689  26       2         68
#> 690  28       2         68
#> 691  28       2         69
#> 692  22       1         69
#> 693  26       2         69
#> 694  30       2         69
#> 695  29       2         69
#> 696  27       2         69
#> 697  29       2         69
#> 698  33       2         69
#> 699  26       2         69
#> 700  26       2         69
#> 701  28       2         70
#> 702  27       2         70
#> 703  33       2         70
#> 704  27       2         70
#> 705  30       2         70
#> 706  25       2         70
#> 707  28       2         70
#> 708  28       2         70
#> 709  29       2         70
#> 710  27       2         70
#> 711  25       2         71
#> 712  27       2         71
#> 713  28       2         71
#> 714  30       2         71
#> 715  23       1         71
#> 716  28       2         71
#> 717  29       2         71
#> 718  25       2         71
#> 719  31       2         71
#> 720  28       2         71
#> 721  30       2         72
#> 722  31       2         72
#> 723  29       2         72
#> 724  32       2         72
#> 725  20       1         72
#> 726  31       2         72
#> 727  29       2         72
#> 728  29       2         72
#> 729  26       2         72
#> 730  21       1         72
#> 731  28       2         73
#> 732  32       2         73
#> 733  27       2         73
#> 734  31       2         73
#> 735  30       2         73
#> 736  31       2         73
#> 737  32       2         73
#> 738  31       2         73
#> 739  29       2         73
#> 740  31       2         73
#> 741  31       2         74
#> 742  32       2         74
#> 743  29       2         74
#> 744  24       1         74
#> 745  30       2         74
#> 746  28       2         74
#> 747  36       2         74
#> 748  28       2         74
#> 749  31       2         74
#> 750  28       2         74
#> 751  33       2         75
#> 752  34       2         75
#> 753  29       2         75
#> 754  31       2         75
#> 755  28       2         75
#> 756  35       2         75
#> 757  29       2         75
#> 758  29       2         75
#> 759  30       2         75
#> 760  30       2         75
#> 761  35       2         76
#> 762  33       2         76
#> 763  28       2         76
#> 764  28       2         76
#> 765  28       2         76
#> 766  31       2         76
#> 767  33       2         76
#> 768  31       2         76
#> 769  30       2         76
#> 770  37       3         76
#> 771  35       2         77
#> 772  33       2         77
#> 773  32       2         77
#> 774  33       2         77
#> 775  32       2         77
#> 776  30       2         77
#> 777  27       2         77
#> 778  30       2         77
#> 779  32       2         77
#> 780  33       2         77
#> 781  33       2         78
#> 782  31       2         78
#> 783  30       2         78
#> 784  29       2         78
#> 785  32       2         78
#> 786  30       2         78
#> 787  30       2         78
#> 788  32       2         78
#> 789  34       2         78
#> 790  27       2         78
#> 791  32       2         79
#> 792  34       2         79
#> 793  30       2         79
#> 794  32       2         79
#> 795  35       2         79
#> 796  33       2         79
#> 797  34       2         79
#> 798  31       2         79
#> 799  32       2         79
#> 800  30       2         79
#> 801  33       2         80
#> 802  32       2         80
#> 803  32       2         80
#> 804  32       2         80
#> 805  30       2         80
#> 806  32       2         80
#> 807  30       2         80
#> 808  32       2         80
#> 809  31       2         80
#> 810  31       2         80
#> 811  33       2         81
#> 812  34       2         81
#> 813  34       2         81
#> 814  35       2         81
#> 815  31       2         81
#> 816  32       2         81
#> 817  30       2         81
#> 818  31       2         81
#> 819  31       2         81
#> 820  29       2         81
#> 821  33       2         82
#> 822  34       2         82
#> 823  34       2         82
#> 824  35       2         82
#> 825  33       2         82
#> 826  33       2         82
#> 827  31       2         82
#> 828  33       2         82
#> 829  33       2         82
#> 830  29       2         82
#> 831  32       2         83
#> 832  34       2         83
#> 833  31       2         83
#> 834  35       2         83
#> 835  34       2         83
#> 836  31       2         83
#> 837  35       2         83
#> 838  34       2         83
#> 839  25       2         83
#> 840  33       2         83
#> 841  32       2         84
#> 842  29       2         84
#> 843  40       3         84
#> 844  33       2         84
#> 845  35       2         84
#> 846  35       2         84
#> 847  35       2         84
#> 848  36       2         84
#> 849  35       2         84
#> 850  35       2         84
#> 851  35       2         85
#> 852  32       2         85
#> 853  36       2         85
#> 854  35       2         85
#> 855  32       2         85
#> 856  31       2         85
#> 857  32       2         85
#> 858  33       2         85
#> 859  35       2         85
#> 860  33       2         85
#> 861  39       3         86
#> 862  35       2         86
#> 863  31       2         86
#> 864  33       2         86
#> 865  34       2         86
#> 866  36       2         86
#> 867  33       2         86
#> 868  35       2         86
#> 869  35       2         86
#> 870  35       2         86
#> 871  36       2         87
#> 872  36       2         87
#> 873  36       2         87
#> 874  35       2         87
#> 875  33       2         87
#> 876  33       2         87
#> 877  36       2         87
#> 878  35       2         87
#> 879  35       2         87
#> 880  34       2         87
#> 881  38       3         88
#> 882  37       3         88
#> 883  35       2         88
#> 884  36       2         88
#> 885  34       2         88
#> 886  34       2         88
#> 887  33       2         88
#> 888  37       3         88
#> 889  38       3         88
#> 890  35       2         88
#> 891  36       2         89
#> 892  34       2         89
#> 893  32       2         89
#> 894  36       2         89
#> 895  37       3         89
#> 896  36       2         89
#> 897  34       2         89
#> 898  39       3         89
#> 899  35       2         89
#> 900  33       2         89
#> 901  37       3         90
#> 902  38       3         90
#> 903  37       3         90
#> 904  39       3         90
#> 905  33       2         90
#> 906  38       3         90
#> 907  35       2         90
#> 908  37       3         90
#> 909  34       2         90
#> 910  38       3         90
#> 911  37       3         91
#> 912  36       2         91
#> 913  38       3         91
#> 914  38       3         91
#> 915  37       3         91
#> 916  37       3         91
#> 917  39       3         91
#> 918  37       3         91
#> 919  37       3         91
#> 920  36       2         91
#> 921  37       3         92
#> 922  36       2         92
#> 923  36       2         92
#> 924  40       3         92
#> 925  38       3         92
#> 926  37       3         92
#> 927  37       3         92
#> 928  36       2         92
#> 929  39       3         92
#> 930  34       2         92
#> 931  37       3         93
#> 932  37       3         93
#> 933  39       3         93
#> 934  37       3         93
#> 935  34       2         93
#> 936  38       3         93
#> 937  35       2         93
#> 938  36       2         93
#> 939  37       3         93
#> 940  38       3         93
#> 941  36       2         94
#> 942  38       3         94
#> 943  39       3         94
#> 944  39       3         94
#> 945  37       3         94
#> 946  40       3         94
#> 947  37       3         94
#> 948  35       2         94
#> 949  38       3         94
#> 950  38       3         94
#> 951  37       3         95
#> 952  39       3         95
#> 953  38       3         95
#> 954  39       3         95
#> 955  36       2         95
#> 956  36       2         95
#> 957  32       2         95
#> 958  39       3         95
#> 959  38       3         95
#> 960  39       3         95
#> 961  37       3         96
#> 962  39       3         96
#> 963  39       3         96
#> 964  38       3         96
#> 965  38       3         96
#> 966  38       3         96
#> 967  40       3         96
#> 968  38       3         96
#> 969  38       3         96
#> 970  39       3         96
#> 971  39       3         97
#> 972  40       3         97
#> 973  38       3         97
#> 974  40       3         97
#> 975  38       3         97
#> 976  39       3         97
#> 977  38       3         97
#> 978  39       3         97
#> 979  40       3         97
#> 980  40       3         97
#> 981  39       3         98
#> 982  39       3         98
#> 983  39       3         98
#> 984  38       3         98
#> 985  39       3         98
#> 986  37       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  39       3         98
#> 991  40       3         99
#> 992  40       3         99
#> 993  38       3         99
#> 994  39       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  38       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 39       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100