Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    0       1          1
#> 13    0       1          1
#> 14    0       1          1
#> 15    0       1          1
#> 16    1       1          1
#> 17    1       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    1       1          2
#> 23    2       1          2
#> 24    2       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    1       1          2
#> 28    0       1          2
#> 29    0       1          2
#> 30    1       1          2
#> 31    1       1          3
#> 32    2       1          3
#> 33    4       1          3
#> 34    1       1          3
#> 35    1       1          3
#> 36    0       1          3
#> 37    1       1          3
#> 38    3       1          3
#> 39    1       1          3
#> 40    2       1          3
#> 41    1       1          4
#> 42    2       1          4
#> 43    1       1          4
#> 44    1       1          4
#> 45    2       1          4
#> 46    1       1          4
#> 47    2       1          4
#> 48    2       1          4
#> 49    1       1          4
#> 50    4       1          4
#> 51    2       1          5
#> 52    3       1          5
#> 53    1       1          5
#> 54    1       1          5
#> 55    1       1          5
#> 56    0       1          5
#> 57    3       1          5
#> 58    1       1          5
#> 59    1       1          5
#> 60    3       1          5
#> 61    1       1          6
#> 62    2       1          6
#> 63    3       1          6
#> 64    2       1          6
#> 65    3       1          6
#> 66    1       1          6
#> 67    2       1          6
#> 68    1       1          6
#> 69    7       1          6
#> 70    3       1          6
#> 71    4       1          7
#> 72    2       1          7
#> 73    3       1          7
#> 74    1       1          7
#> 75    2       1          7
#> 76    2       1          7
#> 77    2       1          7
#> 78    3       1          7
#> 79    3       1          7
#> 80    4       1          7
#> 81    0       1          8
#> 82    1       1          8
#> 83    1       1          8
#> 84    3       1          8
#> 85    3       1          8
#> 86    3       1          8
#> 87    5       1          8
#> 88    2       1          8
#> 89    6       1          8
#> 90    1       1          8
#> 91    3       1          9
#> 92    1       1          9
#> 93    4       1          9
#> 94    3       1          9
#> 95    5       1          9
#> 96    5       1          9
#> 97    2       1          9
#> 98    3       1          9
#> 99    7       1          9
#> 100   1       1          9
#> 101   2       1         10
#> 102   2       1         10
#> 103   4       1         10
#> 104   0       1         10
#> 105   4       1         10
#> 106   4       1         10
#> 107   7       1         10
#> 108   2       1         10
#> 109   6       1         10
#> 110  10       1         10
#> 111   5       1         11
#> 112   5       1         11
#> 113   4       1         11
#> 114   7       1         11
#> 115   4       1         11
#> 116   5       1         11
#> 117   4       1         11
#> 118   6       1         11
#> 119   5       1         11
#> 120   7       1         11
#> 121   5       1         12
#> 122   3       1         12
#> 123   9       1         12
#> 124   6       1         12
#> 125   5       1         12
#> 126   5       1         12
#> 127   8       1         12
#> 128   3       1         12
#> 129   5       1         12
#> 130   7       1         12
#> 131   3       1         13
#> 132   3       1         13
#> 133   1       1         13
#> 134   4       1         13
#> 135   7       1         13
#> 136   7       1         13
#> 137   6       1         13
#> 138   6       1         13
#> 139   4       1         13
#> 140   2       1         13
#> 141   2       1         14
#> 142   1       1         14
#> 143  10       1         14
#> 144   6       1         14
#> 145   6       1         14
#> 146  10       1         14
#> 147   9       1         14
#> 148   4       1         14
#> 149   3       1         14
#> 150   9       1         14
#> 151   5       1         15
#> 152   7       1         15
#> 153   3       1         15
#> 154   9       1         15
#> 155   3       1         15
#> 156   9       1         15
#> 157   3       1         15
#> 158   6       1         15
#> 159   4       1         15
#> 160  10       1         15
#> 161   4       1         16
#> 162   8       1         16
#> 163   6       1         16
#> 164  10       1         16
#> 165   4       1         16
#> 166   5       1         16
#> 167   4       1         16
#> 168   8       1         16
#> 169   1       1         16
#> 170   1       1         16
#> 171   8       1         17
#> 172   5       1         17
#> 173   9       1         17
#> 174   8       1         17
#> 175   8       1         17
#> 176   4       1         17
#> 177   4       1         17
#> 178   9       1         17
#> 179   7       1         17
#> 180   7       1         17
#> 181   6       1         18
#> 182   8       1         18
#> 183   7       1         18
#> 184   8       1         18
#> 185   7       1         18
#> 186   5       1         18
#> 187   7       1         18
#> 188   8       1         18
#> 189  11       1         18
#> 190  11       1         18
#> 191   8       1         19
#> 192  13       1         19
#> 193   8       1         19
#> 194   4       1         19
#> 195   5       1         19
#> 196   6       1         19
#> 197   8       1         19
#> 198  12       1         19
#> 199   7       1         19
#> 200   8       1         19
#> 201   7       1         20
#> 202  12       1         20
#> 203   6       1         20
#> 204  10       1         20
#> 205   6       1         20
#> 206   7       1         20
#> 207   6       1         20
#> 208  12       1         20
#> 209  11       1         20
#> 210   5       1         20
#> 211   9       1         21
#> 212   8       1         21
#> 213   9       1         21
#> 214   6       1         21
#> 215  12       1         21
#> 216   9       1         21
#> 217   9       1         21
#> 218   9       1         21
#> 219  11       1         21
#> 220  11       1         21
#> 221   8       1         22
#> 222  10       1         22
#> 223   6       1         22
#> 224   8       1         22
#> 225  11       1         22
#> 226   6       1         22
#> 227   8       1         22
#> 228   8       1         22
#> 229   8       1         22
#> 230   5       1         22
#> 231   8       1         23
#> 232   7       1         23
#> 233   9       1         23
#> 234  11       1         23
#> 235  12       1         23
#> 236  11       1         23
#> 237   9       1         23
#> 238  10       1         23
#> 239   5       1         23
#> 240   7       1         23
#> 241   7       1         24
#> 242  13       1         24
#> 243   8       1         24
#> 244   6       1         24
#> 245  11       1         24
#> 246   9       1         24
#> 247   7       1         24
#> 248   8       1         24
#> 249   7       1         24
#> 250  13       1         24
#> 251  10       1         25
#> 252   8       1         25
#> 253  10       1         25
#> 254   7       1         25
#> 255  13       1         25
#> 256   9       1         25
#> 257  13       1         25
#> 258   9       1         25
#> 259   6       1         25
#> 260   6       1         25
#> 261  12       1         26
#> 262  12       1         26
#> 263   6       1         26
#> 264  14       1         26
#> 265  13       1         26
#> 266  10       1         26
#> 267  16       1         26
#> 268   9       1         26
#> 269   8       1         26
#> 270   6       1         26
#> 271  19       1         27
#> 272  12       1         27
#> 273   9       1         27
#> 274   6       1         27
#> 275  10       1         27
#> 276   8       1         27
#> 277   9       1         27
#> 278   6       1         27
#> 279  16       1         27
#> 280   8       1         27
#> 281  16       1         28
#> 282  14       1         28
#> 283  16       1         28
#> 284  11       1         28
#> 285   7       1         28
#> 286   9       1         28
#> 287  10       1         28
#> 288  13       1         28
#> 289   9       1         28
#> 290   4       1         28
#> 291  14       1         29
#> 292  15       1         29
#> 293  13       1         29
#> 294  15       1         29
#> 295  14       1         29
#> 296   9       1         29
#> 297   9       1         29
#> 298   7       1         29
#> 299   8       1         29
#> 300  10       1         29
#> 301  14       1         30
#> 302   8       1         30
#> 303  14       1         30
#> 304  10       1         30
#> 305  15       1         30
#> 306  14       1         30
#> 307   6       1         30
#> 308   8       1         30
#> 309  11       1         30
#> 310  13       1         30
#> 311  11       1         31
#> 312   9       1         31
#> 313  12       1         31
#> 314  14       1         31
#> 315  14       1         31
#> 316  12       1         31
#> 317  11       1         31
#> 318   9       1         31
#> 319  10       1         31
#> 320  14       1         31
#> 321  14       1         32
#> 322  13       1         32
#> 323  12       1         32
#> 324  11       1         32
#> 325   9       1         32
#> 326  18       1         32
#> 327   8       1         32
#> 328  15       1         32
#> 329  17       1         32
#> 330  16       1         32
#> 331   9       1         33
#> 332  12       1         33
#> 333  17       1         33
#> 334  15       1         33
#> 335  16       1         33
#> 336  12       1         33
#> 337  18       1         33
#> 338  14       1         33
#> 339  13       1         33
#> 340  14       1         33
#> 341  18       1         34
#> 342  19       1         34
#> 343  17       1         34
#> 344  16       1         34
#> 345  12       1         34
#> 346  14       1         34
#> 347  14       1         34
#> 348  14       1         34
#> 349  10       1         34
#> 350  12       1         34
#> 351  10       1         35
#> 352  12       1         35
#> 353  11       1         35
#> 354  15       1         35
#> 355  17       1         35
#> 356  10       1         35
#> 357  12       1         35
#> 358  10       1         35
#> 359   8       1         35
#> 360  11       1         35
#> 361  15       1         36
#> 362  15       1         36
#> 363  15       1         36
#> 364  18       1         36
#> 365  16       1         36
#> 366  18       1         36
#> 367  13       1         36
#> 368  14       1         36
#> 369  14       1         36
#> 370  10       1         36
#> 371  18       1         37
#> 372  12       1         37
#> 373  16       1         37
#> 374  17       1         37
#> 375  18       1         37
#> 376  19       1         37
#> 377  13       1         37
#> 378  12       1         37
#> 379  18       1         37
#> 380  15       1         37
#> 381  13       1         38
#> 382  11       1         38
#> 383  15       1         38
#> 384  17       1         38
#> 385  13       1         38
#> 386  12       1         38
#> 387  16       1         38
#> 388  15       1         38
#> 389  13       1         38
#> 390  17       1         38
#> 391   9       1         39
#> 392  14       1         39
#> 393  18       1         39
#> 394  11       1         39
#> 395  15       1         39
#> 396  15       1         39
#> 397  13       1         39
#> 398  18       1         39
#> 399  24       1         39
#> 400  19       1         39
#> 401  16       1         40
#> 402  16       1         40
#> 403  20       1         40
#> 404  14       1         40
#> 405  20       1         40
#> 406  12       1         40
#> 407  16       1         40
#> 408  20       1         40
#> 409  12       1         40
#> 410  17       1         40
#> 411  17       1         41
#> 412  13       1         41
#> 413  21       1         41
#> 414  24       1         41
#> 415  20       1         41
#> 416  15       1         41
#> 417  21       1         41
#> 418  13       1         41
#> 419  12       1         41
#> 420  17       1         41
#> 421  23       1         42
#> 422  14       1         42
#> 423  17       1         42
#> 424  12       1         42
#> 425  18       1         42
#> 426  18       1         42
#> 427  21       1         42
#> 428  16       1         42
#> 429  11       1         42
#> 430  23       1         42
#> 431  23       1         43
#> 432  20       1         43
#> 433  17       1         43
#> 434  20       1         43
#> 435  17       1         43
#> 436  16       1         43
#> 437  10       1         43
#> 438  19       1         43
#> 439  13       1         43
#> 440  19       1         43
#> 441  24       1         44
#> 442  12       1         44
#> 443   9       1         44
#> 444  20       1         44
#> 445  21       1         44
#> 446  18       1         44
#> 447  17       1         44
#> 448  22       1         44
#> 449  19       1         44
#> 450  18       1         44
#> 451  19       1         45
#> 452  24       1         45
#> 453  19       1         45
#> 454  13       1         45
#> 455  14       1         45
#> 456  24       1         45
#> 457  16       1         45
#> 458  15       1         45
#> 459  22       1         45
#> 460  18       1         45
#> 461  21       1         46
#> 462  18       1         46
#> 463  18       1         46
#> 464  15       1         46
#> 465  18       1         46
#> 466  14       1         46
#> 467  18       1         46
#> 468  10       1         46
#> 469  18       1         46
#> 470  16       1         46
#> 471  24       1         47
#> 472  20       1         47
#> 473  21       1         47
#> 474  21       1         47
#> 475  17       1         47
#> 476  20       1         47
#> 477  18       1         47
#> 478  21       1         47
#> 479  20       1         47
#> 480  26       2         47
#> 481  20       1         48
#> 482  20       1         48
#> 483  19       1         48
#> 484  25       2         48
#> 485  17       1         48
#> 486  20       1         48
#> 487  20       1         48
#> 488  24       1         48
#> 489  22       1         48
#> 490  13       1         48
#> 491  20       1         49
#> 492  21       1         49
#> 493  15       1         49
#> 494  18       1         49
#> 495  22       1         49
#> 496  13       1         49
#> 497  27       2         49
#> 498  14       1         49
#> 499  19       1         49
#> 500  23       1         49
#> 501  14       1         50
#> 502  21       1         50
#> 503  22       1         50
#> 504  23       1         50
#> 505  20       1         50
#> 506  19       1         50
#> 507  18       1         50
#> 508  17       1         50
#> 509  24       1         50
#> 510  17       1         50
#> 511  20       1         51
#> 512  22       1         51
#> 513  21       1         51
#> 514  25       2         51
#> 515  18       1         51
#> 516  21       1         51
#> 517  21       1         51
#> 518  14       1         51
#> 519  20       1         51
#> 520  15       1         51
#> 521  21       1         52
#> 522  17       1         52
#> 523  21       1         52
#> 524  22       1         52
#> 525  22       1         52
#> 526  21       1         52
#> 527  22       1         52
#> 528  21       1         52
#> 529  26       2         52
#> 530  22       1         52
#> 531  22       1         53
#> 532  21       1         53
#> 533  25       2         53
#> 534  26       2         53
#> 535  20       1         53
#> 536  18       1         53
#> 537  23       1         53
#> 538  25       2         53
#> 539  24       1         53
#> 540  19       1         53
#> 541  24       1         54
#> 542  22       1         54
#> 543  22       1         54
#> 544  24       1         54
#> 545  20       1         54
#> 546  24       1         54
#> 547  23       1         54
#> 548  20       1         54
#> 549  21       1         54
#> 550  24       1         54
#> 551  20       1         55
#> 552  19       1         55
#> 553  23       1         55
#> 554  21       1         55
#> 555  13       1         55
#> 556  18       1         55
#> 557  22       1         55
#> 558  19       1         55
#> 559  26       2         55
#> 560  24       1         55
#> 561  20       1         56
#> 562  28       2         56
#> 563  21       1         56
#> 564  27       2         56
#> 565  22       1         56
#> 566  21       1         56
#> 567  21       1         56
#> 568  19       1         56
#> 569  25       2         56
#> 570  20       1         56
#> 571  25       2         57
#> 572  19       1         57
#> 573  23       1         57
#> 574  26       2         57
#> 575  24       1         57
#> 576  25       2         57
#> 577  26       2         57
#> 578  25       2         57
#> 579  24       1         57
#> 580  24       1         57
#> 581  22       1         58
#> 582  23       1         58
#> 583  21       1         58
#> 584  24       1         58
#> 585  25       2         58
#> 586  25       2         58
#> 587  26       2         58
#> 588  23       1         58
#> 589  25       2         58
#> 590  18       1         58
#> 591  23       1         59
#> 592  26       2         59
#> 593  28       2         59
#> 594  28       2         59
#> 595  26       2         59
#> 596  23       1         59
#> 597  22       1         59
#> 598  23       1         59
#> 599  18       1         59
#> 600  22       1         59
#> 601  29       2         60
#> 602  25       2         60
#> 603  21       1         60
#> 604  29       2         60
#> 605  27       2         60
#> 606  23       1         60
#> 607  18       1         60
#> 608  23       1         60
#> 609  21       1         60
#> 610  31       2         60
#> 611  22       1         61
#> 612  25       2         61
#> 613  22       1         61
#> 614  22       1         61
#> 615  22       1         61
#> 616  21       1         61
#> 617  24       1         61
#> 618  23       1         61
#> 619  24       1         61
#> 620  29       2         61
#> 621  26       2         62
#> 622  25       2         62
#> 623  17       1         62
#> 624  24       1         62
#> 625  22       1         62
#> 626  25       2         62
#> 627  29       2         62
#> 628  21       1         62
#> 629  23       1         62
#> 630  33       2         62
#> 631  28       2         63
#> 632  23       1         63
#> 633  23       1         63
#> 634  22       1         63
#> 635  31       2         63
#> 636  25       2         63
#> 637  19       1         63
#> 638  25       2         63
#> 639  25       2         63
#> 640  25       2         63
#> 641  27       2         64
#> 642  21       1         64
#> 643  30       2         64
#> 644  26       2         64
#> 645  23       1         64
#> 646  26       2         64
#> 647  30       2         64
#> 648  29       2         64
#> 649  31       2         64
#> 650  26       2         64
#> 651  26       2         65
#> 652  26       2         65
#> 653  30       2         65
#> 654  27       2         65
#> 655  26       2         65
#> 656  23       1         65
#> 657  25       2         65
#> 658  26       2         65
#> 659  19       1         65
#> 660  28       2         65
#> 661  26       2         66
#> 662  30       2         66
#> 663  23       1         66
#> 664  26       2         66
#> 665  23       1         66
#> 666  27       2         66
#> 667  25       2         66
#> 668  30       2         66
#> 669  33       2         66
#> 670  22       1         66
#> 671  26       2         67
#> 672  31       2         67
#> 673  24       1         67
#> 674  28       2         67
#> 675  28       2         67
#> 676  23       1         67
#> 677  27       2         67
#> 678  33       2         67
#> 679  21       1         67
#> 680  26       2         67
#> 681  34       2         68
#> 682  32       2         68
#> 683  29       2         68
#> 684  26       2         68
#> 685  24       1         68
#> 686  30       2         68
#> 687  35       2         68
#> 688  30       2         68
#> 689  22       1         68
#> 690  26       2         68
#> 691  28       2         69
#> 692  25       2         69
#> 693  26       2         69
#> 694  30       2         69
#> 695  29       2         69
#> 696  27       2         69
#> 697  28       2         69
#> 698  26       2         69
#> 699  28       2         69
#> 700  25       2         69
#> 701  29       2         70
#> 702  27       2         70
#> 703  28       2         70
#> 704  31       2         70
#> 705  28       2         70
#> 706  26       2         70
#> 707  30       2         70
#> 708  29       2         70
#> 709  26       2         70
#> 710  26       2         70
#> 711  26       2         71
#> 712  29       2         71
#> 713  30       2         71
#> 714  32       2         71
#> 715  25       2         71
#> 716  24       1         71
#> 717  31       2         71
#> 718  30       2         71
#> 719  28       2         71
#> 720  28       2         71
#> 721  25       2         72
#> 722  32       2         72
#> 723  30       2         72
#> 724  30       2         72
#> 725  29       2         72
#> 726  28       2         72
#> 727  25       2         72
#> 728  32       2         72
#> 729  29       2         72
#> 730  33       2         72
#> 731  26       2         73
#> 732  31       2         73
#> 733  29       2         73
#> 734  29       2         73
#> 735  34       2         73
#> 736  30       2         73
#> 737  31       2         73
#> 738  28       2         73
#> 739  31       2         73
#> 740  30       2         73
#> 741  36       2         74
#> 742  26       2         74
#> 743  27       2         74
#> 744  27       2         74
#> 745  31       2         74
#> 746  31       2         74
#> 747  32       2         74
#> 748  32       2         74
#> 749  33       2         74
#> 750  27       2         74
#> 751  33       2         75
#> 752  31       2         75
#> 753  25       2         75
#> 754  29       2         75
#> 755  30       2         75
#> 756  33       2         75
#> 757  30       2         75
#> 758  29       2         75
#> 759  29       2         75
#> 760  26       2         75
#> 761  29       2         76
#> 762  26       2         76
#> 763  35       2         76
#> 764  28       2         76
#> 765  35       2         76
#> 766  27       2         76
#> 767  31       2         76
#> 768  27       2         76
#> 769  32       2         76
#> 770  32       2         76
#> 771  32       2         77
#> 772  34       2         77
#> 773  30       2         77
#> 774  29       2         77
#> 775  29       2         77
#> 776  31       2         77
#> 777  26       2         77
#> 778  30       2         77
#> 779  34       2         77
#> 780  36       2         77
#> 781  33       2         78
#> 782  31       2         78
#> 783  36       2         78
#> 784  27       2         78
#> 785  34       2         78
#> 786  28       2         78
#> 787  28       2         78
#> 788  27       2         78
#> 789  26       2         78
#> 790  30       2         78
#> 791  30       2         79
#> 792  34       2         79
#> 793  33       2         79
#> 794  30       2         79
#> 795  29       2         79
#> 796  29       2         79
#> 797  36       2         79
#> 798  34       2         79
#> 799  30       2         79
#> 800  33       2         79
#> 801  32       2         80
#> 802  34       2         80
#> 803  35       2         80
#> 804  28       2         80
#> 805  27       2         80
#> 806  34       2         80
#> 807  30       2         80
#> 808  31       2         80
#> 809  23       1         80
#> 810  31       2         80
#> 811  32       2         81
#> 812  26       2         81
#> 813  27       2         81
#> 814  31       2         81
#> 815  34       2         81
#> 816  27       2         81
#> 817  33       2         81
#> 818  32       2         81
#> 819  31       2         81
#> 820  33       2         81
#> 821  32       2         82
#> 822  30       2         82
#> 823  34       2         82
#> 824  34       2         82
#> 825  32       2         82
#> 826  33       2         82
#> 827  35       2         82
#> 828  35       2         82
#> 829  30       2         82
#> 830  29       2         82
#> 831  29       2         83
#> 832  34       2         83
#> 833  34       2         83
#> 834  33       2         83
#> 835  30       2         83
#> 836  33       2         83
#> 837  35       2         83
#> 838  32       2         83
#> 839  31       2         83
#> 840  37       3         83
#> 841  34       2         84
#> 842  38       3         84
#> 843  38       3         84
#> 844  30       2         84
#> 845  36       2         84
#> 846  33       2         84
#> 847  31       2         84
#> 848  39       3         84
#> 849  36       2         84
#> 850  34       2         84
#> 851  32       2         85
#> 852  37       3         85
#> 853  37       3         85
#> 854  35       2         85
#> 855  34       2         85
#> 856  37       3         85
#> 857  31       2         85
#> 858  36       2         85
#> 859  34       2         85
#> 860  35       2         85
#> 861  34       2         86
#> 862  34       2         86
#> 863  36       2         86
#> 864  35       2         86
#> 865  34       2         86
#> 866  28       2         86
#> 867  34       2         86
#> 868  34       2         86
#> 869  28       2         86
#> 870  38       3         86
#> 871  32       2         87
#> 872  38       3         87
#> 873  33       2         87
#> 874  34       2         87
#> 875  33       2         87
#> 876  34       2         87
#> 877  36       2         87
#> 878  29       2         87
#> 879  35       2         87
#> 880  36       2         87
#> 881  34       2         88
#> 882  33       2         88
#> 883  38       3         88
#> 884  38       3         88
#> 885  36       2         88
#> 886  35       2         88
#> 887  36       2         88
#> 888  35       2         88
#> 889  36       2         88
#> 890  37       3         88
#> 891  33       2         89
#> 892  35       2         89
#> 893  36       2         89
#> 894  32       2         89
#> 895  37       3         89
#> 896  37       3         89
#> 897  35       2         89
#> 898  35       2         89
#> 899  35       2         89
#> 900  34       2         89
#> 901  35       2         90
#> 902  35       2         90
#> 903  38       3         90
#> 904  34       2         90
#> 905  36       2         90
#> 906  37       3         90
#> 907  36       2         90
#> 908  36       2         90
#> 909  38       3         90
#> 910  36       2         90
#> 911  35       2         91
#> 912  36       2         91
#> 913  34       2         91
#> 914  35       2         91
#> 915  37       3         91
#> 916  37       3         91
#> 917  36       2         91
#> 918  34       2         91
#> 919  39       3         91
#> 920  36       2         91
#> 921  35       2         92
#> 922  35       2         92
#> 923  35       2         92
#> 924  38       3         92
#> 925  36       2         92
#> 926  37       3         92
#> 927  33       2         92
#> 928  37       3         92
#> 929  39       3         92
#> 930  36       2         92
#> 931  39       3         93
#> 932  37       3         93
#> 933  36       2         93
#> 934  37       3         93
#> 935  39       3         93
#> 936  39       3         93
#> 937  38       3         93
#> 938  39       3         93
#> 939  39       3         93
#> 940  39       3         93
#> 941  36       2         94
#> 942  39       3         94
#> 943  38       3         94
#> 944  36       2         94
#> 945  36       2         94
#> 946  38       3         94
#> 947  36       2         94
#> 948  38       3         94
#> 949  39       3         94
#> 950  39       3         94
#> 951  39       3         95
#> 952  36       2         95
#> 953  39       3         95
#> 954  39       3         95
#> 955  38       3         95
#> 956  38       3         95
#> 957  39       3         95
#> 958  36       2         95
#> 959  38       3         95
#> 960  38       3         95
#> 961  37       3         96
#> 962  38       3         96
#> 963  38       3         96
#> 964  39       3         96
#> 965  39       3         96
#> 966  38       3         96
#> 967  39       3         96
#> 968  39       3         96
#> 969  38       3         96
#> 970  40       3         96
#> 971  38       3         97
#> 972  39       3         97
#> 973  40       3         97
#> 974  39       3         97
#> 975  40       3         97
#> 976  37       3         97
#> 977  38       3         97
#> 978  38       3         97
#> 979  40       3         97
#> 980  40       3         97
#> 981  39       3         98
#> 982  40       3         98
#> 983  40       3         98
#> 984  40       3         98
#> 985  37       3         98
#> 986  39       3         98
#> 987  40       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  40       3         98
#> 991  40       3         99
#> 992  39       3         99
#> 993  40       3         99
#> 994  39       3         99
#> 995  39       3         99
#> 996  40       3         99
#> 997  38       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100