Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    1       1          1
#> 12    0       1          1
#> 13    1       1          1
#> 14    0       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    1       1          1
#> 18    1       1          1
#> 19    1       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    0       1          2
#> 23    2       1          2
#> 24    3       1          2
#> 25    1       1          2
#> 26    0       1          2
#> 27    3       1          2
#> 28    1       1          2
#> 29    0       1          2
#> 30    0       1          2
#> 31    0       1          3
#> 32    1       1          3
#> 33    1       1          3
#> 34    2       1          3
#> 35    1       1          3
#> 36    1       1          3
#> 37    2       1          3
#> 38    1       1          3
#> 39    1       1          3
#> 40    2       1          3
#> 41    1       1          4
#> 42    1       1          4
#> 43    1       1          4
#> 44    4       1          4
#> 45    1       1          4
#> 46    4       1          4
#> 47    1       1          4
#> 48    3       1          4
#> 49    0       1          4
#> 50    0       1          4
#> 51    1       1          5
#> 52    5       1          5
#> 53    0       1          5
#> 54    5       1          5
#> 55    2       1          5
#> 56    1       1          5
#> 57    1       1          5
#> 58    0       1          5
#> 59    3       1          5
#> 60    1       1          5
#> 61    4       1          6
#> 62    0       1          6
#> 63    3       1          6
#> 64    8       1          6
#> 65    3       1          6
#> 66    4       1          6
#> 67    5       1          6
#> 68    3       1          6
#> 69    5       1          6
#> 70    4       1          6
#> 71    3       1          7
#> 72    1       1          7
#> 73    2       1          7
#> 74    1       1          7
#> 75    4       1          7
#> 76    2       1          7
#> 77    5       1          7
#> 78    2       1          7
#> 79    4       1          7
#> 80    1       1          7
#> 81    5       1          8
#> 82    3       1          8
#> 83    4       1          8
#> 84    8       1          8
#> 85    4       1          8
#> 86    4       1          8
#> 87    4       1          8
#> 88    1       1          8
#> 89    3       1          8
#> 90    6       1          8
#> 91    1       1          9
#> 92    3       1          9
#> 93    5       1          9
#> 94    3       1          9
#> 95    6       1          9
#> 96    3       1          9
#> 97    3       1          9
#> 98    1       1          9
#> 99    3       1          9
#> 100   4       1          9
#> 101   5       1         10
#> 102   3       1         10
#> 103   4       1         10
#> 104   3       1         10
#> 105   3       1         10
#> 106   7       1         10
#> 107   6       1         10
#> 108   6       1         10
#> 109   5       1         10
#> 110   4       1         10
#> 111   6       1         11
#> 112   7       1         11
#> 113   3       1         11
#> 114   3       1         11
#> 115   5       1         11
#> 116   4       1         11
#> 117   5       1         11
#> 118   3       1         11
#> 119   4       1         11
#> 120   5       1         11
#> 121   4       1         12
#> 122   3       1         12
#> 123   5       1         12
#> 124   6       1         12
#> 125   6       1         12
#> 126   3       1         12
#> 127   5       1         12
#> 128   4       1         12
#> 129   4       1         12
#> 130   6       1         12
#> 131   9       1         13
#> 132   5       1         13
#> 133   5       1         13
#> 134   2       1         13
#> 135   5       1         13
#> 136   7       1         13
#> 137   8       1         13
#> 138   6       1         13
#> 139   3       1         13
#> 140   6       1         13
#> 141   5       1         14
#> 142   4       1         14
#> 143   5       1         14
#> 144   6       1         14
#> 145   7       1         14
#> 146   6       1         14
#> 147   5       1         14
#> 148   3       1         14
#> 149   7       1         14
#> 150   5       1         14
#> 151   7       1         15
#> 152   6       1         15
#> 153   8       1         15
#> 154  10       1         15
#> 155   5       1         15
#> 156   6       1         15
#> 157   4       1         15
#> 158   9       1         15
#> 159   8       1         15
#> 160   4       1         15
#> 161  10       1         16
#> 162   3       1         16
#> 163  10       1         16
#> 164   5       1         16
#> 165   4       1         16
#> 166   6       1         16
#> 167   6       1         16
#> 168   6       1         16
#> 169   7       1         16
#> 170   4       1         16
#> 171  10       1         17
#> 172   6       1         17
#> 173   5       1         17
#> 174   9       1         17
#> 175   6       1         17
#> 176   3       1         17
#> 177   8       1         17
#> 178   9       1         17
#> 179   5       1         17
#> 180   9       1         17
#> 181   9       1         18
#> 182   9       1         18
#> 183   6       1         18
#> 184   6       1         18
#> 185   3       1         18
#> 186   5       1         18
#> 187   7       1         18
#> 188   7       1         18
#> 189  13       1         18
#> 190   9       1         18
#> 191   5       1         19
#> 192   5       1         19
#> 193  10       1         19
#> 194   9       1         19
#> 195   8       1         19
#> 196  10       1         19
#> 197   5       1         19
#> 198   8       1         19
#> 199  10       1         19
#> 200   6       1         19
#> 201   4       1         20
#> 202  15       1         20
#> 203  10       1         20
#> 204  11       1         20
#> 205   5       1         20
#> 206   3       1         20
#> 207  11       1         20
#> 208  10       1         20
#> 209   7       1         20
#> 210   8       1         20
#> 211  10       1         21
#> 212   8       1         21
#> 213   7       1         21
#> 214   9       1         21
#> 215  11       1         21
#> 216   9       1         21
#> 217   9       1         21
#> 218  13       1         21
#> 219  10       1         21
#> 220   9       1         21
#> 221   7       1         22
#> 222   5       1         22
#> 223   9       1         22
#> 224   9       1         22
#> 225   7       1         22
#> 226   8       1         22
#> 227  10       1         22
#> 228   9       1         22
#> 229   5       1         22
#> 230   6       1         22
#> 231  11       1         23
#> 232   8       1         23
#> 233  10       1         23
#> 234   3       1         23
#> 235   9       1         23
#> 236  12       1         23
#> 237   9       1         23
#> 238   7       1         23
#> 239   7       1         23
#> 240  11       1         23
#> 241  13       1         24
#> 242   8       1         24
#> 243   7       1         24
#> 244  13       1         24
#> 245   9       1         24
#> 246   9       1         24
#> 247   9       1         24
#> 248  11       1         24
#> 249  12       1         24
#> 250  14       1         24
#> 251  11       1         25
#> 252  10       1         25
#> 253  12       1         25
#> 254   9       1         25
#> 255  15       1         25
#> 256  12       1         25
#> 257  14       1         25
#> 258   9       1         25
#> 259  16       1         25
#> 260  13       1         25
#> 261  12       1         26
#> 262  12       1         26
#> 263  13       1         26
#> 264  11       1         26
#> 265   9       1         26
#> 266  10       1         26
#> 267  11       1         26
#> 268  12       1         26
#> 269   8       1         26
#> 270   8       1         26
#> 271  12       1         27
#> 272  12       1         27
#> 273  12       1         27
#> 274  12       1         27
#> 275   8       1         27
#> 276  13       1         27
#> 277   8       1         27
#> 278  13       1         27
#> 279  12       1         27
#> 280  11       1         27
#> 281   9       1         28
#> 282  12       1         28
#> 283  14       1         28
#> 284  11       1         28
#> 285   9       1         28
#> 286  13       1         28
#> 287   8       1         28
#> 288   6       1         28
#> 289  10       1         28
#> 290  14       1         28
#> 291  12       1         29
#> 292  14       1         29
#> 293   9       1         29
#> 294  13       1         29
#> 295  12       1         29
#> 296  10       1         29
#> 297  19       1         29
#> 298  10       1         29
#> 299  15       1         29
#> 300   8       1         29
#> 301   9       1         30
#> 302   9       1         30
#> 303  10       1         30
#> 304  14       1         30
#> 305  10       1         30
#> 306  16       1         30
#> 307  15       1         30
#> 308  10       1         30
#> 309  14       1         30
#> 310  18       1         30
#> 311  14       1         31
#> 312  13       1         31
#> 313  13       1         31
#> 314  12       1         31
#> 315  13       1         31
#> 316  12       1         31
#> 317   7       1         31
#> 318  14       1         31
#> 319  12       1         31
#> 320   9       1         31
#> 321   8       1         32
#> 322  16       1         32
#> 323  13       1         32
#> 324  10       1         32
#> 325  10       1         32
#> 326  17       1         32
#> 327  13       1         32
#> 328  14       1         32
#> 329  13       1         32
#> 330  17       1         32
#> 331  10       1         33
#> 332  10       1         33
#> 333  10       1         33
#> 334  17       1         33
#> 335  11       1         33
#> 336  15       1         33
#> 337  18       1         33
#> 338  13       1         33
#> 339  11       1         33
#> 340   9       1         33
#> 341  14       1         34
#> 342  13       1         34
#> 343  20       1         34
#> 344  14       1         34
#> 345  12       1         34
#> 346  11       1         34
#> 347  13       1         34
#> 348  15       1         34
#> 349  12       1         34
#> 350  11       1         34
#> 351  10       1         35
#> 352  16       1         35
#> 353  10       1         35
#> 354  14       1         35
#> 355  15       1         35
#> 356  13       1         35
#> 357  15       1         35
#> 358  12       1         35
#> 359  11       1         35
#> 360  12       1         35
#> 361  12       1         36
#> 362  12       1         36
#> 363  18       1         36
#> 364  17       1         36
#> 365  14       1         36
#> 366  14       1         36
#> 367  17       1         36
#> 368  16       1         36
#> 369  16       1         36
#> 370  13       1         36
#> 371  16       1         37
#> 372  11       1         37
#> 373  13       1         37
#> 374  16       1         37
#> 375  14       1         37
#> 376  18       1         37
#> 377  19       1         37
#> 378  14       1         37
#> 379  11       1         37
#> 380  17       1         37
#> 381  19       1         38
#> 382  14       1         38
#> 383  15       1         38
#> 384  18       1         38
#> 385  14       1         38
#> 386   9       1         38
#> 387  19       1         38
#> 388  15       1         38
#> 389  15       1         38
#> 390  16       1         38
#> 391  15       1         39
#> 392  18       1         39
#> 393  22       1         39
#> 394  22       1         39
#> 395  15       1         39
#> 396  20       1         39
#> 397  16       1         39
#> 398  16       1         39
#> 399  17       1         39
#> 400  19       1         39
#> 401  17       1         40
#> 402  17       1         40
#> 403  19       1         40
#> 404  17       1         40
#> 405  16       1         40
#> 406  13       1         40
#> 407  13       1         40
#> 408  19       1         40
#> 409  18       1         40
#> 410  15       1         40
#> 411  16       1         41
#> 412  18       1         41
#> 413  15       1         41
#> 414  21       1         41
#> 415  14       1         41
#> 416  16       1         41
#> 417  13       1         41
#> 418  18       1         41
#> 419  13       1         41
#> 420  21       1         41
#> 421  17       1         42
#> 422  20       1         42
#> 423  17       1         42
#> 424  18       1         42
#> 425  15       1         42
#> 426  23       1         42
#> 427  19       1         42
#> 428  19       1         42
#> 429  21       1         42
#> 430  20       1         42
#> 431  20       1         43
#> 432  18       1         43
#> 433  20       1         43
#> 434  10       1         43
#> 435  11       1         43
#> 436  23       1         43
#> 437  13       1         43
#> 438  21       1         43
#> 439  15       1         43
#> 440  17       1         43
#> 441  21       1         44
#> 442  19       1         44
#> 443  16       1         44
#> 444  16       1         44
#> 445  20       1         44
#> 446  16       1         44
#> 447  18       1         44
#> 448  19       1         44
#> 449  17       1         44
#> 450  15       1         44
#> 451  16       1         45
#> 452  22       1         45
#> 453  20       1         45
#> 454  19       1         45
#> 455  17       1         45
#> 456  14       1         45
#> 457  19       1         45
#> 458  13       1         45
#> 459  16       1         45
#> 460  16       1         45
#> 461  20       1         46
#> 462  16       1         46
#> 463  15       1         46
#> 464  15       1         46
#> 465  23       1         46
#> 466  22       1         46
#> 467  16       1         46
#> 468  21       1         46
#> 469  18       1         46
#> 470  17       1         46
#> 471  18       1         47
#> 472  20       1         47
#> 473  20       1         47
#> 474  19       1         47
#> 475  19       1         47
#> 476  23       1         47
#> 477  23       1         47
#> 478  20       1         47
#> 479  12       1         47
#> 480  17       1         47
#> 481  18       1         48
#> 482  20       1         48
#> 483  22       1         48
#> 484  18       1         48
#> 485  18       1         48
#> 486  21       1         48
#> 487  20       1         48
#> 488  24       1         48
#> 489  21       1         48
#> 490  18       1         48
#> 491  20       1         49
#> 492  21       1         49
#> 493  17       1         49
#> 494  24       1         49
#> 495  17       1         49
#> 496  18       1         49
#> 497  21       1         49
#> 498  16       1         49
#> 499  21       1         49
#> 500  17       1         49
#> 501  15       1         50
#> 502  16       1         50
#> 503  21       1         50
#> 504  18       1         50
#> 505  16       1         50
#> 506  20       1         50
#> 507  17       1         50
#> 508  23       1         50
#> 509  17       1         50
#> 510  22       1         50
#> 511  20       1         51
#> 512  19       1         51
#> 513  18       1         51
#> 514  25       2         51
#> 515  22       1         51
#> 516  19       1         51
#> 517  17       1         51
#> 518  24       1         51
#> 519  22       1         51
#> 520  21       1         51
#> 521  19       1         52
#> 522  24       1         52
#> 523  24       1         52
#> 524  16       1         52
#> 525  17       1         52
#> 526  22       1         52
#> 527  19       1         52
#> 528  16       1         52
#> 529  21       1         52
#> 530  20       1         52
#> 531  26       2         53
#> 532  18       1         53
#> 533  21       1         53
#> 534  17       1         53
#> 535  21       1         53
#> 536  17       1         53
#> 537  23       1         53
#> 538  19       1         53
#> 539  22       1         53
#> 540  19       1         53
#> 541  26       2         54
#> 542  25       2         54
#> 543  25       2         54
#> 544  24       1         54
#> 545  25       2         54
#> 546  24       1         54
#> 547  22       1         54
#> 548  19       1         54
#> 549  24       1         54
#> 550  20       1         54
#> 551  19       1         55
#> 552  23       1         55
#> 553  25       2         55
#> 554  22       1         55
#> 555  22       1         55
#> 556  18       1         55
#> 557  20       1         55
#> 558  28       2         55
#> 559  20       1         55
#> 560  21       1         55
#> 561  21       1         56
#> 562  22       1         56
#> 563  24       1         56
#> 564  18       1         56
#> 565  25       2         56
#> 566  23       1         56
#> 567  23       1         56
#> 568  21       1         56
#> 569  22       1         56
#> 570  20       1         56
#> 571  18       1         57
#> 572  24       1         57
#> 573  23       1         57
#> 574  20       1         57
#> 575  29       2         57
#> 576  23       1         57
#> 577  24       1         57
#> 578  21       1         57
#> 579  22       1         57
#> 580  27       2         57
#> 581  23       1         58
#> 582  21       1         58
#> 583  25       2         58
#> 584  29       2         58
#> 585  25       2         58
#> 586  23       1         58
#> 587  23       1         58
#> 588  26       2         58
#> 589  21       1         58
#> 590  24       1         58
#> 591  23       1         59
#> 592  22       1         59
#> 593  25       2         59
#> 594  26       2         59
#> 595  26       2         59
#> 596  20       1         59
#> 597  21       1         59
#> 598  24       1         59
#> 599  23       1         59
#> 600  26       2         59
#> 601  23       1         60
#> 602  28       2         60
#> 603  20       1         60
#> 604  26       2         60
#> 605  21       1         60
#> 606  25       2         60
#> 607  24       1         60
#> 608  25       2         60
#> 609  26       2         60
#> 610  22       1         60
#> 611  30       2         61
#> 612  23       1         61
#> 613  21       1         61
#> 614  20       1         61
#> 615  28       2         61
#> 616  27       2         61
#> 617  21       1         61
#> 618  25       2         61
#> 619  27       2         61
#> 620  21       1         61
#> 621  23       1         62
#> 622  28       2         62
#> 623  27       2         62
#> 624  22       1         62
#> 625  20       1         62
#> 626  26       2         62
#> 627  25       2         62
#> 628  23       1         62
#> 629  33       2         62
#> 630  30       2         62
#> 631  29       2         63
#> 632  24       1         63
#> 633  22       1         63
#> 634  22       1         63
#> 635  26       2         63
#> 636  25       2         63
#> 637  33       2         63
#> 638  27       2         63
#> 639  25       2         63
#> 640  24       1         63
#> 641  30       2         64
#> 642  22       1         64
#> 643  26       2         64
#> 644  29       2         64
#> 645  28       2         64
#> 646  31       2         64
#> 647  28       2         64
#> 648  20       1         64
#> 649  29       2         64
#> 650  30       2         64
#> 651  22       1         65
#> 652  33       2         65
#> 653  24       1         65
#> 654  23       1         65
#> 655  27       2         65
#> 656  27       2         65
#> 657  24       1         65
#> 658  24       1         65
#> 659  27       2         65
#> 660  28       2         65
#> 661  21       1         66
#> 662  25       2         66
#> 663  28       2         66
#> 664  29       2         66
#> 665  28       2         66
#> 666  24       1         66
#> 667  24       1         66
#> 668  27       2         66
#> 669  26       2         66
#> 670  28       2         66
#> 671  30       2         67
#> 672  29       2         67
#> 673  29       2         67
#> 674  25       2         67
#> 675  25       2         67
#> 676  31       2         67
#> 677  26       2         67
#> 678  23       1         67
#> 679  27       2         67
#> 680  27       2         67
#> 681  29       2         68
#> 682  26       2         68
#> 683  30       2         68
#> 684  27       2         68
#> 685  28       2         68
#> 686  28       2         68
#> 687  30       2         68
#> 688  32       2         68
#> 689  27       2         68
#> 690  30       2         68
#> 691  29       2         69
#> 692  31       2         69
#> 693  22       1         69
#> 694  23       1         69
#> 695  26       2         69
#> 696  28       2         69
#> 697  31       2         69
#> 698  31       2         69
#> 699  26       2         69
#> 700  30       2         69
#> 701  32       2         70
#> 702  27       2         70
#> 703  31       2         70
#> 704  27       2         70
#> 705  35       2         70
#> 706  29       2         70
#> 707  29       2         70
#> 708  30       2         70
#> 709  25       2         70
#> 710  33       2         70
#> 711  30       2         71
#> 712  28       2         71
#> 713  31       2         71
#> 714  28       2         71
#> 715  29       2         71
#> 716  28       2         71
#> 717  26       2         71
#> 718  31       2         71
#> 719  26       2         71
#> 720  34       2         71
#> 721  35       2         72
#> 722  33       2         72
#> 723  31       2         72
#> 724  29       2         72
#> 725  30       2         72
#> 726  29       2         72
#> 727  26       2         72
#> 728  34       2         72
#> 729  29       2         72
#> 730  34       2         72
#> 731  28       2         73
#> 732  26       2         73
#> 733  30       2         73
#> 734  27       2         73
#> 735  33       2         73
#> 736  28       2         73
#> 737  27       2         73
#> 738  30       2         73
#> 739  31       2         73
#> 740  27       2         73
#> 741  29       2         74
#> 742  29       2         74
#> 743  29       2         74
#> 744  28       2         74
#> 745  31       2         74
#> 746  28       2         74
#> 747  30       2         74
#> 748  34       2         74
#> 749  28       2         74
#> 750  22       1         74
#> 751  29       2         75
#> 752  35       2         75
#> 753  30       2         75
#> 754  32       2         75
#> 755  37       3         75
#> 756  29       2         75
#> 757  28       2         75
#> 758  34       2         75
#> 759  33       2         75
#> 760  28       2         75
#> 761  30       2         76
#> 762  29       2         76
#> 763  29       2         76
#> 764  29       2         76
#> 765  34       2         76
#> 766  33       2         76
#> 767  29       2         76
#> 768  29       2         76
#> 769  32       2         76
#> 770  27       2         76
#> 771  32       2         77
#> 772  35       2         77
#> 773  32       2         77
#> 774  35       2         77
#> 775  33       2         77
#> 776  32       2         77
#> 777  29       2         77
#> 778  31       2         77
#> 779  31       2         77
#> 780  26       2         77
#> 781  29       2         78
#> 782  36       2         78
#> 783  31       2         78
#> 784  30       2         78
#> 785  35       2         78
#> 786  31       2         78
#> 787  34       2         78
#> 788  31       2         78
#> 789  31       2         78
#> 790  32       2         78
#> 791  33       2         79
#> 792  31       2         79
#> 793  32       2         79
#> 794  30       2         79
#> 795  33       2         79
#> 796  25       2         79
#> 797  34       2         79
#> 798  31       2         79
#> 799  34       2         79
#> 800  31       2         79
#> 801  30       2         80
#> 802  30       2         80
#> 803  34       2         80
#> 804  32       2         80
#> 805  31       2         80
#> 806  25       2         80
#> 807  32       2         80
#> 808  34       2         80
#> 809  32       2         80
#> 810  31       2         80
#> 811  31       2         81
#> 812  32       2         81
#> 813  32       2         81
#> 814  38       3         81
#> 815  31       2         81
#> 816  35       2         81
#> 817  33       2         81
#> 818  36       2         81
#> 819  35       2         81
#> 820  31       2         81
#> 821  31       2         82
#> 822  36       2         82
#> 823  33       2         82
#> 824  34       2         82
#> 825  30       2         82
#> 826  28       2         82
#> 827  31       2         82
#> 828  30       2         82
#> 829  32       2         82
#> 830  36       2         82
#> 831  37       3         83
#> 832  31       2         83
#> 833  33       2         83
#> 834  35       2         83
#> 835  35       2         83
#> 836  32       2         83
#> 837  36       2         83
#> 838  37       3         83
#> 839  32       2         83
#> 840  35       2         83
#> 841  35       2         84
#> 842  35       2         84
#> 843  33       2         84
#> 844  35       2         84
#> 845  37       3         84
#> 846  31       2         84
#> 847  32       2         84
#> 848  38       3         84
#> 849  27       2         84
#> 850  34       2         84
#> 851  36       2         85
#> 852  39       3         85
#> 853  36       2         85
#> 854  34       2         85
#> 855  36       2         85
#> 856  34       2         85
#> 857  36       2         85
#> 858  32       2         85
#> 859  36       2         85
#> 860  34       2         85
#> 861  30       2         86
#> 862  34       2         86
#> 863  36       2         86
#> 864  33       2         86
#> 865  33       2         86
#> 866  32       2         86
#> 867  37       3         86
#> 868  35       2         86
#> 869  36       2         86
#> 870  35       2         86
#> 871  38       3         87
#> 872  38       3         87
#> 873  38       3         87
#> 874  32       2         87
#> 875  35       2         87
#> 876  32       2         87
#> 877  33       2         87
#> 878  32       2         87
#> 879  31       2         87
#> 880  40       3         87
#> 881  32       2         88
#> 882  34       2         88
#> 883  37       3         88
#> 884  33       2         88
#> 885  34       2         88
#> 886  33       2         88
#> 887  34       2         88
#> 888  38       3         88
#> 889  32       2         88
#> 890  37       3         88
#> 891  33       2         89
#> 892  36       2         89
#> 893  36       2         89
#> 894  33       2         89
#> 895  33       2         89
#> 896  37       3         89
#> 897  37       3         89
#> 898  38       3         89
#> 899  32       2         89
#> 900  37       3         89
#> 901  39       3         90
#> 902  34       2         90
#> 903  38       3         90
#> 904  40       3         90
#> 905  37       3         90
#> 906  38       3         90
#> 907  36       2         90
#> 908  36       2         90
#> 909  35       2         90
#> 910  35       2         90
#> 911  33       2         91
#> 912  34       2         91
#> 913  33       2         91
#> 914  36       2         91
#> 915  35       2         91
#> 916  36       2         91
#> 917  39       3         91
#> 918  36       2         91
#> 919  38       3         91
#> 920  37       3         91
#> 921  38       3         92
#> 922  38       3         92
#> 923  38       3         92
#> 924  36       2         92
#> 925  35       2         92
#> 926  38       3         92
#> 927  37       3         92
#> 928  39       3         92
#> 929  38       3         92
#> 930  39       3         92
#> 931  39       3         93
#> 932  39       3         93
#> 933  39       3         93
#> 934  34       2         93
#> 935  39       3         93
#> 936  35       2         93
#> 937  36       2         93
#> 938  35       2         93
#> 939  35       2         93
#> 940  38       3         93
#> 941  38       3         94
#> 942  38       3         94
#> 943  39       3         94
#> 944  37       3         94
#> 945  36       2         94
#> 946  34       2         94
#> 947  38       3         94
#> 948  37       3         94
#> 949  40       3         94
#> 950  38       3         94
#> 951  38       3         95
#> 952  34       2         95
#> 953  37       3         95
#> 954  36       2         95
#> 955  39       3         95
#> 956  36       2         95
#> 957  39       3         95
#> 958  38       3         95
#> 959  37       3         95
#> 960  39       3         95
#> 961  39       3         96
#> 962  37       3         96
#> 963  39       3         96
#> 964  38       3         96
#> 965  37       3         96
#> 966  39       3         96
#> 967  38       3         96
#> 968  38       3         96
#> 969  38       3         96
#> 970  37       3         96
#> 971  39       3         97
#> 972  39       3         97
#> 973  37       3         97
#> 974  39       3         97
#> 975  39       3         97
#> 976  39       3         97
#> 977  39       3         97
#> 978  39       3         97
#> 979  39       3         97
#> 980  39       3         97
#> 981  38       3         98
#> 982  39       3         98
#> 983  39       3         98
#> 984  39       3         98
#> 985  38       3         98
#> 986  39       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  40       3         98
#> 991  40       3         99
#> 992  40       3         99
#> 993  39       3         99
#> 994  40       3         99
#> 995  39       3         99
#> 996  39       3         99
#> 997  40       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100