Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    1       1          1
#> 13    0       1          1
#> 14    0       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    1       1          1
#> 21    0       1          2
#> 22    1       1          2
#> 23    0       1          2
#> 24    2       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    1       1          2
#> 28    1       1          2
#> 29    1       1          2
#> 30    1       1          2
#> 31    2       1          3
#> 32    1       1          3
#> 33    1       1          3
#> 34    1       1          3
#> 35    1       1          3
#> 36    2       1          3
#> 37    2       1          3
#> 38    1       1          3
#> 39    1       1          3
#> 40    1       1          3
#> 41    0       1          4
#> 42    1       1          4
#> 43    1       1          4
#> 44    1       1          4
#> 45    2       1          4
#> 46    1       1          4
#> 47    0       1          4
#> 48    0       1          4
#> 49    4       1          4
#> 50    0       1          4
#> 51    0       1          5
#> 52    1       1          5
#> 53    0       1          5
#> 54    1       1          5
#> 55    1       1          5
#> 56    1       1          5
#> 57    0       1          5
#> 58    2       1          5
#> 59    1       1          5
#> 60    1       1          5
#> 61    1       1          6
#> 62    3       1          6
#> 63    2       1          6
#> 64    1       1          6
#> 65    5       1          6
#> 66    6       1          6
#> 67    2       1          6
#> 68    3       1          6
#> 69    2       1          6
#> 70    4       1          6
#> 71    2       1          7
#> 72    4       1          7
#> 73    1       1          7
#> 74    6       1          7
#> 75    2       1          7
#> 76    2       1          7
#> 77    3       1          7
#> 78    4       1          7
#> 79    2       1          7
#> 80    0       1          7
#> 81    3       1          8
#> 82    6       1          8
#> 83    2       1          8
#> 84    4       1          8
#> 85    0       1          8
#> 86    2       1          8
#> 87    0       1          8
#> 88    6       1          8
#> 89    3       1          8
#> 90    2       1          8
#> 91    5       1          9
#> 92    4       1          9
#> 93    4       1          9
#> 94    5       1          9
#> 95    3       1          9
#> 96    4       1          9
#> 97    4       1          9
#> 98    8       1          9
#> 99    2       1          9
#> 100   1       1          9
#> 101   4       1         10
#> 102   0       1         10
#> 103   1       1         10
#> 104   5       1         10
#> 105   2       1         10
#> 106   2       1         10
#> 107   4       1         10
#> 108   9       1         10
#> 109   9       1         10
#> 110   6       1         10
#> 111   7       1         11
#> 112   5       1         11
#> 113   3       1         11
#> 114   6       1         11
#> 115   5       1         11
#> 116   3       1         11
#> 117   5       1         11
#> 118   4       1         11
#> 119   6       1         11
#> 120   4       1         11
#> 121   3       1         12
#> 122   6       1         12
#> 123   5       1         12
#> 124   5       1         12
#> 125   3       1         12
#> 126   7       1         12
#> 127   5       1         12
#> 128   4       1         12
#> 129   3       1         12
#> 130   4       1         12
#> 131   7       1         13
#> 132   7       1         13
#> 133   3       1         13
#> 134   4       1         13
#> 135   3       1         13
#> 136   4       1         13
#> 137   4       1         13
#> 138   5       1         13
#> 139  11       1         13
#> 140   9       1         13
#> 141   8       1         14
#> 142   7       1         14
#> 143   8       1         14
#> 144  10       1         14
#> 145   9       1         14
#> 146   1       1         14
#> 147   5       1         14
#> 148   4       1         14
#> 149  11       1         14
#> 150   7       1         14
#> 151   8       1         15
#> 152   5       1         15
#> 153   4       1         15
#> 154   3       1         15
#> 155   3       1         15
#> 156   8       1         15
#> 157   9       1         15
#> 158   9       1         15
#> 159   4       1         15
#> 160   8       1         15
#> 161   8       1         16
#> 162   5       1         16
#> 163   3       1         16
#> 164   5       1         16
#> 165   5       1         16
#> 166   7       1         16
#> 167   8       1         16
#> 168   8       1         16
#> 169   9       1         16
#> 170   8       1         16
#> 171   5       1         17
#> 172   4       1         17
#> 173   7       1         17
#> 174   6       1         17
#> 175  10       1         17
#> 176  10       1         17
#> 177  10       1         17
#> 178   8       1         17
#> 179   6       1         17
#> 180   6       1         17
#> 181   9       1         18
#> 182   6       1         18
#> 183   8       1         18
#> 184   7       1         18
#> 185   5       1         18
#> 186   8       1         18
#> 187   6       1         18
#> 188   7       1         18
#> 189   7       1         18
#> 190   5       1         18
#> 191   8       1         19
#> 192   9       1         19
#> 193  10       1         19
#> 194  10       1         19
#> 195   5       1         19
#> 196   7       1         19
#> 197   8       1         19
#> 198   7       1         19
#> 199   7       1         19
#> 200   7       1         19
#> 201   6       1         20
#> 202   7       1         20
#> 203   8       1         20
#> 204   9       1         20
#> 205   5       1         20
#> 206   8       1         20
#> 207  10       1         20
#> 208   8       1         20
#> 209   9       1         20
#> 210   5       1         20
#> 211   8       1         21
#> 212   8       1         21
#> 213  11       1         21
#> 214  11       1         21
#> 215  12       1         21
#> 216   6       1         21
#> 217   5       1         21
#> 218  13       1         21
#> 219  10       1         21
#> 220   7       1         21
#> 221  10       1         22
#> 222   9       1         22
#> 223   9       1         22
#> 224   6       1         22
#> 225  10       1         22
#> 226   8       1         22
#> 227  12       1         22
#> 228  11       1         22
#> 229  10       1         22
#> 230   9       1         22
#> 231  12       1         23
#> 232   7       1         23
#> 233   7       1         23
#> 234  13       1         23
#> 235   8       1         23
#> 236  16       1         23
#> 237   9       1         23
#> 238  10       1         23
#> 239   8       1         23
#> 240   4       1         23
#> 241   7       1         24
#> 242  10       1         24
#> 243  13       1         24
#> 244  10       1         24
#> 245   9       1         24
#> 246   6       1         24
#> 247   6       1         24
#> 248  10       1         24
#> 249   9       1         24
#> 250  11       1         24
#> 251  15       1         25
#> 252  11       1         25
#> 253   8       1         25
#> 254   8       1         25
#> 255  12       1         25
#> 256   9       1         25
#> 257   9       1         25
#> 258   8       1         25
#> 259  14       1         25
#> 260  10       1         25
#> 261   7       1         26
#> 262   5       1         26
#> 263   8       1         26
#> 264  14       1         26
#> 265   9       1         26
#> 266   9       1         26
#> 267   7       1         26
#> 268   9       1         26
#> 269   8       1         26
#> 270  11       1         26
#> 271  13       1         27
#> 272  14       1         27
#> 273  11       1         27
#> 274  15       1         27
#> 275  13       1         27
#> 276  10       1         27
#> 277   9       1         27
#> 278   9       1         27
#> 279  15       1         27
#> 280   9       1         27
#> 281   8       1         28
#> 282  12       1         28
#> 283  13       1         28
#> 284   4       1         28
#> 285   8       1         28
#> 286   7       1         28
#> 287   8       1         28
#> 288  13       1         28
#> 289  11       1         28
#> 290  15       1         28
#> 291  11       1         29
#> 292  12       1         29
#> 293  15       1         29
#> 294   8       1         29
#> 295  14       1         29
#> 296   9       1         29
#> 297  12       1         29
#> 298  12       1         29
#> 299   5       1         29
#> 300  15       1         29
#> 301  20       1         30
#> 302  10       1         30
#> 303  10       1         30
#> 304   6       1         30
#> 305  10       1         30
#> 306   8       1         30
#> 307   9       1         30
#> 308  16       1         30
#> 309   9       1         30
#> 310  13       1         30
#> 311  12       1         31
#> 312  11       1         31
#> 313  18       1         31
#> 314  11       1         31
#> 315  13       1         31
#> 316  12       1         31
#> 317  14       1         31
#> 318  18       1         31
#> 319  14       1         31
#> 320  15       1         31
#> 321  10       1         32
#> 322  13       1         32
#> 323  10       1         32
#> 324  16       1         32
#> 325  11       1         32
#> 326  23       1         32
#> 327  13       1         32
#> 328  11       1         32
#> 329   9       1         32
#> 330  11       1         32
#> 331  14       1         33
#> 332  11       1         33
#> 333  15       1         33
#> 334   8       1         33
#> 335  13       1         33
#> 336  13       1         33
#> 337  13       1         33
#> 338  16       1         33
#> 339  19       1         33
#> 340  11       1         33
#> 341  13       1         34
#> 342  12       1         34
#> 343  14       1         34
#> 344  14       1         34
#> 345  20       1         34
#> 346  14       1         34
#> 347  13       1         34
#> 348  11       1         34
#> 349  16       1         34
#> 350  11       1         34
#> 351  21       1         35
#> 352  17       1         35
#> 353  16       1         35
#> 354  14       1         35
#> 355  18       1         35
#> 356  11       1         35
#> 357  11       1         35
#> 358  17       1         35
#> 359  14       1         35
#> 360  13       1         35
#> 361  20       1         36
#> 362  13       1         36
#> 363  21       1         36
#> 364  11       1         36
#> 365  16       1         36
#> 366  18       1         36
#> 367  14       1         36
#> 368  16       1         36
#> 369  20       1         36
#> 370  19       1         36
#> 371  12       1         37
#> 372  17       1         37
#> 373  19       1         37
#> 374  18       1         37
#> 375  17       1         37
#> 376  17       1         37
#> 377  14       1         37
#> 378  13       1         37
#> 379  12       1         37
#> 380   7       1         37
#> 381  18       1         38
#> 382  20       1         38
#> 383  19       1         38
#> 384  16       1         38
#> 385  20       1         38
#> 386  14       1         38
#> 387  15       1         38
#> 388  10       1         38
#> 389  19       1         38
#> 390  15       1         38
#> 391  14       1         39
#> 392  22       1         39
#> 393  17       1         39
#> 394  14       1         39
#> 395  17       1         39
#> 396  10       1         39
#> 397  17       1         39
#> 398  18       1         39
#> 399  16       1         39
#> 400   7       1         39
#> 401  20       1         40
#> 402  14       1         40
#> 403  19       1         40
#> 404  18       1         40
#> 405  18       1         40
#> 406  17       1         40
#> 407  18       1         40
#> 408  19       1         40
#> 409  14       1         40
#> 410  15       1         40
#> 411  11       1         41
#> 412  15       1         41
#> 413  13       1         41
#> 414  16       1         41
#> 415  18       1         41
#> 416  14       1         41
#> 417  19       1         41
#> 418  15       1         41
#> 419  12       1         41
#> 420  16       1         41
#> 421  12       1         42
#> 422  13       1         42
#> 423  14       1         42
#> 424  19       1         42
#> 425  18       1         42
#> 426  14       1         42
#> 427  17       1         42
#> 428  17       1         42
#> 429  17       1         42
#> 430  16       1         42
#> 431  18       1         43
#> 432  10       1         43
#> 433  23       1         43
#> 434  16       1         43
#> 435  16       1         43
#> 436  17       1         43
#> 437  20       1         43
#> 438  23       1         43
#> 439  18       1         43
#> 440  17       1         43
#> 441  13       1         44
#> 442  19       1         44
#> 443  19       1         44
#> 444  16       1         44
#> 445  18       1         44
#> 446  15       1         44
#> 447  13       1         44
#> 448  20       1         44
#> 449  21       1         44
#> 450  18       1         44
#> 451  19       1         45
#> 452  20       1         45
#> 453  17       1         45
#> 454  16       1         45
#> 455  15       1         45
#> 456  13       1         45
#> 457  17       1         45
#> 458  20       1         45
#> 459  20       1         45
#> 460  11       1         45
#> 461  17       1         46
#> 462  15       1         46
#> 463  19       1         46
#> 464  17       1         46
#> 465  24       1         46
#> 466  16       1         46
#> 467  19       1         46
#> 468  16       1         46
#> 469  22       1         46
#> 470  18       1         46
#> 471  17       1         47
#> 472  15       1         47
#> 473  20       1         47
#> 474  17       1         47
#> 475  20       1         47
#> 476  21       1         47
#> 477  16       1         47
#> 478  22       1         47
#> 479  18       1         47
#> 480  16       1         47
#> 481  21       1         48
#> 482  21       1         48
#> 483  23       1         48
#> 484  18       1         48
#> 485  19       1         48
#> 486  23       1         48
#> 487  22       1         48
#> 488  18       1         48
#> 489  18       1         48
#> 490  18       1         48
#> 491  22       1         49
#> 492  20       1         49
#> 493  19       1         49
#> 494  23       1         49
#> 495  23       1         49
#> 496  19       1         49
#> 497  21       1         49
#> 498  18       1         49
#> 499  18       1         49
#> 500  16       1         49
#> 501  18       1         50
#> 502  15       1         50
#> 503  17       1         50
#> 504  13       1         50
#> 505  22       1         50
#> 506  17       1         50
#> 507  22       1         50
#> 508  19       1         50
#> 509  21       1         50
#> 510  26       2         50
#> 511  18       1         51
#> 512  23       1         51
#> 513  20       1         51
#> 514  20       1         51
#> 515  21       1         51
#> 516  21       1         51
#> 517  21       1         51
#> 518  23       1         51
#> 519  21       1         51
#> 520  23       1         51
#> 521  23       1         52
#> 522  25       2         52
#> 523  20       1         52
#> 524  18       1         52
#> 525  24       1         52
#> 526  18       1         52
#> 527  22       1         52
#> 528  21       1         52
#> 529  19       1         52
#> 530  24       1         52
#> 531  19       1         53
#> 532  21       1         53
#> 533  24       1         53
#> 534  21       1         53
#> 535  23       1         53
#> 536  22       1         53
#> 537  23       1         53
#> 538  21       1         53
#> 539  20       1         53
#> 540  18       1         53
#> 541  19       1         54
#> 542  20       1         54
#> 543  27       2         54
#> 544  24       1         54
#> 545  16       1         54
#> 546  20       1         54
#> 547  25       2         54
#> 548  25       2         54
#> 549  24       1         54
#> 550  22       1         54
#> 551  24       1         55
#> 552  25       2         55
#> 553  24       1         55
#> 554  17       1         55
#> 555  19       1         55
#> 556  25       2         55
#> 557  22       1         55
#> 558  24       1         55
#> 559  19       1         55
#> 560  20       1         55
#> 561  23       1         56
#> 562  23       1         56
#> 563  31       2         56
#> 564  22       1         56
#> 565  19       1         56
#> 566  24       1         56
#> 567  23       1         56
#> 568  26       2         56
#> 569  24       1         56
#> 570  25       2         56
#> 571  22       1         57
#> 572  22       1         57
#> 573  20       1         57
#> 574  23       1         57
#> 575  22       1         57
#> 576  20       1         57
#> 577  22       1         57
#> 578  26       2         57
#> 579  23       1         57
#> 580  26       2         57
#> 581  23       1         58
#> 582  21       1         58
#> 583  29       2         58
#> 584  19       1         58
#> 585  29       2         58
#> 586  28       2         58
#> 587  25       2         58
#> 588  24       1         58
#> 589  23       1         58
#> 590  26       2         58
#> 591  24       1         59
#> 592  26       2         59
#> 593  21       1         59
#> 594  22       1         59
#> 595  20       1         59
#> 596  29       2         59
#> 597  16       1         59
#> 598  26       2         59
#> 599  21       1         59
#> 600  24       1         59
#> 601  20       1         60
#> 602  21       1         60
#> 603  26       2         60
#> 604  26       2         60
#> 605  22       1         60
#> 606  26       2         60
#> 607  22       1         60
#> 608  27       2         60
#> 609  26       2         60
#> 610  25       2         60
#> 611  27       2         61
#> 612  27       2         61
#> 613  18       1         61
#> 614  26       2         61
#> 615  27       2         61
#> 616  29       2         61
#> 617  23       1         61
#> 618  28       2         61
#> 619  24       1         61
#> 620  25       2         61
#> 621  23       1         62
#> 622  27       2         62
#> 623  26       2         62
#> 624  27       2         62
#> 625  27       2         62
#> 626  24       1         62
#> 627  23       1         62
#> 628  27       2         62
#> 629  23       1         62
#> 630  26       2         62
#> 631  23       1         63
#> 632  25       2         63
#> 633  29       2         63
#> 634  29       2         63
#> 635  25       2         63
#> 636  25       2         63
#> 637  29       2         63
#> 638  25       2         63
#> 639  29       2         63
#> 640  23       1         63
#> 641  23       1         64
#> 642  32       2         64
#> 643  26       2         64
#> 644  28       2         64
#> 645  29       2         64
#> 646  25       2         64
#> 647  22       1         64
#> 648  31       2         64
#> 649  29       2         64
#> 650  28       2         64
#> 651  24       1         65
#> 652  26       2         65
#> 653  24       1         65
#> 654  31       2         65
#> 655  30       2         65
#> 656  26       2         65
#> 657  23       1         65
#> 658  24       1         65
#> 659  24       1         65
#> 660  24       1         65
#> 661  26       2         66
#> 662  27       2         66
#> 663  27       2         66
#> 664  24       1         66
#> 665  29       2         66
#> 666  27       2         66
#> 667  29       2         66
#> 668  32       2         66
#> 669  23       1         66
#> 670  27       2         66
#> 671  30       2         67
#> 672  28       2         67
#> 673  22       1         67
#> 674  28       2         67
#> 675  29       2         67
#> 676  28       2         67
#> 677  30       2         67
#> 678  28       2         67
#> 679  28       2         67
#> 680  25       2         67
#> 681  28       2         68
#> 682  30       2         68
#> 683  30       2         68
#> 684  24       1         68
#> 685  30       2         68
#> 686  30       2         68
#> 687  32       2         68
#> 688  29       2         68
#> 689  26       2         68
#> 690  30       2         68
#> 691  24       1         69
#> 692  34       2         69
#> 693  29       2         69
#> 694  25       2         69
#> 695  30       2         69
#> 696  24       1         69
#> 697  28       2         69
#> 698  22       1         69
#> 699  34       2         69
#> 700  26       2         69
#> 701  34       2         70
#> 702  27       2         70
#> 703  28       2         70
#> 704  30       2         70
#> 705  29       2         70
#> 706  28       2         70
#> 707  27       2         70
#> 708  34       2         70
#> 709  28       2         70
#> 710  28       2         70
#> 711  26       2         71
#> 712  27       2         71
#> 713  33       2         71
#> 714  22       1         71
#> 715  31       2         71
#> 716  32       2         71
#> 717  29       2         71
#> 718  22       1         71
#> 719  31       2         71
#> 720  27       2         71
#> 721  25       2         72
#> 722  31       2         72
#> 723  34       2         72
#> 724  34       2         72
#> 725  29       2         72
#> 726  21       1         72
#> 727  27       2         72
#> 728  27       2         72
#> 729  32       2         72
#> 730  30       2         72
#> 731  29       2         73
#> 732  25       2         73
#> 733  29       2         73
#> 734  25       2         73
#> 735  30       2         73
#> 736  33       2         73
#> 737  32       2         73
#> 738  32       2         73
#> 739  29       2         73
#> 740  28       2         73
#> 741  28       2         74
#> 742  26       2         74
#> 743  29       2         74
#> 744  32       2         74
#> 745  30       2         74
#> 746  25       2         74
#> 747  34       2         74
#> 748  30       2         74
#> 749  30       2         74
#> 750  30       2         74
#> 751  34       2         75
#> 752  24       1         75
#> 753  30       2         75
#> 754  30       2         75
#> 755  34       2         75
#> 756  32       2         75
#> 757  30       2         75
#> 758  28       2         75
#> 759  30       2         75
#> 760  23       1         75
#> 761  30       2         76
#> 762  21       1         76
#> 763  30       2         76
#> 764  31       2         76
#> 765  29       2         76
#> 766  30       2         76
#> 767  33       2         76
#> 768  30       2         76
#> 769  27       2         76
#> 770  28       2         76
#> 771  32       2         77
#> 772  32       2         77
#> 773  29       2         77
#> 774  31       2         77
#> 775  27       2         77
#> 776  33       2         77
#> 777  33       2         77
#> 778  31       2         77
#> 779  34       2         77
#> 780  23       1         77
#> 781  34       2         78
#> 782  33       2         78
#> 783  30       2         78
#> 784  33       2         78
#> 785  27       2         78
#> 786  30       2         78
#> 787  28       2         78
#> 788  29       2         78
#> 789  34       2         78
#> 790  26       2         78
#> 791  31       2         79
#> 792  32       2         79
#> 793  31       2         79
#> 794  29       2         79
#> 795  30       2         79
#> 796  35       2         79
#> 797  32       2         79
#> 798  31       2         79
#> 799  24       1         79
#> 800  28       2         79
#> 801  32       2         80
#> 802  35       2         80
#> 803  33       2         80
#> 804  31       2         80
#> 805  35       2         80
#> 806  32       2         80
#> 807  27       2         80
#> 808  33       2         80
#> 809  31       2         80
#> 810  30       2         80
#> 811  36       2         81
#> 812  31       2         81
#> 813  29       2         81
#> 814  29       2         81
#> 815  33       2         81
#> 816  32       2         81
#> 817  36       2         81
#> 818  32       2         81
#> 819  33       2         81
#> 820  34       2         81
#> 821  34       2         82
#> 822  35       2         82
#> 823  31       2         82
#> 824  33       2         82
#> 825  33       2         82
#> 826  33       2         82
#> 827  33       2         82
#> 828  30       2         82
#> 829  35       2         82
#> 830  30       2         82
#> 831  31       2         83
#> 832  37       3         83
#> 833  33       2         83
#> 834  35       2         83
#> 835  33       2         83
#> 836  32       2         83
#> 837  34       2         83
#> 838  34       2         83
#> 839  32       2         83
#> 840  32       2         83
#> 841  36       2         84
#> 842  35       2         84
#> 843  36       2         84
#> 844  29       2         84
#> 845  33       2         84
#> 846  35       2         84
#> 847  30       2         84
#> 848  32       2         84
#> 849  29       2         84
#> 850  35       2         84
#> 851  33       2         85
#> 852  34       2         85
#> 853  33       2         85
#> 854  36       2         85
#> 855  33       2         85
#> 856  34       2         85
#> 857  34       2         85
#> 858  36       2         85
#> 859  30       2         85
#> 860  36       2         85
#> 861  34       2         86
#> 862  35       2         86
#> 863  32       2         86
#> 864  35       2         86
#> 865  34       2         86
#> 866  36       2         86
#> 867  37       3         86
#> 868  37       3         86
#> 869  34       2         86
#> 870  38       3         86
#> 871  35       2         87
#> 872  36       2         87
#> 873  32       2         87
#> 874  38       3         87
#> 875  37       3         87
#> 876  35       2         87
#> 877  38       3         87
#> 878  34       2         87
#> 879  36       2         87
#> 880  34       2         87
#> 881  38       3         88
#> 882  34       2         88
#> 883  33       2         88
#> 884  35       2         88
#> 885  37       3         88
#> 886  32       2         88
#> 887  36       2         88
#> 888  36       2         88
#> 889  36       2         88
#> 890  37       3         88
#> 891  34       2         89
#> 892  36       2         89
#> 893  33       2         89
#> 894  35       2         89
#> 895  36       2         89
#> 896  35       2         89
#> 897  36       2         89
#> 898  30       2         89
#> 899  34       2         89
#> 900  36       2         89
#> 901  36       2         90
#> 902  35       2         90
#> 903  34       2         90
#> 904  36       2         90
#> 905  36       2         90
#> 906  37       3         90
#> 907  35       2         90
#> 908  36       2         90
#> 909  39       3         90
#> 910  37       3         90
#> 911  34       2         91
#> 912  33       2         91
#> 913  35       2         91
#> 914  37       3         91
#> 915  36       2         91
#> 916  36       2         91
#> 917  37       3         91
#> 918  35       2         91
#> 919  37       3         91
#> 920  35       2         91
#> 921  36       2         92
#> 922  36       2         92
#> 923  37       3         92
#> 924  35       2         92
#> 925  39       3         92
#> 926  39       3         92
#> 927  36       2         92
#> 928  39       3         92
#> 929  37       3         92
#> 930  38       3         92
#> 931  37       3         93
#> 932  38       3         93
#> 933  35       2         93
#> 934  37       3         93
#> 935  38       3         93
#> 936  39       3         93
#> 937  38       3         93
#> 938  38       3         93
#> 939  35       2         93
#> 940  37       3         93
#> 941  37       3         94
#> 942  38       3         94
#> 943  38       3         94
#> 944  35       2         94
#> 945  37       3         94
#> 946  37       3         94
#> 947  36       2         94
#> 948  40       3         94
#> 949  37       3         94
#> 950  35       2         94
#> 951  37       3         95
#> 952  39       3         95
#> 953  37       3         95
#> 954  39       3         95
#> 955  40       3         95
#> 956  37       3         95
#> 957  40       3         95
#> 958  39       3         95
#> 959  37       3         95
#> 960  38       3         95
#> 961  36       2         96
#> 962  39       3         96
#> 963  37       3         96
#> 964  38       3         96
#> 965  38       3         96
#> 966  40       3         96
#> 967  38       3         96
#> 968  37       3         96
#> 969  38       3         96
#> 970  38       3         96
#> 971  38       3         97
#> 972  40       3         97
#> 973  40       3         97
#> 974  36       2         97
#> 975  40       3         97
#> 976  40       3         97
#> 977  39       3         97
#> 978  39       3         97
#> 979  40       3         97
#> 980  39       3         97
#> 981  37       3         98
#> 982  39       3         98
#> 983  40       3         98
#> 984  39       3         98
#> 985  38       3         98
#> 986  40       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  39       3         98
#> 991  39       3         99
#> 992  40       3         99
#> 993  40       3         99
#> 994  40       3         99
#> 995  39       3         99
#> 996  40       3         99
#> 997  39       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100