Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    1       1          1
#> 13    1       1          1
#> 14    1       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    1       1          1
#> 20    0       1          1
#> 21    1       1          2
#> 22    2       1          2
#> 23    0       1          2
#> 24    1       1          2
#> 25    1       1          2
#> 26    0       1          2
#> 27    1       1          2
#> 28    0       1          2
#> 29    2       1          2
#> 30    0       1          2
#> 31    0       1          3
#> 32    1       1          3
#> 33    1       1          3
#> 34    3       1          3
#> 35    1       1          3
#> 36    2       1          3
#> 37    2       1          3
#> 38    0       1          3
#> 39    1       1          3
#> 40    2       1          3
#> 41    3       1          4
#> 42    2       1          4
#> 43    0       1          4
#> 44    3       1          4
#> 45    1       1          4
#> 46    4       1          4
#> 47    1       1          4
#> 48    3       1          4
#> 49    2       1          4
#> 50    2       1          4
#> 51    0       1          5
#> 52    2       1          5
#> 53    3       1          5
#> 54    0       1          5
#> 55    1       1          5
#> 56    1       1          5
#> 57    3       1          5
#> 58    2       1          5
#> 59    1       1          5
#> 60    0       1          5
#> 61    4       1          6
#> 62    4       1          6
#> 63    2       1          6
#> 64    0       1          6
#> 65    0       1          6
#> 66    1       1          6
#> 67    3       1          6
#> 68    1       1          6
#> 69    2       1          6
#> 70    2       1          6
#> 71    5       1          7
#> 72    1       1          7
#> 73    1       1          7
#> 74    0       1          7
#> 75    4       1          7
#> 76    5       1          7
#> 77    3       1          7
#> 78    4       1          7
#> 79    1       1          7
#> 80    2       1          7
#> 81    3       1          8
#> 82    3       1          8
#> 83    0       1          8
#> 84    1       1          8
#> 85    6       1          8
#> 86    2       1          8
#> 87    4       1          8
#> 88    1       1          8
#> 89    3       1          8
#> 90    3       1          8
#> 91    5       1          9
#> 92    3       1          9
#> 93    5       1          9
#> 94    3       1          9
#> 95    2       1          9
#> 96    6       1          9
#> 97    3       1          9
#> 98    2       1          9
#> 99    4       1          9
#> 100   5       1          9
#> 101   1       1         10
#> 102   3       1         10
#> 103   3       1         10
#> 104   4       1         10
#> 105   3       1         10
#> 106   3       1         10
#> 107   5       1         10
#> 108   6       1         10
#> 109   6       1         10
#> 110   1       1         10
#> 111   7       1         11
#> 112   5       1         11
#> 113   3       1         11
#> 114   4       1         11
#> 115   7       1         11
#> 116   8       1         11
#> 117   8       1         11
#> 118   5       1         11
#> 119   0       1         11
#> 120   2       1         11
#> 121   2       1         12
#> 122   8       1         12
#> 123   4       1         12
#> 124   2       1         12
#> 125   4       1         12
#> 126   4       1         12
#> 127   3       1         12
#> 128   9       1         12
#> 129   3       1         12
#> 130   3       1         12
#> 131   4       1         13
#> 132   8       1         13
#> 133   1       1         13
#> 134   9       1         13
#> 135   9       1         13
#> 136   3       1         13
#> 137   5       1         13
#> 138   6       1         13
#> 139   3       1         13
#> 140   6       1         13
#> 141   5       1         14
#> 142   6       1         14
#> 143   5       1         14
#> 144   9       1         14
#> 145   5       1         14
#> 146   6       1         14
#> 147   3       1         14
#> 148   5       1         14
#> 149   7       1         14
#> 150   7       1         14
#> 151   9       1         15
#> 152   6       1         15
#> 153   6       1         15
#> 154   8       1         15
#> 155   4       1         15
#> 156   3       1         15
#> 157   2       1         15
#> 158   7       1         15
#> 159   7       1         15
#> 160   8       1         15
#> 161   7       1         16
#> 162   3       1         16
#> 163   5       1         16
#> 164   8       1         16
#> 165   2       1         16
#> 166   5       1         16
#> 167   4       1         16
#> 168   4       1         16
#> 169   5       1         16
#> 170   2       1         16
#> 171   8       1         17
#> 172   1       1         17
#> 173  10       1         17
#> 174   7       1         17
#> 175   8       1         17
#> 176   5       1         17
#> 177   7       1         17
#> 178   5       1         17
#> 179   6       1         17
#> 180   8       1         17
#> 181   6       1         18
#> 182  11       1         18
#> 183   9       1         18
#> 184   4       1         18
#> 185   6       1         18
#> 186   5       1         18
#> 187   3       1         18
#> 188   4       1         18
#> 189   7       1         18
#> 190   7       1         18
#> 191   2       1         19
#> 192   6       1         19
#> 193  10       1         19
#> 194  10       1         19
#> 195  12       1         19
#> 196   6       1         19
#> 197   3       1         19
#> 198  12       1         19
#> 199   2       1         19
#> 200   8       1         19
#> 201  10       1         20
#> 202  11       1         20
#> 203   9       1         20
#> 204   7       1         20
#> 205   9       1         20
#> 206   8       1         20
#> 207   7       1         20
#> 208   8       1         20
#> 209   8       1         20
#> 210  11       1         20
#> 211   4       1         21
#> 212   9       1         21
#> 213   7       1         21
#> 214   9       1         21
#> 215   7       1         21
#> 216   7       1         21
#> 217  11       1         21
#> 218   7       1         21
#> 219   7       1         21
#> 220   8       1         21
#> 221   7       1         22
#> 222   9       1         22
#> 223   6       1         22
#> 224  10       1         22
#> 225   7       1         22
#> 226  10       1         22
#> 227   9       1         22
#> 228   6       1         22
#> 229  10       1         22
#> 230   4       1         22
#> 231  12       1         23
#> 232   7       1         23
#> 233   8       1         23
#> 234  13       1         23
#> 235  10       1         23
#> 236  14       1         23
#> 237   5       1         23
#> 238   3       1         23
#> 239   9       1         23
#> 240  12       1         23
#> 241   7       1         24
#> 242  12       1         24
#> 243   8       1         24
#> 244  12       1         24
#> 245  10       1         24
#> 246   7       1         24
#> 247  11       1         24
#> 248   9       1         24
#> 249  11       1         24
#> 250  13       1         24
#> 251  12       1         25
#> 252   8       1         25
#> 253   6       1         25
#> 254  13       1         25
#> 255   9       1         25
#> 256  15       1         25
#> 257  13       1         25
#> 258   8       1         25
#> 259   8       1         25
#> 260  10       1         25
#> 261  14       1         26
#> 262  12       1         26
#> 263  14       1         26
#> 264   9       1         26
#> 265  10       1         26
#> 266  13       1         26
#> 267  11       1         26
#> 268  11       1         26
#> 269   8       1         26
#> 270   9       1         26
#> 271  12       1         27
#> 272  13       1         27
#> 273   9       1         27
#> 274   9       1         27
#> 275   8       1         27
#> 276  11       1         27
#> 277   8       1         27
#> 278  12       1         27
#> 279   7       1         27
#> 280  10       1         27
#> 281  12       1         28
#> 282  12       1         28
#> 283  11       1         28
#> 284  14       1         28
#> 285  13       1         28
#> 286  15       1         28
#> 287  14       1         28
#> 288  11       1         28
#> 289  12       1         28
#> 290  11       1         28
#> 291  15       1         29
#> 292  12       1         29
#> 293   9       1         29
#> 294   7       1         29
#> 295  12       1         29
#> 296  11       1         29
#> 297   9       1         29
#> 298  11       1         29
#> 299   9       1         29
#> 300  10       1         29
#> 301   5       1         30
#> 302  13       1         30
#> 303  11       1         30
#> 304  13       1         30
#> 305  14       1         30
#> 306  10       1         30
#> 307  12       1         30
#> 308  11       1         30
#> 309  11       1         30
#> 310  12       1         30
#> 311   8       1         31
#> 312  10       1         31
#> 313  16       1         31
#> 314   9       1         31
#> 315  14       1         31
#> 316  13       1         31
#> 317   8       1         31
#> 318  19       1         31
#> 319   9       1         31
#> 320  11       1         31
#> 321  17       1         32
#> 322  10       1         32
#> 323  11       1         32
#> 324  10       1         32
#> 325  15       1         32
#> 326  13       1         32
#> 327   9       1         32
#> 328  11       1         32
#> 329  13       1         32
#> 330   9       1         32
#> 331  12       1         33
#> 332  11       1         33
#> 333  14       1         33
#> 334  19       1         33
#> 335   9       1         33
#> 336  12       1         33
#> 337  13       1         33
#> 338  12       1         33
#> 339  14       1         33
#> 340  13       1         33
#> 341  15       1         34
#> 342   8       1         34
#> 343  13       1         34
#> 344  10       1         34
#> 345  13       1         34
#> 346  16       1         34
#> 347  17       1         34
#> 348  13       1         34
#> 349  17       1         34
#> 350  14       1         34
#> 351  17       1         35
#> 352  14       1         35
#> 353  11       1         35
#> 354  20       1         35
#> 355  16       1         35
#> 356  15       1         35
#> 357  14       1         35
#> 358  14       1         35
#> 359  14       1         35
#> 360  16       1         35
#> 361  18       1         36
#> 362  15       1         36
#> 363  13       1         36
#> 364  15       1         36
#> 365  11       1         36
#> 366  17       1         36
#> 367  13       1         36
#> 368  11       1         36
#> 369  14       1         36
#> 370  18       1         36
#> 371  17       1         37
#> 372  14       1         37
#> 373  18       1         37
#> 374  16       1         37
#> 375  14       1         37
#> 376  11       1         37
#> 377  18       1         37
#> 378  14       1         37
#> 379   8       1         37
#> 380  18       1         37
#> 381  15       1         38
#> 382  18       1         38
#> 383  16       1         38
#> 384  14       1         38
#> 385  15       1         38
#> 386  20       1         38
#> 387  19       1         38
#> 388  14       1         38
#> 389  18       1         38
#> 390  19       1         38
#> 391  17       1         39
#> 392  16       1         39
#> 393  21       1         39
#> 394  17       1         39
#> 395   8       1         39
#> 396  15       1         39
#> 397  14       1         39
#> 398  17       1         39
#> 399  15       1         39
#> 400  19       1         39
#> 401  15       1         40
#> 402  11       1         40
#> 403  18       1         40
#> 404  15       1         40
#> 405  16       1         40
#> 406  21       1         40
#> 407  11       1         40
#> 408  14       1         40
#> 409  12       1         40
#> 410  18       1         40
#> 411  20       1         41
#> 412  21       1         41
#> 413  14       1         41
#> 414  20       1         41
#> 415  15       1         41
#> 416  15       1         41
#> 417  17       1         41
#> 418  22       1         41
#> 419  15       1         41
#> 420  18       1         41
#> 421  19       1         42
#> 422  10       1         42
#> 423  22       1         42
#> 424  17       1         42
#> 425  14       1         42
#> 426  18       1         42
#> 427  17       1         42
#> 428  19       1         42
#> 429  16       1         42
#> 430  13       1         42
#> 431  15       1         43
#> 432  21       1         43
#> 433  17       1         43
#> 434  12       1         43
#> 435  16       1         43
#> 436  14       1         43
#> 437  22       1         43
#> 438  22       1         43
#> 439  14       1         43
#> 440   9       1         43
#> 441  20       1         44
#> 442  26       2         44
#> 443  23       1         44
#> 444  22       1         44
#> 445  19       1         44
#> 446  22       1         44
#> 447  11       1         44
#> 448  17       1         44
#> 449  15       1         44
#> 450  16       1         44
#> 451  15       1         45
#> 452  20       1         45
#> 453  14       1         45
#> 454  19       1         45
#> 455  15       1         45
#> 456  11       1         45
#> 457  14       1         45
#> 458  15       1         45
#> 459  23       1         45
#> 460  21       1         45
#> 461  18       1         46
#> 462  17       1         46
#> 463  19       1         46
#> 464  20       1         46
#> 465  18       1         46
#> 466  17       1         46
#> 467  16       1         46
#> 468  20       1         46
#> 469  17       1         46
#> 470  19       1         46
#> 471  19       1         47
#> 472  19       1         47
#> 473  14       1         47
#> 474  21       1         47
#> 475  13       1         47
#> 476  22       1         47
#> 477  24       1         47
#> 478  20       1         47
#> 479  19       1         47
#> 480  17       1         47
#> 481  20       1         48
#> 482  20       1         48
#> 483  19       1         48
#> 484  20       1         48
#> 485  22       1         48
#> 486  19       1         48
#> 487  16       1         48
#> 488  18       1         48
#> 489  18       1         48
#> 490  18       1         48
#> 491  23       1         49
#> 492  18       1         49
#> 493  21       1         49
#> 494  15       1         49
#> 495  20       1         49
#> 496  16       1         49
#> 497  22       1         49
#> 498  20       1         49
#> 499  16       1         49
#> 500  25       2         49
#> 501  15       1         50
#> 502  19       1         50
#> 503  21       1         50
#> 504  20       1         50
#> 505  26       2         50
#> 506  21       1         50
#> 507  22       1         50
#> 508  18       1         50
#> 509  17       1         50
#> 510  19       1         50
#> 511  20       1         51
#> 512  20       1         51
#> 513  19       1         51
#> 514  22       1         51
#> 515  22       1         51
#> 516  24       1         51
#> 517  22       1         51
#> 518  17       1         51
#> 519  18       1         51
#> 520  20       1         51
#> 521  18       1         52
#> 522  24       1         52
#> 523  20       1         52
#> 524  22       1         52
#> 525  24       1         52
#> 526  18       1         52
#> 527  19       1         52
#> 528  22       1         52
#> 529  21       1         52
#> 530  19       1         52
#> 531  25       2         53
#> 532  20       1         53
#> 533  19       1         53
#> 534  18       1         53
#> 535  20       1         53
#> 536  21       1         53
#> 537  18       1         53
#> 538  19       1         53
#> 539  22       1         53
#> 540  25       2         53
#> 541  23       1         54
#> 542  16       1         54
#> 543  23       1         54
#> 544  19       1         54
#> 545  20       1         54
#> 546  22       1         54
#> 547  24       1         54
#> 548  20       1         54
#> 549  22       1         54
#> 550  22       1         54
#> 551  20       1         55
#> 552  21       1         55
#> 553  20       1         55
#> 554  21       1         55
#> 555  27       2         55
#> 556  20       1         55
#> 557  26       2         55
#> 558  20       1         55
#> 559  19       1         55
#> 560  23       1         55
#> 561  22       1         56
#> 562  21       1         56
#> 563  22       1         56
#> 564  21       1         56
#> 565  21       1         56
#> 566  28       2         56
#> 567  24       1         56
#> 568  20       1         56
#> 569  22       1         56
#> 570  24       1         56
#> 571  21       1         57
#> 572  16       1         57
#> 573  27       2         57
#> 574  22       1         57
#> 575  21       1         57
#> 576  22       1         57
#> 577  26       2         57
#> 578  22       1         57
#> 579  20       1         57
#> 580  20       1         57
#> 581  23       1         58
#> 582  20       1         58
#> 583  23       1         58
#> 584  14       1         58
#> 585  22       1         58
#> 586  21       1         58
#> 587  22       1         58
#> 588  21       1         58
#> 589  28       2         58
#> 590  23       1         58
#> 591  28       2         59
#> 592  27       2         59
#> 593  25       2         59
#> 594  26       2         59
#> 595  25       2         59
#> 596  22       1         59
#> 597  25       2         59
#> 598  19       1         59
#> 599  20       1         59
#> 600  23       1         59
#> 601  27       2         60
#> 602  27       2         60
#> 603  22       1         60
#> 604  25       2         60
#> 605  20       1         60
#> 606  25       2         60
#> 607  23       1         60
#> 608  21       1         60
#> 609  20       1         60
#> 610  19       1         60
#> 611  25       2         61
#> 612  22       1         61
#> 613  26       2         61
#> 614  28       2         61
#> 615  25       2         61
#> 616  22       1         61
#> 617  29       2         61
#> 618  24       1         61
#> 619  24       1         61
#> 620  22       1         61
#> 621  27       2         62
#> 622  23       1         62
#> 623  25       2         62
#> 624  22       1         62
#> 625  25       2         62
#> 626  32       2         62
#> 627  25       2         62
#> 628  21       1         62
#> 629  23       1         62
#> 630  21       1         62
#> 631  25       2         63
#> 632  24       1         63
#> 633  24       1         63
#> 634  29       2         63
#> 635  21       1         63
#> 636  21       1         63
#> 637  30       2         63
#> 638  29       2         63
#> 639  25       2         63
#> 640  29       2         63
#> 641  30       2         64
#> 642  29       2         64
#> 643  29       2         64
#> 644  29       2         64
#> 645  20       1         64
#> 646  28       2         64
#> 647  25       2         64
#> 648  31       2         64
#> 649  28       2         64
#> 650  24       1         64
#> 651  24       1         65
#> 652  23       1         65
#> 653  30       2         65
#> 654  25       2         65
#> 655  31       2         65
#> 656  28       2         65
#> 657  23       1         65
#> 658  24       1         65
#> 659  29       2         65
#> 660  27       2         65
#> 661  27       2         66
#> 662  24       1         66
#> 663  25       2         66
#> 664  25       2         66
#> 665  29       2         66
#> 666  27       2         66
#> 667  25       2         66
#> 668  28       2         66
#> 669  26       2         66
#> 670  23       1         66
#> 671  29       2         67
#> 672  25       2         67
#> 673  27       2         67
#> 674  29       2         67
#> 675  27       2         67
#> 676  26       2         67
#> 677  26       2         67
#> 678  25       2         67
#> 679  29       2         67
#> 680  24       1         67
#> 681  27       2         68
#> 682  29       2         68
#> 683  27       2         68
#> 684  28       2         68
#> 685  29       2         68
#> 686  26       2         68
#> 687  31       2         68
#> 688  21       1         68
#> 689  25       2         68
#> 690  36       2         68
#> 691  20       1         69
#> 692  26       2         69
#> 693  27       2         69
#> 694  29       2         69
#> 695  26       2         69
#> 696  28       2         69
#> 697  24       1         69
#> 698  27       2         69
#> 699  29       2         69
#> 700  32       2         69
#> 701  29       2         70
#> 702  29       2         70
#> 703  29       2         70
#> 704  23       1         70
#> 705  23       1         70
#> 706  30       2         70
#> 707  27       2         70
#> 708  29       2         70
#> 709  30       2         70
#> 710  30       2         70
#> 711  26       2         71
#> 712  22       1         71
#> 713  31       2         71
#> 714  26       2         71
#> 715  29       2         71
#> 716  30       2         71
#> 717  31       2         71
#> 718  26       2         71
#> 719  24       1         71
#> 720  27       2         71
#> 721  25       2         72
#> 722  27       2         72
#> 723  26       2         72
#> 724  25       2         72
#> 725  29       2         72
#> 726  31       2         72
#> 727  32       2         72
#> 728  30       2         72
#> 729  26       2         72
#> 730  32       2         72
#> 731  28       2         73
#> 732  23       1         73
#> 733  20       1         73
#> 734  24       1         73
#> 735  30       2         73
#> 736  24       1         73
#> 737  30       2         73
#> 738  26       2         73
#> 739  35       2         73
#> 740  26       2         73
#> 741  29       2         74
#> 742  31       2         74
#> 743  33       2         74
#> 744  28       2         74
#> 745  25       2         74
#> 746  31       2         74
#> 747  30       2         74
#> 748  29       2         74
#> 749  27       2         74
#> 750  31       2         74
#> 751  33       2         75
#> 752  30       2         75
#> 753  24       1         75
#> 754  33       2         75
#> 755  29       2         75
#> 756  27       2         75
#> 757  33       2         75
#> 758  28       2         75
#> 759  31       2         75
#> 760  29       2         75
#> 761  28       2         76
#> 762  28       2         76
#> 763  25       2         76
#> 764  31       2         76
#> 765  34       2         76
#> 766  32       2         76
#> 767  28       2         76
#> 768  28       2         76
#> 769  34       2         76
#> 770  29       2         76
#> 771  34       2         77
#> 772  30       2         77
#> 773  32       2         77
#> 774  35       2         77
#> 775  28       2         77
#> 776  31       2         77
#> 777  26       2         77
#> 778  29       2         77
#> 779  31       2         77
#> 780  27       2         77
#> 781  33       2         78
#> 782  32       2         78
#> 783  32       2         78
#> 784  28       2         78
#> 785  33       2         78
#> 786  31       2         78
#> 787  33       2         78
#> 788  33       2         78
#> 789  33       2         78
#> 790  33       2         78
#> 791  35       2         79
#> 792  33       2         79
#> 793  33       2         79
#> 794  34       2         79
#> 795  34       2         79
#> 796  29       2         79
#> 797  32       2         79
#> 798  29       2         79
#> 799  31       2         79
#> 800  33       2         79
#> 801  26       2         80
#> 802  33       2         80
#> 803  31       2         80
#> 804  32       2         80
#> 805  30       2         80
#> 806  31       2         80
#> 807  33       2         80
#> 808  33       2         80
#> 809  29       2         80
#> 810  33       2         80
#> 811  35       2         81
#> 812  31       2         81
#> 813  32       2         81
#> 814  29       2         81
#> 815  34       2         81
#> 816  31       2         81
#> 817  33       2         81
#> 818  28       2         81
#> 819  34       2         81
#> 820  29       2         81
#> 821  34       2         82
#> 822  36       2         82
#> 823  38       3         82
#> 824  32       2         82
#> 825  35       2         82
#> 826  34       2         82
#> 827  28       2         82
#> 828  34       2         82
#> 829  30       2         82
#> 830  32       2         82
#> 831  33       2         83
#> 832  31       2         83
#> 833  36       2         83
#> 834  37       3         83
#> 835  36       2         83
#> 836  28       2         83
#> 837  36       2         83
#> 838  31       2         83
#> 839  29       2         83
#> 840  30       2         83
#> 841  32       2         84
#> 842  32       2         84
#> 843  34       2         84
#> 844  33       2         84
#> 845  26       2         84
#> 846  34       2         84
#> 847  35       2         84
#> 848  33       2         84
#> 849  37       3         84
#> 850  32       2         84
#> 851  36       2         85
#> 852  38       3         85
#> 853  36       2         85
#> 854  36       2         85
#> 855  30       2         85
#> 856  36       2         85
#> 857  35       2         85
#> 858  30       2         85
#> 859  33       2         85
#> 860  33       2         85
#> 861  35       2         86
#> 862  30       2         86
#> 863  34       2         86
#> 864  34       2         86
#> 865  36       2         86
#> 866  36       2         86
#> 867  34       2         86
#> 868  37       3         86
#> 869  37       3         86
#> 870  36       2         86
#> 871  36       2         87
#> 872  32       2         87
#> 873  35       2         87
#> 874  34       2         87
#> 875  36       2         87
#> 876  34       2         87
#> 877  36       2         87
#> 878  37       3         87
#> 879  35       2         87
#> 880  30       2         87
#> 881  35       2         88
#> 882  35       2         88
#> 883  34       2         88
#> 884  36       2         88
#> 885  35       2         88
#> 886  35       2         88
#> 887  38       3         88
#> 888  35       2         88
#> 889  33       2         88
#> 890  37       3         88
#> 891  37       3         89
#> 892  36       2         89
#> 893  32       2         89
#> 894  39       3         89
#> 895  37       3         89
#> 896  36       2         89
#> 897  36       2         89
#> 898  37       3         89
#> 899  35       2         89
#> 900  33       2         89
#> 901  36       2         90
#> 902  37       3         90
#> 903  35       2         90
#> 904  36       2         90
#> 905  38       3         90
#> 906  39       3         90
#> 907  31       2         90
#> 908  32       2         90
#> 909  35       2         90
#> 910  35       2         90
#> 911  37       3         91
#> 912  40       3         91
#> 913  35       2         91
#> 914  35       2         91
#> 915  39       3         91
#> 916  37       3         91
#> 917  37       3         91
#> 918  37       3         91
#> 919  38       3         91
#> 920  38       3         91
#> 921  39       3         92
#> 922  37       3         92
#> 923  36       2         92
#> 924  40       3         92
#> 925  37       3         92
#> 926  39       3         92
#> 927  36       2         92
#> 928  38       3         92
#> 929  37       3         92
#> 930  36       2         92
#> 931  35       2         93
#> 932  39       3         93
#> 933  37       3         93
#> 934  32       2         93
#> 935  35       2         93
#> 936  38       3         93
#> 937  37       3         93
#> 938  36       2         93
#> 939  40       3         93
#> 940  34       2         93
#> 941  36       2         94
#> 942  37       3         94
#> 943  38       3         94
#> 944  40       3         94
#> 945  36       2         94
#> 946  40       3         94
#> 947  39       3         94
#> 948  39       3         94
#> 949  39       3         94
#> 950  38       3         94
#> 951  39       3         95
#> 952  38       3         95
#> 953  39       3         95
#> 954  39       3         95
#> 955  40       3         95
#> 956  39       3         95
#> 957  37       3         95
#> 958  38       3         95
#> 959  35       2         95
#> 960  37       3         95
#> 961  39       3         96
#> 962  40       3         96
#> 963  40       3         96
#> 964  39       3         96
#> 965  38       3         96
#> 966  40       3         96
#> 967  36       2         96
#> 968  36       2         96
#> 969  40       3         96
#> 970  38       3         96
#> 971  39       3         97
#> 972  40       3         97
#> 973  39       3         97
#> 974  39       3         97
#> 975  38       3         97
#> 976  40       3         97
#> 977  40       3         97
#> 978  40       3         97
#> 979  39       3         97
#> 980  39       3         97
#> 981  39       3         98
#> 982  39       3         98
#> 983  40       3         98
#> 984  39       3         98
#> 985  39       3         98
#> 986  39       3         98
#> 987  40       3         98
#> 988  40       3         98
#> 989  39       3         98
#> 990  39       3         98
#> 991  40       3         99
#> 992  40       3         99
#> 993  40       3         99
#> 994  40       3         99
#> 995  40       3         99
#> 996  39       3         99
#> 997  40       3         99
#> 998  40       3         99
#> 999  39       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100