Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    1       1          1
#> 12    1       1          1
#> 13    0       1          1
#> 14    0       1          1
#> 15    1       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    1       1          1
#> 19    0       1          1
#> 20    1       1          1
#> 21    0       1          2
#> 22    1       1          2
#> 23    0       1          2
#> 24    0       1          2
#> 25    1       1          2
#> 26    1       1          2
#> 27    3       1          2
#> 28    1       1          2
#> 29    1       1          2
#> 30    0       1          2
#> 31    2       1          3
#> 32    1       1          3
#> 33    1       1          3
#> 34    2       1          3
#> 35    3       1          3
#> 36    0       1          3
#> 37    1       1          3
#> 38    0       1          3
#> 39    1       1          3
#> 40    3       1          3
#> 41    1       1          4
#> 42    1       1          4
#> 43    3       1          4
#> 44    2       1          4
#> 45    2       1          4
#> 46    2       1          4
#> 47    1       1          4
#> 48    3       1          4
#> 49    0       1          4
#> 50    2       1          4
#> 51    1       1          5
#> 52    3       1          5
#> 53    3       1          5
#> 54    1       1          5
#> 55    1       1          5
#> 56    3       1          5
#> 57    2       1          5
#> 58    3       1          5
#> 59    3       1          5
#> 60    3       1          5
#> 61    2       1          6
#> 62    1       1          6
#> 63    2       1          6
#> 64    0       1          6
#> 65    2       1          6
#> 66    6       1          6
#> 67    2       1          6
#> 68    8       1          6
#> 69    4       1          6
#> 70    2       1          6
#> 71    3       1          7
#> 72    2       1          7
#> 73    3       1          7
#> 74    2       1          7
#> 75    3       1          7
#> 76    1       1          7
#> 77    5       1          7
#> 78    4       1          7
#> 79    1       1          7
#> 80    1       1          7
#> 81    6       1          8
#> 82    5       1          8
#> 83    3       1          8
#> 84    3       1          8
#> 85    4       1          8
#> 86    2       1          8
#> 87    3       1          8
#> 88    2       1          8
#> 89    1       1          8
#> 90    5       1          8
#> 91    1       1          9
#> 92    1       1          9
#> 93    4       1          9
#> 94    4       1          9
#> 95    1       1          9
#> 96    0       1          9
#> 97    3       1          9
#> 98    3       1          9
#> 99    2       1          9
#> 100   1       1          9
#> 101   5       1         10
#> 102   5       1         10
#> 103   6       1         10
#> 104   3       1         10
#> 105   4       1         10
#> 106   5       1         10
#> 107   3       1         10
#> 108   3       1         10
#> 109   3       1         10
#> 110   5       1         10
#> 111   6       1         11
#> 112   0       1         11
#> 113   3       1         11
#> 114   4       1         11
#> 115   1       1         11
#> 116   3       1         11
#> 117   7       1         11
#> 118   4       1         11
#> 119   6       1         11
#> 120   2       1         11
#> 121   5       1         12
#> 122   5       1         12
#> 123   2       1         12
#> 124   3       1         12
#> 125   2       1         12
#> 126   5       1         12
#> 127   3       1         12
#> 128   5       1         12
#> 129   2       1         12
#> 130   6       1         12
#> 131   1       1         13
#> 132   6       1         13
#> 133   5       1         13
#> 134   3       1         13
#> 135   7       1         13
#> 136   8       1         13
#> 137   4       1         13
#> 138   6       1         13
#> 139  13       1         13
#> 140   4       1         13
#> 141   4       1         14
#> 142   6       1         14
#> 143   5       1         14
#> 144   5       1         14
#> 145   5       1         14
#> 146   4       1         14
#> 147   5       1         14
#> 148   7       1         14
#> 149   7       1         14
#> 150   7       1         14
#> 151   8       1         15
#> 152   8       1         15
#> 153   8       1         15
#> 154   8       1         15
#> 155   7       1         15
#> 156   8       1         15
#> 157   7       1         15
#> 158   7       1         15
#> 159   5       1         15
#> 160   5       1         15
#> 161   7       1         16
#> 162   5       1         16
#> 163  10       1         16
#> 164   7       1         16
#> 165   5       1         16
#> 166   7       1         16
#> 167   3       1         16
#> 168   5       1         16
#> 169   9       1         16
#> 170   6       1         16
#> 171  10       1         17
#> 172   7       1         17
#> 173   6       1         17
#> 174   6       1         17
#> 175   7       1         17
#> 176   5       1         17
#> 177   7       1         17
#> 178   4       1         17
#> 179   5       1         17
#> 180   7       1         17
#> 181   7       1         18
#> 182   7       1         18
#> 183   7       1         18
#> 184   6       1         18
#> 185   6       1         18
#> 186   5       1         18
#> 187  11       1         18
#> 188   7       1         18
#> 189  11       1         18
#> 190   8       1         18
#> 191   5       1         19
#> 192  10       1         19
#> 193  10       1         19
#> 194  10       1         19
#> 195   7       1         19
#> 196   7       1         19
#> 197   8       1         19
#> 198   1       1         19
#> 199   3       1         19
#> 200   5       1         19
#> 201   9       1         20
#> 202  10       1         20
#> 203   9       1         20
#> 204   6       1         20
#> 205   4       1         20
#> 206   7       1         20
#> 207   5       1         20
#> 208   7       1         20
#> 209  13       1         20
#> 210  15       1         20
#> 211  11       1         21
#> 212   9       1         21
#> 213  11       1         21
#> 214   8       1         21
#> 215  14       1         21
#> 216   6       1         21
#> 217   7       1         21
#> 218  10       1         21
#> 219   7       1         21
#> 220  11       1         21
#> 221   7       1         22
#> 222   8       1         22
#> 223   9       1         22
#> 224   8       1         22
#> 225  10       1         22
#> 226   9       1         22
#> 227  10       1         22
#> 228   9       1         22
#> 229  14       1         22
#> 230  11       1         22
#> 231  13       1         23
#> 232  10       1         23
#> 233   9       1         23
#> 234  10       1         23
#> 235   7       1         23
#> 236   9       1         23
#> 237  13       1         23
#> 238   7       1         23
#> 239   7       1         23
#> 240  13       1         23
#> 241   9       1         24
#> 242   4       1         24
#> 243  10       1         24
#> 244   9       1         24
#> 245  12       1         24
#> 246   7       1         24
#> 247  14       1         24
#> 248   7       1         24
#> 249   7       1         24
#> 250  14       1         24
#> 251   5       1         25
#> 252   8       1         25
#> 253   7       1         25
#> 254  18       1         25
#> 255   7       1         25
#> 256  11       1         25
#> 257  10       1         25
#> 258   8       1         25
#> 259  10       1         25
#> 260  10       1         25
#> 261  16       1         26
#> 262  10       1         26
#> 263   9       1         26
#> 264  13       1         26
#> 265  11       1         26
#> 266  12       1         26
#> 267  14       1         26
#> 268   9       1         26
#> 269   8       1         26
#> 270   9       1         26
#> 271  14       1         27
#> 272  11       1         27
#> 273  15       1         27
#> 274  12       1         27
#> 275   8       1         27
#> 276  14       1         27
#> 277  12       1         27
#> 278  13       1         27
#> 279  12       1         27
#> 280  13       1         27
#> 281  10       1         28
#> 282  10       1         28
#> 283  14       1         28
#> 284  17       1         28
#> 285  13       1         28
#> 286  13       1         28
#> 287  16       1         28
#> 288  12       1         28
#> 289  12       1         28
#> 290  10       1         28
#> 291   6       1         29
#> 292  11       1         29
#> 293  10       1         29
#> 294  10       1         29
#> 295  14       1         29
#> 296  11       1         29
#> 297  14       1         29
#> 298  14       1         29
#> 299   8       1         29
#> 300  18       1         29
#> 301  10       1         30
#> 302  15       1         30
#> 303   9       1         30
#> 304  18       1         30
#> 305   9       1         30
#> 306  13       1         30
#> 307   9       1         30
#> 308   8       1         30
#> 309  12       1         30
#> 310   9       1         30
#> 311   9       1         31
#> 312  14       1         31
#> 313  16       1         31
#> 314  10       1         31
#> 315  11       1         31
#> 316  11       1         31
#> 317  13       1         31
#> 318  11       1         31
#> 319  11       1         31
#> 320  15       1         31
#> 321  16       1         32
#> 322  11       1         32
#> 323  10       1         32
#> 324   7       1         32
#> 325  14       1         32
#> 326  14       1         32
#> 327   7       1         32
#> 328  16       1         32
#> 329  16       1         32
#> 330  15       1         32
#> 331  13       1         33
#> 332  12       1         33
#> 333  17       1         33
#> 334  12       1         33
#> 335  16       1         33
#> 336   7       1         33
#> 337  11       1         33
#> 338  10       1         33
#> 339  11       1         33
#> 340  14       1         33
#> 341  17       1         34
#> 342  13       1         34
#> 343  10       1         34
#> 344  13       1         34
#> 345  14       1         34
#> 346  11       1         34
#> 347  13       1         34
#> 348  15       1         34
#> 349  10       1         34
#> 350  14       1         34
#> 351  14       1         35
#> 352  14       1         35
#> 353  15       1         35
#> 354  13       1         35
#> 355  19       1         35
#> 356  10       1         35
#> 357  12       1         35
#> 358  16       1         35
#> 359  16       1         35
#> 360  16       1         35
#> 361  17       1         36
#> 362  13       1         36
#> 363  12       1         36
#> 364  14       1         36
#> 365  12       1         36
#> 366  14       1         36
#> 367  10       1         36
#> 368  15       1         36
#> 369  14       1         36
#> 370  13       1         36
#> 371  16       1         37
#> 372  19       1         37
#> 373  18       1         37
#> 374  15       1         37
#> 375  12       1         37
#> 376  17       1         37
#> 377  18       1         37
#> 378  14       1         37
#> 379  15       1         37
#> 380  13       1         37
#> 381  18       1         38
#> 382  17       1         38
#> 383  13       1         38
#> 384  17       1         38
#> 385  16       1         38
#> 386  18       1         38
#> 387  17       1         38
#> 388  14       1         38
#> 389  15       1         38
#> 390  16       1         38
#> 391  15       1         39
#> 392  14       1         39
#> 393  17       1         39
#> 394  18       1         39
#> 395  19       1         39
#> 396  15       1         39
#> 397  14       1         39
#> 398  14       1         39
#> 399  15       1         39
#> 400  11       1         39
#> 401  17       1         40
#> 402  17       1         40
#> 403  11       1         40
#> 404  14       1         40
#> 405  16       1         40
#> 406  13       1         40
#> 407  19       1         40
#> 408  16       1         40
#> 409  10       1         40
#> 410  18       1         40
#> 411  19       1         41
#> 412  16       1         41
#> 413  13       1         41
#> 414  18       1         41
#> 415  17       1         41
#> 416  23       1         41
#> 417  14       1         41
#> 418  19       1         41
#> 419  15       1         41
#> 420  19       1         41
#> 421  13       1         42
#> 422  15       1         42
#> 423  14       1         42
#> 424  16       1         42
#> 425  15       1         42
#> 426  13       1         42
#> 427  14       1         42
#> 428  19       1         42
#> 429  19       1         42
#> 430  21       1         42
#> 431  19       1         43
#> 432  30       2         43
#> 433  20       1         43
#> 434  16       1         43
#> 435  25       2         43
#> 436  13       1         43
#> 437  16       1         43
#> 438  17       1         43
#> 439  14       1         43
#> 440  20       1         43
#> 441  17       1         44
#> 442  19       1         44
#> 443  12       1         44
#> 444  16       1         44
#> 445  11       1         44
#> 446  16       1         44
#> 447  17       1         44
#> 448  23       1         44
#> 449  16       1         44
#> 450  20       1         44
#> 451  20       1         45
#> 452  20       1         45
#> 453  17       1         45
#> 454  17       1         45
#> 455  14       1         45
#> 456  18       1         45
#> 457  21       1         45
#> 458  21       1         45
#> 459  17       1         45
#> 460  17       1         45
#> 461  16       1         46
#> 462  16       1         46
#> 463  18       1         46
#> 464  13       1         46
#> 465  15       1         46
#> 466  17       1         46
#> 467  22       1         46
#> 468  17       1         46
#> 469  18       1         46
#> 470  21       1         46
#> 471  11       1         47
#> 472  18       1         47
#> 473  18       1         47
#> 474  18       1         47
#> 475  21       1         47
#> 476  20       1         47
#> 477  19       1         47
#> 478  18       1         47
#> 479  22       1         47
#> 480  16       1         47
#> 481  15       1         48
#> 482  20       1         48
#> 483  22       1         48
#> 484  19       1         48
#> 485  22       1         48
#> 486  20       1         48
#> 487  22       1         48
#> 488  20       1         48
#> 489  23       1         48
#> 490  15       1         48
#> 491  15       1         49
#> 492  21       1         49
#> 493  21       1         49
#> 494  20       1         49
#> 495  22       1         49
#> 496  21       1         49
#> 497  16       1         49
#> 498  14       1         49
#> 499  18       1         49
#> 500  21       1         49
#> 501  24       1         50
#> 502  12       1         50
#> 503  20       1         50
#> 504  18       1         50
#> 505  24       1         50
#> 506  24       1         50
#> 507  21       1         50
#> 508  20       1         50
#> 509  22       1         50
#> 510  21       1         50
#> 511  23       1         51
#> 512  23       1         51
#> 513  21       1         51
#> 514  26       2         51
#> 515  27       2         51
#> 516  20       1         51
#> 517  22       1         51
#> 518  20       1         51
#> 519  18       1         51
#> 520  21       1         51
#> 521  23       1         52
#> 522  16       1         52
#> 523  26       2         52
#> 524  24       1         52
#> 525  14       1         52
#> 526  20       1         52
#> 527  17       1         52
#> 528  23       1         52
#> 529  20       1         52
#> 530  21       1         52
#> 531  28       2         53
#> 532  23       1         53
#> 533  18       1         53
#> 534  26       2         53
#> 535  18       1         53
#> 536  22       1         53
#> 537  23       1         53
#> 538  16       1         53
#> 539  19       1         53
#> 540  23       1         53
#> 541  22       1         54
#> 542  20       1         54
#> 543  21       1         54
#> 544  23       1         54
#> 545  28       2         54
#> 546  21       1         54
#> 547  22       1         54
#> 548  23       1         54
#> 549  19       1         54
#> 550  14       1         54
#> 551  26       2         55
#> 552  17       1         55
#> 553  19       1         55
#> 554  21       1         55
#> 555  21       1         55
#> 556  24       1         55
#> 557  23       1         55
#> 558  25       2         55
#> 559  20       1         55
#> 560  19       1         55
#> 561  26       2         56
#> 562  27       2         56
#> 563  20       1         56
#> 564  22       1         56
#> 565  26       2         56
#> 566  24       1         56
#> 567  26       2         56
#> 568  24       1         56
#> 569  19       1         56
#> 570  16       1         56
#> 571  25       2         57
#> 572  22       1         57
#> 573  25       2         57
#> 574  16       1         57
#> 575  24       1         57
#> 576  18       1         57
#> 577  26       2         57
#> 578  24       1         57
#> 579  27       2         57
#> 580  24       1         57
#> 581  19       1         58
#> 582  27       2         58
#> 583  24       1         58
#> 584  23       1         58
#> 585  27       2         58
#> 586  23       1         58
#> 587  27       2         58
#> 588  25       2         58
#> 589  21       1         58
#> 590  20       1         58
#> 591  28       2         59
#> 592  25       2         59
#> 593  24       1         59
#> 594  24       1         59
#> 595  25       2         59
#> 596  24       1         59
#> 597  18       1         59
#> 598  25       2         59
#> 599  27       2         59
#> 600  25       2         59
#> 601  22       1         60
#> 602  21       1         60
#> 603  26       2         60
#> 604  22       1         60
#> 605  26       2         60
#> 606  29       2         60
#> 607  19       1         60
#> 608  23       1         60
#> 609  26       2         60
#> 610  24       1         60
#> 611  27       2         61
#> 612  26       2         61
#> 613  21       1         61
#> 614  21       1         61
#> 615  25       2         61
#> 616  25       2         61
#> 617  24       1         61
#> 618  26       2         61
#> 619  22       1         61
#> 620  20       1         61
#> 621  29       2         62
#> 622  29       2         62
#> 623  24       1         62
#> 624  25       2         62
#> 625  23       1         62
#> 626  22       1         62
#> 627  26       2         62
#> 628  29       2         62
#> 629  24       1         62
#> 630  24       1         62
#> 631  25       2         63
#> 632  25       2         63
#> 633  20       1         63
#> 634  23       1         63
#> 635  24       1         63
#> 636  24       1         63
#> 637  23       1         63
#> 638  33       2         63
#> 639  24       1         63
#> 640  28       2         63
#> 641  31       2         64
#> 642  27       2         64
#> 643  23       1         64
#> 644  25       2         64
#> 645  30       2         64
#> 646  26       2         64
#> 647  29       2         64
#> 648  31       2         64
#> 649  26       2         64
#> 650  29       2         64
#> 651  26       2         65
#> 652  24       1         65
#> 653  29       2         65
#> 654  23       1         65
#> 655  28       2         65
#> 656  28       2         65
#> 657  25       2         65
#> 658  25       2         65
#> 659  26       2         65
#> 660  26       2         65
#> 661  26       2         66
#> 662  30       2         66
#> 663  28       2         66
#> 664  26       2         66
#> 665  28       2         66
#> 666  27       2         66
#> 667  26       2         66
#> 668  27       2         66
#> 669  30       2         66
#> 670  30       2         66
#> 671  28       2         67
#> 672  26       2         67
#> 673  32       2         67
#> 674  26       2         67
#> 675  27       2         67
#> 676  27       2         67
#> 677  31       2         67
#> 678  27       2         67
#> 679  27       2         67
#> 680  27       2         67
#> 681  27       2         68
#> 682  27       2         68
#> 683  23       1         68
#> 684  28       2         68
#> 685  28       2         68
#> 686  30       2         68
#> 687  21       1         68
#> 688  30       2         68
#> 689  28       2         68
#> 690  29       2         68
#> 691  29       2         69
#> 692  25       2         69
#> 693  27       2         69
#> 694  26       2         69
#> 695  23       1         69
#> 696  29       2         69
#> 697  29       2         69
#> 698  24       1         69
#> 699  28       2         69
#> 700  25       2         69
#> 701  29       2         70
#> 702  29       2         70
#> 703  26       2         70
#> 704  28       2         70
#> 705  26       2         70
#> 706  26       2         70
#> 707  26       2         70
#> 708  27       2         70
#> 709  24       1         70
#> 710  27       2         70
#> 711  26       2         71
#> 712  29       2         71
#> 713  28       2         71
#> 714  29       2         71
#> 715  28       2         71
#> 716  28       2         71
#> 717  25       2         71
#> 718  26       2         71
#> 719  24       1         71
#> 720  26       2         71
#> 721  35       2         72
#> 722  23       1         72
#> 723  32       2         72
#> 724  28       2         72
#> 725  32       2         72
#> 726  22       1         72
#> 727  33       2         72
#> 728  30       2         72
#> 729  31       2         72
#> 730  27       2         72
#> 731  34       2         73
#> 732  31       2         73
#> 733  28       2         73
#> 734  30       2         73
#> 735  27       2         73
#> 736  31       2         73
#> 737  29       2         73
#> 738  25       2         73
#> 739  30       2         73
#> 740  24       1         73
#> 741  29       2         74
#> 742  33       2         74
#> 743  33       2         74
#> 744  32       2         74
#> 745  32       2         74
#> 746  32       2         74
#> 747  27       2         74
#> 748  23       1         74
#> 749  33       2         74
#> 750  28       2         74
#> 751  28       2         75
#> 752  31       2         75
#> 753  33       2         75
#> 754  32       2         75
#> 755  30       2         75
#> 756  32       2         75
#> 757  31       2         75
#> 758  32       2         75
#> 759  32       2         75
#> 760  31       2         75
#> 761  31       2         76
#> 762  31       2         76
#> 763  29       2         76
#> 764  32       2         76
#> 765  29       2         76
#> 766  36       2         76
#> 767  32       2         76
#> 768  31       2         76
#> 769  29       2         76
#> 770  32       2         76
#> 771  35       2         77
#> 772  30       2         77
#> 773  29       2         77
#> 774  32       2         77
#> 775  31       2         77
#> 776  30       2         77
#> 777  34       2         77
#> 778  33       2         77
#> 779  27       2         77
#> 780  28       2         77
#> 781  27       2         78
#> 782  35       2         78
#> 783  30       2         78
#> 784  31       2         78
#> 785  33       2         78
#> 786  33       2         78
#> 787  24       1         78
#> 788  35       2         78
#> 789  31       2         78
#> 790  28       2         78
#> 791  31       2         79
#> 792  27       2         79
#> 793  30       2         79
#> 794  32       2         79
#> 795  28       2         79
#> 796  35       2         79
#> 797  35       2         79
#> 798  32       2         79
#> 799  27       2         79
#> 800  32       2         79
#> 801  30       2         80
#> 802  33       2         80
#> 803  32       2         80
#> 804  34       2         80
#> 805  28       2         80
#> 806  32       2         80
#> 807  32       2         80
#> 808  35       2         80
#> 809  32       2         80
#> 810  32       2         80
#> 811  31       2         81
#> 812  31       2         81
#> 813  33       2         81
#> 814  31       2         81
#> 815  31       2         81
#> 816  31       2         81
#> 817  32       2         81
#> 818  32       2         81
#> 819  32       2         81
#> 820  32       2         81
#> 821  29       2         82
#> 822  32       2         82
#> 823  35       2         82
#> 824  34       2         82
#> 825  33       2         82
#> 826  28       2         82
#> 827  30       2         82
#> 828  35       2         82
#> 829  34       2         82
#> 830  33       2         82
#> 831  34       2         83
#> 832  34       2         83
#> 833  33       2         83
#> 834  35       2         83
#> 835  34       2         83
#> 836  34       2         83
#> 837  33       2         83
#> 838  37       3         83
#> 839  32       2         83
#> 840  34       2         83
#> 841  35       2         84
#> 842  34       2         84
#> 843  34       2         84
#> 844  31       2         84
#> 845  36       2         84
#> 846  35       2         84
#> 847  30       2         84
#> 848  36       2         84
#> 849  32       2         84
#> 850  35       2         84
#> 851  32       2         85
#> 852  38       3         85
#> 853  32       2         85
#> 854  35       2         85
#> 855  33       2         85
#> 856  34       2         85
#> 857  39       3         85
#> 858  36       2         85
#> 859  33       2         85
#> 860  31       2         85
#> 861  32       2         86
#> 862  34       2         86
#> 863  35       2         86
#> 864  35       2         86
#> 865  36       2         86
#> 866  33       2         86
#> 867  34       2         86
#> 868  34       2         86
#> 869  34       2         86
#> 870  36       2         86
#> 871  35       2         87
#> 872  39       3         87
#> 873  37       3         87
#> 874  37       3         87
#> 875  34       2         87
#> 876  40       3         87
#> 877  34       2         87
#> 878  32       2         87
#> 879  37       3         87
#> 880  37       3         87
#> 881  35       2         88
#> 882  36       2         88
#> 883  36       2         88
#> 884  36       2         88
#> 885  37       3         88
#> 886  35       2         88
#> 887  34       2         88
#> 888  34       2         88
#> 889  35       2         88
#> 890  36       2         88
#> 891  37       3         89
#> 892  33       2         89
#> 893  33       2         89
#> 894  35       2         89
#> 895  38       3         89
#> 896  35       2         89
#> 897  34       2         89
#> 898  37       3         89
#> 899  36       2         89
#> 900  38       3         89
#> 901  35       2         90
#> 902  38       3         90
#> 903  36       2         90
#> 904  36       2         90
#> 905  36       2         90
#> 906  37       3         90
#> 907  35       2         90
#> 908  38       3         90
#> 909  34       2         90
#> 910  34       2         90
#> 911  34       2         91
#> 912  37       3         91
#> 913  38       3         91
#> 914  36       2         91
#> 915  37       3         91
#> 916  35       2         91
#> 917  35       2         91
#> 918  37       3         91
#> 919  39       3         91
#> 920  36       2         91
#> 921  38       3         92
#> 922  35       2         92
#> 923  36       2         92
#> 924  37       3         92
#> 925  38       3         92
#> 926  37       3         92
#> 927  38       3         92
#> 928  38       3         92
#> 929  39       3         92
#> 930  38       3         92
#> 931  38       3         93
#> 932  40       3         93
#> 933  36       2         93
#> 934  38       3         93
#> 935  40       3         93
#> 936  40       3         93
#> 937  40       3         93
#> 938  38       3         93
#> 939  38       3         93
#> 940  36       2         93
#> 941  36       2         94
#> 942  39       3         94
#> 943  39       3         94
#> 944  38       3         94
#> 945  38       3         94
#> 946  37       3         94
#> 947  37       3         94
#> 948  39       3         94
#> 949  35       2         94
#> 950  35       2         94
#> 951  40       3         95
#> 952  38       3         95
#> 953  38       3         95
#> 954  37       3         95
#> 955  39       3         95
#> 956  36       2         95
#> 957  38       3         95
#> 958  35       2         95
#> 959  40       3         95
#> 960  38       3         95
#> 961  40       3         96
#> 962  33       2         96
#> 963  40       3         96
#> 964  38       3         96
#> 965  40       3         96
#> 966  39       3         96
#> 967  40       3         96
#> 968  39       3         96
#> 969  40       3         96
#> 970  38       3         96
#> 971  39       3         97
#> 972  38       3         97
#> 973  40       3         97
#> 974  40       3         97
#> 975  37       3         97
#> 976  40       3         97
#> 977  38       3         97
#> 978  40       3         97
#> 979  39       3         97
#> 980  40       3         97
#> 981  37       3         98
#> 982  40       3         98
#> 983  40       3         98
#> 984  39       3         98
#> 985  40       3         98
#> 986  39       3         98
#> 987  39       3         98
#> 988  40       3         98
#> 989  40       3         98
#> 990  39       3         98
#> 991  40       3         99
#> 992  39       3         99
#> 993  39       3         99
#> 994  40       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  40       3         99
#> 998  39       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100