Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    1       1          1
#> 12    0       1          1
#> 13    0       1          1
#> 14    1       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    0       1          2
#> 23    1       1          2
#> 24    0       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    1       1          2
#> 28    1       1          2
#> 29    1       1          2
#> 30    1       1          2
#> 31    1       1          3
#> 32    0       1          3
#> 33    2       1          3
#> 34    1       1          3
#> 35    0       1          3
#> 36    0       1          3
#> 37    2       1          3
#> 38    2       1          3
#> 39    1       1          3
#> 40    3       1          3
#> 41    2       1          4
#> 42    1       1          4
#> 43    2       1          4
#> 44    1       1          4
#> 45    1       1          4
#> 46    1       1          4
#> 47    4       1          4
#> 48    1       1          4
#> 49    0       1          4
#> 50    1       1          4
#> 51    2       1          5
#> 52    1       1          5
#> 53    1       1          5
#> 54    1       1          5
#> 55    0       1          5
#> 56    2       1          5
#> 57    1       1          5
#> 58    1       1          5
#> 59    3       1          5
#> 60    5       1          5
#> 61    3       1          6
#> 62    2       1          6
#> 63    1       1          6
#> 64    1       1          6
#> 65    1       1          6
#> 66    7       1          6
#> 67    4       1          6
#> 68    0       1          6
#> 69    1       1          6
#> 70    1       1          6
#> 71    7       1          7
#> 72    3       1          7
#> 73    3       1          7
#> 74    1       1          7
#> 75    2       1          7
#> 76    4       1          7
#> 77    2       1          7
#> 78    1       1          7
#> 79    4       1          7
#> 80    2       1          7
#> 81    2       1          8
#> 82    0       1          8
#> 83    4       1          8
#> 84    2       1          8
#> 85    1       1          8
#> 86    2       1          8
#> 87    3       1          8
#> 88    2       1          8
#> 89    7       1          8
#> 90    5       1          8
#> 91    4       1          9
#> 92    3       1          9
#> 93    1       1          9
#> 94    6       1          9
#> 95    3       1          9
#> 96    4       1          9
#> 97    4       1          9
#> 98    5       1          9
#> 99    2       1          9
#> 100   5       1          9
#> 101   6       1         10
#> 102   5       1         10
#> 103   3       1         10
#> 104   3       1         10
#> 105   5       1         10
#> 106   4       1         10
#> 107   7       1         10
#> 108   8       1         10
#> 109   6       1         10
#> 110   6       1         10
#> 111   3       1         11
#> 112   2       1         11
#> 113   2       1         11
#> 114   3       1         11
#> 115   4       1         11
#> 116   5       1         11
#> 117   7       1         11
#> 118   8       1         11
#> 119   3       1         11
#> 120   6       1         11
#> 121   4       1         12
#> 122   5       1         12
#> 123   2       1         12
#> 124   4       1         12
#> 125   2       1         12
#> 126   6       1         12
#> 127   6       1         12
#> 128   5       1         12
#> 129   6       1         12
#> 130   5       1         12
#> 131   2       1         13
#> 132   8       1         13
#> 133   2       1         13
#> 134   2       1         13
#> 135   8       1         13
#> 136   5       1         13
#> 137   5       1         13
#> 138   3       1         13
#> 139   3       1         13
#> 140   5       1         13
#> 141   4       1         14
#> 142   4       1         14
#> 143   5       1         14
#> 144   4       1         14
#> 145   6       1         14
#> 146   4       1         14
#> 147  10       1         14
#> 148   2       1         14
#> 149   8       1         14
#> 150   8       1         14
#> 151   5       1         15
#> 152   4       1         15
#> 153   4       1         15
#> 154   6       1         15
#> 155   8       1         15
#> 156  12       1         15
#> 157   6       1         15
#> 158   8       1         15
#> 159   3       1         15
#> 160  11       1         15
#> 161   7       1         16
#> 162   6       1         16
#> 163   6       1         16
#> 164   3       1         16
#> 165   6       1         16
#> 166   3       1         16
#> 167   3       1         16
#> 168   5       1         16
#> 169   5       1         16
#> 170  11       1         16
#> 171   7       1         17
#> 172   3       1         17
#> 173   5       1         17
#> 174   5       1         17
#> 175   6       1         17
#> 176   7       1         17
#> 177   6       1         17
#> 178   4       1         17
#> 179  10       1         17
#> 180   6       1         17
#> 181   8       1         18
#> 182   7       1         18
#> 183   9       1         18
#> 184   6       1         18
#> 185   5       1         18
#> 186   4       1         18
#> 187   5       1         18
#> 188  12       1         18
#> 189   8       1         18
#> 190   5       1         18
#> 191   7       1         19
#> 192  15       1         19
#> 193   3       1         19
#> 194  12       1         19
#> 195   5       1         19
#> 196   6       1         19
#> 197   4       1         19
#> 198   9       1         19
#> 199   8       1         19
#> 200   6       1         19
#> 201   3       1         20
#> 202   6       1         20
#> 203   8       1         20
#> 204  11       1         20
#> 205  10       1         20
#> 206  11       1         20
#> 207   8       1         20
#> 208   5       1         20
#> 209  11       1         20
#> 210   7       1         20
#> 211  12       1         21
#> 212   7       1         21
#> 213  10       1         21
#> 214   8       1         21
#> 215   9       1         21
#> 216  13       1         21
#> 217   5       1         21
#> 218   6       1         21
#> 219   9       1         21
#> 220  11       1         21
#> 221   8       1         22
#> 222   8       1         22
#> 223  10       1         22
#> 224   7       1         22
#> 225  10       1         22
#> 226  11       1         22
#> 227   8       1         22
#> 228   7       1         22
#> 229   5       1         22
#> 230   9       1         22
#> 231   8       1         23
#> 232   8       1         23
#> 233  11       1         23
#> 234   5       1         23
#> 235   8       1         23
#> 236   9       1         23
#> 237  12       1         23
#> 238  10       1         23
#> 239  10       1         23
#> 240  11       1         23
#> 241   6       1         24
#> 242   6       1         24
#> 243   9       1         24
#> 244   9       1         24
#> 245   8       1         24
#> 246  14       1         24
#> 247  11       1         24
#> 248  10       1         24
#> 249  11       1         24
#> 250  12       1         24
#> 251  11       1         25
#> 252  12       1         25
#> 253  10       1         25
#> 254   6       1         25
#> 255  10       1         25
#> 256  11       1         25
#> 257   8       1         25
#> 258  15       1         25
#> 259   9       1         25
#> 260  10       1         25
#> 261   8       1         26
#> 262   6       1         26
#> 263   8       1         26
#> 264  14       1         26
#> 265   9       1         26
#> 266  10       1         26
#> 267  12       1         26
#> 268  12       1         26
#> 269   3       1         26
#> 270  11       1         26
#> 271   8       1         27
#> 272  10       1         27
#> 273   9       1         27
#> 274  11       1         27
#> 275   8       1         27
#> 276   9       1         27
#> 277  10       1         27
#> 278  13       1         27
#> 279   9       1         27
#> 280  13       1         27
#> 281  13       1         28
#> 282   9       1         28
#> 283  14       1         28
#> 284   7       1         28
#> 285  13       1         28
#> 286  15       1         28
#> 287  12       1         28
#> 288   9       1         28
#> 289  11       1         28
#> 290  11       1         28
#> 291   8       1         29
#> 292   9       1         29
#> 293  13       1         29
#> 294  10       1         29
#> 295   9       1         29
#> 296  10       1         29
#> 297  11       1         29
#> 298  10       1         29
#> 299  16       1         29
#> 300  13       1         29
#> 301  10       1         30
#> 302  12       1         30
#> 303  16       1         30
#> 304  12       1         30
#> 305  10       1         30
#> 306   9       1         30
#> 307   8       1         30
#> 308  16       1         30
#> 309   8       1         30
#> 310   8       1         30
#> 311  11       1         31
#> 312  14       1         31
#> 313  12       1         31
#> 314  18       1         31
#> 315  13       1         31
#> 316  14       1         31
#> 317  13       1         31
#> 318  17       1         31
#> 319  13       1         31
#> 320  11       1         31
#> 321  14       1         32
#> 322  13       1         32
#> 323  14       1         32
#> 324  13       1         32
#> 325  11       1         32
#> 326  10       1         32
#> 327  11       1         32
#> 328  17       1         32
#> 329  13       1         32
#> 330  16       1         32
#> 331  10       1         33
#> 332  10       1         33
#> 333  10       1         33
#> 334  14       1         33
#> 335  16       1         33
#> 336  16       1         33
#> 337  15       1         33
#> 338  10       1         33
#> 339  12       1         33
#> 340  12       1         33
#> 341  15       1         34
#> 342  13       1         34
#> 343  10       1         34
#> 344  11       1         34
#> 345  14       1         34
#> 346  16       1         34
#> 347  10       1         34
#> 348   9       1         34
#> 349  12       1         34
#> 350  14       1         34
#> 351  16       1         35
#> 352  17       1         35
#> 353  12       1         35
#> 354  15       1         35
#> 355  19       1         35
#> 356  10       1         35
#> 357  16       1         35
#> 358  19       1         35
#> 359  12       1         35
#> 360  18       1         35
#> 361  11       1         36
#> 362  19       1         36
#> 363  12       1         36
#> 364   9       1         36
#> 365  21       1         36
#> 366  15       1         36
#> 367  10       1         36
#> 368  15       1         36
#> 369  16       1         36
#> 370   8       1         36
#> 371  14       1         37
#> 372  15       1         37
#> 373  12       1         37
#> 374  15       1         37
#> 375  12       1         37
#> 376  13       1         37
#> 377  11       1         37
#> 378  13       1         37
#> 379   7       1         37
#> 380  13       1         37
#> 381  12       1         38
#> 382  18       1         38
#> 383  24       1         38
#> 384  13       1         38
#> 385  19       1         38
#> 386  20       1         38
#> 387   6       1         38
#> 388  15       1         38
#> 389  18       1         38
#> 390  15       1         38
#> 391  18       1         39
#> 392  19       1         39
#> 393  17       1         39
#> 394  11       1         39
#> 395  21       1         39
#> 396  13       1         39
#> 397  17       1         39
#> 398  18       1         39
#> 399  21       1         39
#> 400  18       1         39
#> 401  17       1         40
#> 402  16       1         40
#> 403  17       1         40
#> 404  11       1         40
#> 405  14       1         40
#> 406  11       1         40
#> 407  22       1         40
#> 408  18       1         40
#> 409  16       1         40
#> 410  20       1         40
#> 411  18       1         41
#> 412  20       1         41
#> 413  12       1         41
#> 414  18       1         41
#> 415  13       1         41
#> 416  15       1         41
#> 417  15       1         41
#> 418  20       1         41
#> 419  17       1         41
#> 420  18       1         41
#> 421  22       1         42
#> 422  16       1         42
#> 423   8       1         42
#> 424  16       1         42
#> 425  18       1         42
#> 426  18       1         42
#> 427  15       1         42
#> 428  13       1         42
#> 429  22       1         42
#> 430  16       1         42
#> 431  15       1         43
#> 432  20       1         43
#> 433  19       1         43
#> 434  19       1         43
#> 435  19       1         43
#> 436  14       1         43
#> 437  17       1         43
#> 438  16       1         43
#> 439  16       1         43
#> 440  17       1         43
#> 441  20       1         44
#> 442  17       1         44
#> 443  21       1         44
#> 444  19       1         44
#> 445  28       2         44
#> 446  19       1         44
#> 447  18       1         44
#> 448  21       1         44
#> 449  17       1         44
#> 450  23       1         44
#> 451  20       1         45
#> 452  17       1         45
#> 453  14       1         45
#> 454  15       1         45
#> 455  20       1         45
#> 456  20       1         45
#> 457  23       1         45
#> 458  23       1         45
#> 459  22       1         45
#> 460  24       1         45
#> 461  21       1         46
#> 462  18       1         46
#> 463  19       1         46
#> 464  17       1         46
#> 465  18       1         46
#> 466  22       1         46
#> 467  17       1         46
#> 468  15       1         46
#> 469  14       1         46
#> 470  17       1         46
#> 471  18       1         47
#> 472  17       1         47
#> 473  22       1         47
#> 474  18       1         47
#> 475  26       2         47
#> 476  16       1         47
#> 477  17       1         47
#> 478  22       1         47
#> 479  19       1         47
#> 480  17       1         47
#> 481  19       1         48
#> 482  15       1         48
#> 483  15       1         48
#> 484  20       1         48
#> 485  13       1         48
#> 486  15       1         48
#> 487  17       1         48
#> 488  15       1         48
#> 489  20       1         48
#> 490  21       1         48
#> 491  16       1         49
#> 492  16       1         49
#> 493  12       1         49
#> 494  19       1         49
#> 495  18       1         49
#> 496  16       1         49
#> 497  21       1         49
#> 498  18       1         49
#> 499  20       1         49
#> 500  18       1         49
#> 501  28       2         50
#> 502  19       1         50
#> 503  21       1         50
#> 504  21       1         50
#> 505  22       1         50
#> 506  26       2         50
#> 507  16       1         50
#> 508  22       1         50
#> 509  21       1         50
#> 510  21       1         50
#> 511  18       1         51
#> 512  19       1         51
#> 513  23       1         51
#> 514  21       1         51
#> 515  19       1         51
#> 516  16       1         51
#> 517  20       1         51
#> 518  15       1         51
#> 519  21       1         51
#> 520  24       1         51
#> 521  28       2         52
#> 522  20       1         52
#> 523  21       1         52
#> 524  11       1         52
#> 525  25       2         52
#> 526  22       1         52
#> 527  22       1         52
#> 528  20       1         52
#> 529  18       1         52
#> 530  20       1         52
#> 531  20       1         53
#> 532  21       1         53
#> 533  20       1         53
#> 534  19       1         53
#> 535  32       2         53
#> 536  23       1         53
#> 537  21       1         53
#> 538  16       1         53
#> 539  11       1         53
#> 540  22       1         53
#> 541  21       1         54
#> 542  18       1         54
#> 543  18       1         54
#> 544  22       1         54
#> 545  23       1         54
#> 546  24       1         54
#> 547  22       1         54
#> 548  20       1         54
#> 549  20       1         54
#> 550  23       1         54
#> 551  19       1         55
#> 552  22       1         55
#> 553  25       2         55
#> 554  23       1         55
#> 555  18       1         55
#> 556  24       1         55
#> 557  21       1         55
#> 558  23       1         55
#> 559  18       1         55
#> 560  20       1         55
#> 561  26       2         56
#> 562  23       1         56
#> 563  17       1         56
#> 564  23       1         56
#> 565  27       2         56
#> 566  23       1         56
#> 567  18       1         56
#> 568  30       2         56
#> 569  24       1         56
#> 570  24       1         56
#> 571  25       2         57
#> 572  28       2         57
#> 573  23       1         57
#> 574  24       1         57
#> 575  24       1         57
#> 576  25       2         57
#> 577  23       1         57
#> 578  28       2         57
#> 579  17       1         57
#> 580  21       1         57
#> 581  24       1         58
#> 582  18       1         58
#> 583  15       1         58
#> 584  23       1         58
#> 585  23       1         58
#> 586  29       2         58
#> 587  16       1         58
#> 588  22       1         58
#> 589  17       1         58
#> 590  24       1         58
#> 591  27       2         59
#> 592  31       2         59
#> 593  21       1         59
#> 594  22       1         59
#> 595  27       2         59
#> 596  26       2         59
#> 597  26       2         59
#> 598  19       1         59
#> 599  20       1         59
#> 600  22       1         59
#> 601  21       1         60
#> 602  25       2         60
#> 603  24       1         60
#> 604  27       2         60
#> 605  25       2         60
#> 606  21       1         60
#> 607  24       1         60
#> 608  29       2         60
#> 609  27       2         60
#> 610  27       2         60
#> 611  25       2         61
#> 612  25       2         61
#> 613  22       1         61
#> 614  20       1         61
#> 615  26       2         61
#> 616  24       1         61
#> 617  26       2         61
#> 618  23       1         61
#> 619  22       1         61
#> 620  28       2         61
#> 621  25       2         62
#> 622  23       1         62
#> 623  24       1         62
#> 624  23       1         62
#> 625  23       1         62
#> 626  28       2         62
#> 627  19       1         62
#> 628  30       2         62
#> 629  17       1         62
#> 630  23       1         62
#> 631  28       2         63
#> 632  24       1         63
#> 633  26       2         63
#> 634  23       1         63
#> 635  25       2         63
#> 636  29       2         63
#> 637  22       1         63
#> 638  29       2         63
#> 639  23       1         63
#> 640  30       2         63
#> 641  27       2         64
#> 642  27       2         64
#> 643  33       2         64
#> 644  20       1         64
#> 645  28       2         64
#> 646  30       2         64
#> 647  24       1         64
#> 648  25       2         64
#> 649  29       2         64
#> 650  24       1         64
#> 651  22       1         65
#> 652  29       2         65
#> 653  24       1         65
#> 654  26       2         65
#> 655  25       2         65
#> 656  27       2         65
#> 657  23       1         65
#> 658  22       1         65
#> 659  25       2         65
#> 660  27       2         65
#> 661  21       1         66
#> 662  27       2         66
#> 663  25       2         66
#> 664  28       2         66
#> 665  25       2         66
#> 666  27       2         66
#> 667  26       2         66
#> 668  28       2         66
#> 669  24       1         66
#> 670  30       2         66
#> 671  29       2         67
#> 672  29       2         67
#> 673  25       2         67
#> 674  27       2         67
#> 675  32       2         67
#> 676  32       2         67
#> 677  29       2         67
#> 678  26       2         67
#> 679  24       1         67
#> 680  28       2         67
#> 681  23       1         68
#> 682  25       2         68
#> 683  25       2         68
#> 684  25       2         68
#> 685  23       1         68
#> 686  29       2         68
#> 687  28       2         68
#> 688  29       2         68
#> 689  28       2         68
#> 690  30       2         68
#> 691  35       2         69
#> 692  28       2         69
#> 693  29       2         69
#> 694  30       2         69
#> 695  26       2         69
#> 696  24       1         69
#> 697  24       1         69
#> 698  26       2         69
#> 699  27       2         69
#> 700  30       2         69
#> 701  22       1         70
#> 702  26       2         70
#> 703  32       2         70
#> 704  31       2         70
#> 705  30       2         70
#> 706  30       2         70
#> 707  26       2         70
#> 708  29       2         70
#> 709  31       2         70
#> 710  28       2         70
#> 711  31       2         71
#> 712  26       2         71
#> 713  27       2         71
#> 714  30       2         71
#> 715  29       2         71
#> 716  27       2         71
#> 717  28       2         71
#> 718  30       2         71
#> 719  31       2         71
#> 720  28       2         71
#> 721  28       2         72
#> 722  28       2         72
#> 723  31       2         72
#> 724  26       2         72
#> 725  26       2         72
#> 726  28       2         72
#> 727  32       2         72
#> 728  33       2         72
#> 729  30       2         72
#> 730  29       2         72
#> 731  34       2         73
#> 732  32       2         73
#> 733  27       2         73
#> 734  28       2         73
#> 735  31       2         73
#> 736  32       2         73
#> 737  29       2         73
#> 738  29       2         73
#> 739  25       2         73
#> 740  32       2         73
#> 741  33       2         74
#> 742  30       2         74
#> 743  28       2         74
#> 744  30       2         74
#> 745  33       2         74
#> 746  34       2         74
#> 747  30       2         74
#> 748  32       2         74
#> 749  33       2         74
#> 750  31       2         74
#> 751  33       2         75
#> 752  27       2         75
#> 753  34       2         75
#> 754  28       2         75
#> 755  30       2         75
#> 756  33       2         75
#> 757  35       2         75
#> 758  30       2         75
#> 759  33       2         75
#> 760  27       2         75
#> 761  35       2         76
#> 762  30       2         76
#> 763  28       2         76
#> 764  29       2         76
#> 765  32       2         76
#> 766  28       2         76
#> 767  28       2         76
#> 768  30       2         76
#> 769  29       2         76
#> 770  32       2         76
#> 771  31       2         77
#> 772  34       2         77
#> 773  31       2         77
#> 774  28       2         77
#> 775  30       2         77
#> 776  29       2         77
#> 777  30       2         77
#> 778  27       2         77
#> 779  30       2         77
#> 780  29       2         77
#> 781  33       2         78
#> 782  34       2         78
#> 783  26       2         78
#> 784  34       2         78
#> 785  32       2         78
#> 786  36       2         78
#> 787  33       2         78
#> 788  28       2         78
#> 789  34       2         78
#> 790  31       2         78
#> 791  35       2         79
#> 792  30       2         79
#> 793  25       2         79
#> 794  35       2         79
#> 795  30       2         79
#> 796  30       2         79
#> 797  30       2         79
#> 798  35       2         79
#> 799  37       3         79
#> 800  32       2         79
#> 801  33       2         80
#> 802  32       2         80
#> 803  30       2         80
#> 804  27       2         80
#> 805  29       2         80
#> 806  31       2         80
#> 807  31       2         80
#> 808  32       2         80
#> 809  30       2         80
#> 810  28       2         80
#> 811  33       2         81
#> 812  34       2         81
#> 813  33       2         81
#> 814  34       2         81
#> 815  32       2         81
#> 816  31       2         81
#> 817  31       2         81
#> 818  28       2         81
#> 819  33       2         81
#> 820  31       2         81
#> 821  39       3         82
#> 822  33       2         82
#> 823  33       2         82
#> 824  34       2         82
#> 825  31       2         82
#> 826  34       2         82
#> 827  26       2         82
#> 828  37       3         82
#> 829  29       2         82
#> 830  35       2         82
#> 831  34       2         83
#> 832  28       2         83
#> 833  34       2         83
#> 834  37       3         83
#> 835  33       2         83
#> 836  32       2         83
#> 837  34       2         83
#> 838  36       2         83
#> 839  33       2         83
#> 840  32       2         83
#> 841  36       2         84
#> 842  37       3         84
#> 843  36       2         84
#> 844  37       3         84
#> 845  35       2         84
#> 846  35       2         84
#> 847  32       2         84
#> 848  33       2         84
#> 849  31       2         84
#> 850  25       2         84
#> 851  36       2         85
#> 852  31       2         85
#> 853  37       3         85
#> 854  29       2         85
#> 855  34       2         85
#> 856  33       2         85
#> 857  35       2         85
#> 858  34       2         85
#> 859  35       2         85
#> 860  35       2         85
#> 861  33       2         86
#> 862  38       3         86
#> 863  33       2         86
#> 864  35       2         86
#> 865  36       2         86
#> 866  31       2         86
#> 867  34       2         86
#> 868  35       2         86
#> 869  35       2         86
#> 870  34       2         86
#> 871  36       2         87
#> 872  30       2         87
#> 873  36       2         87
#> 874  36       2         87
#> 875  39       3         87
#> 876  33       2         87
#> 877  33       2         87
#> 878  30       2         87
#> 879  29       2         87
#> 880  39       3         87
#> 881  30       2         88
#> 882  35       2         88
#> 883  37       3         88
#> 884  37       3         88
#> 885  35       2         88
#> 886  37       3         88
#> 887  37       3         88
#> 888  37       3         88
#> 889  34       2         88
#> 890  34       2         88
#> 891  36       2         89
#> 892  36       2         89
#> 893  35       2         89
#> 894  39       3         89
#> 895  35       2         89
#> 896  35       2         89
#> 897  35       2         89
#> 898  33       2         89
#> 899  36       2         89
#> 900  33       2         89
#> 901  39       3         90
#> 902  35       2         90
#> 903  36       2         90
#> 904  37       3         90
#> 905  38       3         90
#> 906  35       2         90
#> 907  34       2         90
#> 908  37       3         90
#> 909  37       3         90
#> 910  34       2         90
#> 911  39       3         91
#> 912  37       3         91
#> 913  38       3         91
#> 914  37       3         91
#> 915  33       2         91
#> 916  38       3         91
#> 917  38       3         91
#> 918  36       2         91
#> 919  38       3         91
#> 920  37       3         91
#> 921  35       2         92
#> 922  38       3         92
#> 923  35       2         92
#> 924  36       2         92
#> 925  35       2         92
#> 926  34       2         92
#> 927  36       2         92
#> 928  37       3         92
#> 929  38       3         92
#> 930  38       3         92
#> 931  36       2         93
#> 932  39       3         93
#> 933  35       2         93
#> 934  36       2         93
#> 935  38       3         93
#> 936  40       3         93
#> 937  38       3         93
#> 938  36       2         93
#> 939  36       2         93
#> 940  38       3         93
#> 941  35       2         94
#> 942  38       3         94
#> 943  38       3         94
#> 944  38       3         94
#> 945  40       3         94
#> 946  31       2         94
#> 947  37       3         94
#> 948  38       3         94
#> 949  37       3         94
#> 950  38       3         94
#> 951  39       3         95
#> 952  39       3         95
#> 953  38       3         95
#> 954  37       3         95
#> 955  39       3         95
#> 956  37       3         95
#> 957  40       3         95
#> 958  38       3         95
#> 959  36       2         95
#> 960  38       3         95
#> 961  40       3         96
#> 962  39       3         96
#> 963  38       3         96
#> 964  40       3         96
#> 965  39       3         96
#> 966  40       3         96
#> 967  39       3         96
#> 968  37       3         96
#> 969  39       3         96
#> 970  39       3         96
#> 971  40       3         97
#> 972  40       3         97
#> 973  40       3         97
#> 974  40       3         97
#> 975  39       3         97
#> 976  38       3         97
#> 977  39       3         97
#> 978  40       3         97
#> 979  39       3         97
#> 980  40       3         97
#> 981  40       3         98
#> 982  40       3         98
#> 983  40       3         98
#> 984  39       3         98
#> 985  40       3         98
#> 986  38       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  37       3         98
#> 990  39       3         98
#> 991  38       3         99
#> 992  40       3         99
#> 993  38       3         99
#> 994  39       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  39       3         99
#> 998  38       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100