Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    1       1          1
#> 12    1       1          1
#> 13    1       1          1
#> 14    0       1          1
#> 15    0       1          1
#> 16    1       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    1       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    1       1          2
#> 23    1       1          2
#> 24    1       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    1       1          2
#> 28    0       1          2
#> 29    1       1          2
#> 30    1       1          2
#> 31    1       1          3
#> 32    2       1          3
#> 33    2       1          3
#> 34    2       1          3
#> 35    1       1          3
#> 36    1       1          3
#> 37    2       1          3
#> 38    0       1          3
#> 39    0       1          3
#> 40    1       1          3
#> 41    1       1          4
#> 42    1       1          4
#> 43    2       1          4
#> 44    2       1          4
#> 45    2       1          4
#> 46    2       1          4
#> 47    1       1          4
#> 48    1       1          4
#> 49    1       1          4
#> 50    3       1          4
#> 51    3       1          5
#> 52    1       1          5
#> 53    1       1          5
#> 54    0       1          5
#> 55    1       1          5
#> 56    1       1          5
#> 57    1       1          5
#> 58    2       1          5
#> 59    0       1          5
#> 60    0       1          5
#> 61    2       1          6
#> 62    2       1          6
#> 63    0       1          6
#> 64    3       1          6
#> 65    3       1          6
#> 66    3       1          6
#> 67    3       1          6
#> 68    2       1          6
#> 69    0       1          6
#> 70    0       1          6
#> 71    4       1          7
#> 72    0       1          7
#> 73    2       1          7
#> 74    1       1          7
#> 75    1       1          7
#> 76    3       1          7
#> 77    5       1          7
#> 78    3       1          7
#> 79    3       1          7
#> 80    2       1          7
#> 81    2       1          8
#> 82    2       1          8
#> 83    1       1          8
#> 84    2       1          8
#> 85    3       1          8
#> 86    4       1          8
#> 87    2       1          8
#> 88    6       1          8
#> 89    1       1          8
#> 90    2       1          8
#> 91    5       1          9
#> 92    5       1          9
#> 93    1       1          9
#> 94    1       1          9
#> 95    1       1          9
#> 96    5       1          9
#> 97    4       1          9
#> 98    4       1          9
#> 99    1       1          9
#> 100   2       1          9
#> 101   5       1         10
#> 102   3       1         10
#> 103   1       1         10
#> 104   3       1         10
#> 105   9       1         10
#> 106   5       1         10
#> 107   2       1         10
#> 108   2       1         10
#> 109   5       1         10
#> 110   4       1         10
#> 111   8       1         11
#> 112   4       1         11
#> 113   2       1         11
#> 114   3       1         11
#> 115   7       1         11
#> 116   3       1         11
#> 117   7       1         11
#> 118   9       1         11
#> 119   5       1         11
#> 120   5       1         11
#> 121   6       1         12
#> 122   8       1         12
#> 123   6       1         12
#> 124   4       1         12
#> 125   6       1         12
#> 126   6       1         12
#> 127   6       1         12
#> 128   4       1         12
#> 129   6       1         12
#> 130   6       1         12
#> 131   8       1         13
#> 132   6       1         13
#> 133   5       1         13
#> 134   5       1         13
#> 135   7       1         13
#> 136   3       1         13
#> 137   2       1         13
#> 138   2       1         13
#> 139   6       1         13
#> 140   6       1         13
#> 141   4       1         14
#> 142   2       1         14
#> 143   7       1         14
#> 144   5       1         14
#> 145  10       1         14
#> 146   2       1         14
#> 147   5       1         14
#> 148   4       1         14
#> 149   7       1         14
#> 150   6       1         14
#> 151   5       1         15
#> 152   9       1         15
#> 153   4       1         15
#> 154   4       1         15
#> 155   6       1         15
#> 156   5       1         15
#> 157   7       1         15
#> 158   7       1         15
#> 159  13       1         15
#> 160  12       1         15
#> 161   3       1         16
#> 162   3       1         16
#> 163   9       1         16
#> 164   8       1         16
#> 165   8       1         16
#> 166   1       1         16
#> 167   6       1         16
#> 168   5       1         16
#> 169   9       1         16
#> 170   7       1         16
#> 171   5       1         17
#> 172   7       1         17
#> 173   8       1         17
#> 174   7       1         17
#> 175   5       1         17
#> 176   5       1         17
#> 177   5       1         17
#> 178   6       1         17
#> 179   6       1         17
#> 180   9       1         17
#> 181   9       1         18
#> 182   6       1         18
#> 183  10       1         18
#> 184   6       1         18
#> 185   9       1         18
#> 186   8       1         18
#> 187   6       1         18
#> 188   8       1         18
#> 189   8       1         18
#> 190   6       1         18
#> 191   9       1         19
#> 192   4       1         19
#> 193   9       1         19
#> 194  16       1         19
#> 195  10       1         19
#> 196  11       1         19
#> 197  11       1         19
#> 198   6       1         19
#> 199  12       1         19
#> 200   3       1         19
#> 201   9       1         20
#> 202   7       1         20
#> 203   3       1         20
#> 204  12       1         20
#> 205   8       1         20
#> 206   7       1         20
#> 207  11       1         20
#> 208  10       1         20
#> 209   5       1         20
#> 210  10       1         20
#> 211   5       1         21
#> 212   9       1         21
#> 213  12       1         21
#> 214   8       1         21
#> 215   8       1         21
#> 216   9       1         21
#> 217  13       1         21
#> 218   9       1         21
#> 219   7       1         21
#> 220  10       1         21
#> 221   7       1         22
#> 222   5       1         22
#> 223   6       1         22
#> 224  13       1         22
#> 225  11       1         22
#> 226   9       1         22
#> 227   5       1         22
#> 228   8       1         22
#> 229   8       1         22
#> 230  12       1         22
#> 231   7       1         23
#> 232   9       1         23
#> 233   9       1         23
#> 234   7       1         23
#> 235  12       1         23
#> 236  12       1         23
#> 237  12       1         23
#> 238  10       1         23
#> 239  15       1         23
#> 240  13       1         23
#> 241   7       1         24
#> 242   9       1         24
#> 243  11       1         24
#> 244  10       1         24
#> 245   5       1         24
#> 246  13       1         24
#> 247  10       1         24
#> 248   8       1         24
#> 249   8       1         24
#> 250   7       1         24
#> 251   7       1         25
#> 252   5       1         25
#> 253  10       1         25
#> 254   9       1         25
#> 255   8       1         25
#> 256   8       1         25
#> 257  15       1         25
#> 258  10       1         25
#> 259   5       1         25
#> 260   8       1         25
#> 261  12       1         26
#> 262   7       1         26
#> 263   9       1         26
#> 264  12       1         26
#> 265  11       1         26
#> 266  17       1         26
#> 267   4       1         26
#> 268  15       1         26
#> 269  13       1         26
#> 270  12       1         26
#> 271  10       1         27
#> 272   8       1         27
#> 273  11       1         27
#> 274  10       1         27
#> 275  11       1         27
#> 276  13       1         27
#> 277  18       1         27
#> 278   7       1         27
#> 279  16       1         27
#> 280  12       1         27
#> 281  10       1         28
#> 282  12       1         28
#> 283  14       1         28
#> 284  14       1         28
#> 285   8       1         28
#> 286   7       1         28
#> 287   6       1         28
#> 288   9       1         28
#> 289  15       1         28
#> 290   9       1         28
#> 291  13       1         29
#> 292  12       1         29
#> 293  14       1         29
#> 294  16       1         29
#> 295  12       1         29
#> 296   4       1         29
#> 297  15       1         29
#> 298  13       1         29
#> 299  11       1         29
#> 300  14       1         29
#> 301   9       1         30
#> 302  17       1         30
#> 303   6       1         30
#> 304  13       1         30
#> 305  15       1         30
#> 306  14       1         30
#> 307   9       1         30
#> 308  19       1         30
#> 309  12       1         30
#> 310  13       1         30
#> 311  12       1         31
#> 312   9       1         31
#> 313   8       1         31
#> 314  12       1         31
#> 315  10       1         31
#> 316  10       1         31
#> 317   7       1         31
#> 318  15       1         31
#> 319  11       1         31
#> 320   8       1         31
#> 321  13       1         32
#> 322   7       1         32
#> 323   7       1         32
#> 324   9       1         32
#> 325  16       1         32
#> 326  13       1         32
#> 327  13       1         32
#> 328  13       1         32
#> 329  13       1         32
#> 330   7       1         32
#> 331  15       1         33
#> 332  14       1         33
#> 333  12       1         33
#> 334  11       1         33
#> 335   9       1         33
#> 336  14       1         33
#> 337   8       1         33
#> 338  10       1         33
#> 339  13       1         33
#> 340  11       1         33
#> 341   7       1         34
#> 342   9       1         34
#> 343  17       1         34
#> 344  16       1         34
#> 345  10       1         34
#> 346  16       1         34
#> 347  11       1         34
#> 348  13       1         34
#> 349  11       1         34
#> 350  10       1         34
#> 351  14       1         35
#> 352  13       1         35
#> 353  13       1         35
#> 354  18       1         35
#> 355  17       1         35
#> 356  21       1         35
#> 357  11       1         35
#> 358  12       1         35
#> 359  11       1         35
#> 360  11       1         35
#> 361   9       1         36
#> 362  21       1         36
#> 363   9       1         36
#> 364  10       1         36
#> 365  14       1         36
#> 366  15       1         36
#> 367  17       1         36
#> 368  15       1         36
#> 369  13       1         36
#> 370  13       1         36
#> 371  14       1         37
#> 372  16       1         37
#> 373  11       1         37
#> 374  13       1         37
#> 375  14       1         37
#> 376  19       1         37
#> 377  13       1         37
#> 378  13       1         37
#> 379  19       1         37
#> 380  16       1         37
#> 381  13       1         38
#> 382  11       1         38
#> 383  14       1         38
#> 384  17       1         38
#> 385  15       1         38
#> 386  16       1         38
#> 387  14       1         38
#> 388  12       1         38
#> 389  23       1         38
#> 390  17       1         38
#> 391   9       1         39
#> 392  15       1         39
#> 393  13       1         39
#> 394  15       1         39
#> 395  18       1         39
#> 396  15       1         39
#> 397  14       1         39
#> 398  16       1         39
#> 399  17       1         39
#> 400  13       1         39
#> 401  14       1         40
#> 402  13       1         40
#> 403  13       1         40
#> 404  16       1         40
#> 405  11       1         40
#> 406  18       1         40
#> 407  19       1         40
#> 408  16       1         40
#> 409  16       1         40
#> 410  12       1         40
#> 411  16       1         41
#> 412  15       1         41
#> 413  16       1         41
#> 414  18       1         41
#> 415  21       1         41
#> 416  14       1         41
#> 417  22       1         41
#> 418  20       1         41
#> 419  12       1         41
#> 420  17       1         41
#> 421  19       1         42
#> 422  16       1         42
#> 423  16       1         42
#> 424  14       1         42
#> 425  20       1         42
#> 426  20       1         42
#> 427  18       1         42
#> 428  22       1         42
#> 429  19       1         42
#> 430  23       1         42
#> 431  25       2         43
#> 432  20       1         43
#> 433  22       1         43
#> 434  18       1         43
#> 435  17       1         43
#> 436  20       1         43
#> 437  18       1         43
#> 438  13       1         43
#> 439  18       1         43
#> 440  18       1         43
#> 441  21       1         44
#> 442  15       1         44
#> 443  22       1         44
#> 444  22       1         44
#> 445  19       1         44
#> 446  18       1         44
#> 447  19       1         44
#> 448   8       1         44
#> 449  18       1         44
#> 450  20       1         44
#> 451  24       1         45
#> 452  18       1         45
#> 453  20       1         45
#> 454  18       1         45
#> 455  20       1         45
#> 456  16       1         45
#> 457  21       1         45
#> 458  18       1         45
#> 459  15       1         45
#> 460  12       1         45
#> 461  19       1         46
#> 462  14       1         46
#> 463  19       1         46
#> 464  21       1         46
#> 465  26       2         46
#> 466  19       1         46
#> 467  14       1         46
#> 468  17       1         46
#> 469  19       1         46
#> 470  14       1         46
#> 471  20       1         47
#> 472  20       1         47
#> 473  18       1         47
#> 474  19       1         47
#> 475  19       1         47
#> 476  19       1         47
#> 477   9       1         47
#> 478  13       1         47
#> 479  17       1         47
#> 480  21       1         47
#> 481  19       1         48
#> 482  20       1         48
#> 483  23       1         48
#> 484  22       1         48
#> 485  17       1         48
#> 486  20       1         48
#> 487  21       1         48
#> 488  16       1         48
#> 489  16       1         48
#> 490  20       1         48
#> 491  20       1         49
#> 492  24       1         49
#> 493  16       1         49
#> 494  17       1         49
#> 495  17       1         49
#> 496  20       1         49
#> 497  23       1         49
#> 498  21       1         49
#> 499  20       1         49
#> 500  24       1         49
#> 501  22       1         50
#> 502  18       1         50
#> 503  21       1         50
#> 504  18       1         50
#> 505  20       1         50
#> 506  13       1         50
#> 507  19       1         50
#> 508  19       1         50
#> 509  20       1         50
#> 510  20       1         50
#> 511  17       1         51
#> 512  24       1         51
#> 513  20       1         51
#> 514  27       2         51
#> 515  22       1         51
#> 516  16       1         51
#> 517  18       1         51
#> 518  18       1         51
#> 519  18       1         51
#> 520  23       1         51
#> 521  24       1         52
#> 522  21       1         52
#> 523  26       2         52
#> 524  21       1         52
#> 525  20       1         52
#> 526  17       1         52
#> 527  23       1         52
#> 528  20       1         52
#> 529  24       1         52
#> 530  21       1         52
#> 531  16       1         53
#> 532  22       1         53
#> 533  22       1         53
#> 534  21       1         53
#> 535  18       1         53
#> 536  21       1         53
#> 537  15       1         53
#> 538  20       1         53
#> 539  21       1         53
#> 540  23       1         53
#> 541  17       1         54
#> 542  19       1         54
#> 543  22       1         54
#> 544  27       2         54
#> 545  20       1         54
#> 546  24       1         54
#> 547  22       1         54
#> 548  22       1         54
#> 549  26       2         54
#> 550  22       1         54
#> 551  26       2         55
#> 552  20       1         55
#> 553  22       1         55
#> 554  22       1         55
#> 555  23       1         55
#> 556  20       1         55
#> 557  21       1         55
#> 558  20       1         55
#> 559  20       1         55
#> 560  19       1         55
#> 561  22       1         56
#> 562  28       2         56
#> 563  18       1         56
#> 564  18       1         56
#> 565  24       1         56
#> 566  20       1         56
#> 567  25       2         56
#> 568  23       1         56
#> 569  26       2         56
#> 570  25       2         56
#> 571  21       1         57
#> 572  23       1         57
#> 573  26       2         57
#> 574  25       2         57
#> 575  21       1         57
#> 576  21       1         57
#> 577  22       1         57
#> 578  21       1         57
#> 579  12       1         57
#> 580  22       1         57
#> 581  20       1         58
#> 582  26       2         58
#> 583  25       2         58
#> 584  22       1         58
#> 585  25       2         58
#> 586  21       1         58
#> 587  19       1         58
#> 588  20       1         58
#> 589  25       2         58
#> 590  27       2         58
#> 591  23       1         59
#> 592  26       2         59
#> 593  25       2         59
#> 594  25       2         59
#> 595  29       2         59
#> 596  19       1         59
#> 597  22       1         59
#> 598  20       1         59
#> 599  27       2         59
#> 600  26       2         59
#> 601  26       2         60
#> 602  25       2         60
#> 603  27       2         60
#> 604  29       2         60
#> 605  26       2         60
#> 606  24       1         60
#> 607  21       1         60
#> 608  25       2         60
#> 609  25       2         60
#> 610  21       1         60
#> 611  20       1         61
#> 612  22       1         61
#> 613  25       2         61
#> 614  28       2         61
#> 615  25       2         61
#> 616  23       1         61
#> 617  28       2         61
#> 618  27       2         61
#> 619  20       1         61
#> 620  21       1         61
#> 621  30       2         62
#> 622  26       2         62
#> 623  27       2         62
#> 624  27       2         62
#> 625  24       1         62
#> 626  29       2         62
#> 627  21       1         62
#> 628  27       2         62
#> 629  28       2         62
#> 630  23       1         62
#> 631  24       1         63
#> 632  26       2         63
#> 633  24       1         63
#> 634  22       1         63
#> 635  27       2         63
#> 636  29       2         63
#> 637  28       2         63
#> 638  21       1         63
#> 639  27       2         63
#> 640  29       2         63
#> 641  28       2         64
#> 642  27       2         64
#> 643  29       2         64
#> 644  24       1         64
#> 645  24       1         64
#> 646  25       2         64
#> 647  22       1         64
#> 648  27       2         64
#> 649  28       2         64
#> 650  23       1         64
#> 651  18       1         65
#> 652  25       2         65
#> 653  20       1         65
#> 654  22       1         65
#> 655  21       1         65
#> 656  23       1         65
#> 657  26       2         65
#> 658  25       2         65
#> 659  26       2         65
#> 660  26       2         65
#> 661  28       2         66
#> 662  27       2         66
#> 663  32       2         66
#> 664  31       2         66
#> 665  27       2         66
#> 666  28       2         66
#> 667  23       1         66
#> 668  26       2         66
#> 669  27       2         66
#> 670  24       1         66
#> 671  30       2         67
#> 672  23       1         67
#> 673  28       2         67
#> 674  27       2         67
#> 675  28       2         67
#> 676  29       2         67
#> 677  34       2         67
#> 678  29       2         67
#> 679  28       2         67
#> 680  27       2         67
#> 681  26       2         68
#> 682  26       2         68
#> 683  24       1         68
#> 684  27       2         68
#> 685  32       2         68
#> 686  28       2         68
#> 687  20       1         68
#> 688  29       2         68
#> 689  29       2         68
#> 690  22       1         68
#> 691  26       2         69
#> 692  26       2         69
#> 693  29       2         69
#> 694  25       2         69
#> 695  28       2         69
#> 696  23       1         69
#> 697  28       2         69
#> 698  28       2         69
#> 699  32       2         69
#> 700  27       2         69
#> 701  28       2         70
#> 702  32       2         70
#> 703  23       1         70
#> 704  29       2         70
#> 705  26       2         70
#> 706  28       2         70
#> 707  26       2         70
#> 708  30       2         70
#> 709  31       2         70
#> 710  26       2         70
#> 711  23       1         71
#> 712  30       2         71
#> 713  31       2         71
#> 714  29       2         71
#> 715  33       2         71
#> 716  30       2         71
#> 717  31       2         71
#> 718  29       2         71
#> 719  32       2         71
#> 720  30       2         71
#> 721  28       2         72
#> 722  28       2         72
#> 723  28       2         72
#> 724  32       2         72
#> 725  32       2         72
#> 726  26       2         72
#> 727  29       2         72
#> 728  28       2         72
#> 729  28       2         72
#> 730  32       2         72
#> 731  25       2         73
#> 732  34       2         73
#> 733  33       2         73
#> 734  26       2         73
#> 735  32       2         73
#> 736  31       2         73
#> 737  30       2         73
#> 738  25       2         73
#> 739  28       2         73
#> 740  32       2         73
#> 741  29       2         74
#> 742  25       2         74
#> 743  32       2         74
#> 744  34       2         74
#> 745  32       2         74
#> 746  30       2         74
#> 747  29       2         74
#> 748  30       2         74
#> 749  33       2         74
#> 750  30       2         74
#> 751  29       2         75
#> 752  26       2         75
#> 753  31       2         75
#> 754  32       2         75
#> 755  32       2         75
#> 756  27       2         75
#> 757  33       2         75
#> 758  32       2         75
#> 759  31       2         75
#> 760  25       2         75
#> 761  33       2         76
#> 762  31       2         76
#> 763  29       2         76
#> 764  30       2         76
#> 765  34       2         76
#> 766  32       2         76
#> 767  33       2         76
#> 768  33       2         76
#> 769  35       2         76
#> 770  27       2         76
#> 771  33       2         77
#> 772  36       2         77
#> 773  29       2         77
#> 774  33       2         77
#> 775  35       2         77
#> 776  28       2         77
#> 777  34       2         77
#> 778  31       2         77
#> 779  31       2         77
#> 780  33       2         77
#> 781  29       2         78
#> 782  34       2         78
#> 783  34       2         78
#> 784  29       2         78
#> 785  34       2         78
#> 786  26       2         78
#> 787  35       2         78
#> 788  33       2         78
#> 789  32       2         78
#> 790  33       2         78
#> 791  30       2         79
#> 792  31       2         79
#> 793  36       2         79
#> 794  34       2         79
#> 795  30       2         79
#> 796  29       2         79
#> 797  30       2         79
#> 798  29       2         79
#> 799  34       2         79
#> 800  32       2         79
#> 801  31       2         80
#> 802  32       2         80
#> 803  32       2         80
#> 804  28       2         80
#> 805  34       2         80
#> 806  31       2         80
#> 807  33       2         80
#> 808  36       2         80
#> 809  36       2         80
#> 810  34       2         80
#> 811  28       2         81
#> 812  30       2         81
#> 813  33       2         81
#> 814  30       2         81
#> 815  34       2         81
#> 816  32       2         81
#> 817  34       2         81
#> 818  34       2         81
#> 819  34       2         81
#> 820  33       2         81
#> 821  35       2         82
#> 822  35       2         82
#> 823  30       2         82
#> 824  32       2         82
#> 825  30       2         82
#> 826  35       2         82
#> 827  32       2         82
#> 828  32       2         82
#> 829  33       2         82
#> 830  36       2         82
#> 831  38       3         83
#> 832  31       2         83
#> 833  33       2         83
#> 834  33       2         83
#> 835  31       2         83
#> 836  32       2         83
#> 837  36       2         83
#> 838  26       2         83
#> 839  36       2         83
#> 840  31       2         83
#> 841  34       2         84
#> 842  34       2         84
#> 843  34       2         84
#> 844  30       2         84
#> 845  34       2         84
#> 846  32       2         84
#> 847  32       2         84
#> 848  32       2         84
#> 849  31       2         84
#> 850  32       2         84
#> 851  33       2         85
#> 852  36       2         85
#> 853  37       3         85
#> 854  36       2         85
#> 855  28       2         85
#> 856  37       3         85
#> 857  36       2         85
#> 858  39       3         85
#> 859  38       3         85
#> 860  34       2         85
#> 861  30       2         86
#> 862  35       2         86
#> 863  36       2         86
#> 864  37       3         86
#> 865  37       3         86
#> 866  35       2         86
#> 867  38       3         86
#> 868  35       2         86
#> 869  36       2         86
#> 870  35       2         86
#> 871  35       2         87
#> 872  38       3         87
#> 873  34       2         87
#> 874  36       2         87
#> 875  34       2         87
#> 876  35       2         87
#> 877  36       2         87
#> 878  30       2         87
#> 879  36       2         87
#> 880  33       2         87
#> 881  37       3         88
#> 882  37       3         88
#> 883  38       3         88
#> 884  38       3         88
#> 885  32       2         88
#> 886  37       3         88
#> 887  32       2         88
#> 888  39       3         88
#> 889  33       2         88
#> 890  33       2         88
#> 891  35       2         89
#> 892  36       2         89
#> 893  37       3         89
#> 894  37       3         89
#> 895  28       2         89
#> 896  37       3         89
#> 897  37       3         89
#> 898  36       2         89
#> 899  37       3         89
#> 900  35       2         89
#> 901  38       3         90
#> 902  37       3         90
#> 903  40       3         90
#> 904  34       2         90
#> 905  38       3         90
#> 906  34       2         90
#> 907  36       2         90
#> 908  36       2         90
#> 909  36       2         90
#> 910  38       3         90
#> 911  38       3         91
#> 912  38       3         91
#> 913  35       2         91
#> 914  36       2         91
#> 915  37       3         91
#> 916  39       3         91
#> 917  35       2         91
#> 918  34       2         91
#> 919  37       3         91
#> 920  36       2         91
#> 921  39       3         92
#> 922  38       3         92
#> 923  39       3         92
#> 924  39       3         92
#> 925  37       3         92
#> 926  39       3         92
#> 927  37       3         92
#> 928  38       3         92
#> 929  36       2         92
#> 930  39       3         92
#> 931  39       3         93
#> 932  37       3         93
#> 933  37       3         93
#> 934  37       3         93
#> 935  38       3         93
#> 936  37       3         93
#> 937  39       3         93
#> 938  37       3         93
#> 939  38       3         93
#> 940  40       3         93
#> 941  36       2         94
#> 942  40       3         94
#> 943  39       3         94
#> 944  40       3         94
#> 945  35       2         94
#> 946  37       3         94
#> 947  37       3         94
#> 948  39       3         94
#> 949  37       3         94
#> 950  37       3         94
#> 951  38       3         95
#> 952  40       3         95
#> 953  37       3         95
#> 954  39       3         95
#> 955  37       3         95
#> 956  36       2         95
#> 957  38       3         95
#> 958  36       2         95
#> 959  36       2         95
#> 960  37       3         95
#> 961  39       3         96
#> 962  37       3         96
#> 963  40       3         96
#> 964  38       3         96
#> 965  35       2         96
#> 966  37       3         96
#> 967  37       3         96
#> 968  38       3         96
#> 969  35       2         96
#> 970  37       3         96
#> 971  39       3         97
#> 972  40       3         97
#> 973  37       3         97
#> 974  36       2         97
#> 975  40       3         97
#> 976  37       3         97
#> 977  39       3         97
#> 978  40       3         97
#> 979  40       3         97
#> 980  39       3         97
#> 981  39       3         98
#> 982  40       3         98
#> 983  37       3         98
#> 984  40       3         98
#> 985  40       3         98
#> 986  40       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  39       3         98
#> 990  39       3         98
#> 991  39       3         99
#> 992  40       3         99
#> 993  40       3         99
#> 994  39       3         99
#> 995  39       3         99
#> 996  40       3         99
#> 997  38       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100