Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    3       1          1
#> 12    0       1          1
#> 13    0       1          1
#> 14    0       1          1
#> 15    0       1          1
#> 16    1       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    0       1          2
#> 23    1       1          2
#> 24    1       1          2
#> 25    3       1          2
#> 26    0       1          2
#> 27    0       1          2
#> 28    3       1          2
#> 29    0       1          2
#> 30    2       1          2
#> 31    0       1          3
#> 32    2       1          3
#> 33    0       1          3
#> 34    0       1          3
#> 35    2       1          3
#> 36    2       1          3
#> 37    1       1          3
#> 38    1       1          3
#> 39    1       1          3
#> 40    3       1          3
#> 41    3       1          4
#> 42    3       1          4
#> 43    2       1          4
#> 44    2       1          4
#> 45    2       1          4
#> 46    3       1          4
#> 47    1       1          4
#> 48    2       1          4
#> 49    0       1          4
#> 50    0       1          4
#> 51    2       1          5
#> 52    1       1          5
#> 53    1       1          5
#> 54    1       1          5
#> 55    3       1          5
#> 56    3       1          5
#> 57    3       1          5
#> 58    0       1          5
#> 59    2       1          5
#> 60    3       1          5
#> 61    1       1          6
#> 62    2       1          6
#> 63    4       1          6
#> 64    2       1          6
#> 65    1       1          6
#> 66    4       1          6
#> 67    4       1          6
#> 68    4       1          6
#> 69    0       1          6
#> 70    3       1          6
#> 71    3       1          7
#> 72    4       1          7
#> 73    0       1          7
#> 74    5       1          7
#> 75    2       1          7
#> 76    4       1          7
#> 77    2       1          7
#> 78    1       1          7
#> 79    3       1          7
#> 80    2       1          7
#> 81    1       1          8
#> 82    3       1          8
#> 83    2       1          8
#> 84    4       1          8
#> 85    6       1          8
#> 86    2       1          8
#> 87    2       1          8
#> 88    4       1          8
#> 89    4       1          8
#> 90    3       1          8
#> 91    2       1          9
#> 92    3       1          9
#> 93    8       1          9
#> 94    2       1          9
#> 95    6       1          9
#> 96    6       1          9
#> 97    5       1          9
#> 98    3       1          9
#> 99    7       1          9
#> 100   5       1          9
#> 101   2       1         10
#> 102   3       1         10
#> 103   6       1         10
#> 104   4       1         10
#> 105   1       1         10
#> 106   5       1         10
#> 107   3       1         10
#> 108   3       1         10
#> 109   4       1         10
#> 110   2       1         10
#> 111   3       1         11
#> 112   7       1         11
#> 113   3       1         11
#> 114   2       1         11
#> 115   6       1         11
#> 116   7       1         11
#> 117   3       1         11
#> 118   6       1         11
#> 119   2       1         11
#> 120   1       1         11
#> 121   7       1         12
#> 122   5       1         12
#> 123   5       1         12
#> 124   4       1         12
#> 125  10       1         12
#> 126   3       1         12
#> 127   9       1         12
#> 128   5       1         12
#> 129   4       1         12
#> 130   6       1         12
#> 131   5       1         13
#> 132   3       1         13
#> 133   4       1         13
#> 134   4       1         13
#> 135   5       1         13
#> 136   4       1         13
#> 137   3       1         13
#> 138   5       1         13
#> 139   6       1         13
#> 140   9       1         13
#> 141   3       1         14
#> 142   5       1         14
#> 143   8       1         14
#> 144   3       1         14
#> 145   4       1         14
#> 146   4       1         14
#> 147   7       1         14
#> 148   8       1         14
#> 149   6       1         14
#> 150   5       1         14
#> 151   6       1         15
#> 152   4       1         15
#> 153  12       1         15
#> 154   9       1         15
#> 155   5       1         15
#> 156   6       1         15
#> 157   5       1         15
#> 158   6       1         15
#> 159   3       1         15
#> 160   7       1         15
#> 161   7       1         16
#> 162   9       1         16
#> 163   5       1         16
#> 164   7       1         16
#> 165   6       1         16
#> 166   5       1         16
#> 167   9       1         16
#> 168   7       1         16
#> 169   2       1         16
#> 170   6       1         16
#> 171   3       1         17
#> 172   8       1         17
#> 173  10       1         17
#> 174   5       1         17
#> 175   7       1         17
#> 176   5       1         17
#> 177  11       1         17
#> 178   9       1         17
#> 179   3       1         17
#> 180   7       1         17
#> 181   4       1         18
#> 182   7       1         18
#> 183   5       1         18
#> 184  11       1         18
#> 185   6       1         18
#> 186   7       1         18
#> 187   8       1         18
#> 188   9       1         18
#> 189   4       1         18
#> 190   2       1         18
#> 191   4       1         19
#> 192   7       1         19
#> 193   9       1         19
#> 194   6       1         19
#> 195   7       1         19
#> 196  12       1         19
#> 197   7       1         19
#> 198  11       1         19
#> 199  10       1         19
#> 200  10       1         19
#> 201  11       1         20
#> 202   6       1         20
#> 203   6       1         20
#> 204   8       1         20
#> 205  10       1         20
#> 206   9       1         20
#> 207  10       1         20
#> 208  10       1         20
#> 209  11       1         20
#> 210   8       1         20
#> 211   7       1         21
#> 212   7       1         21
#> 213  11       1         21
#> 214   9       1         21
#> 215   4       1         21
#> 216   9       1         21
#> 217  11       1         21
#> 218   7       1         21
#> 219   7       1         21
#> 220   8       1         21
#> 221   7       1         22
#> 222   8       1         22
#> 223  12       1         22
#> 224  11       1         22
#> 225  10       1         22
#> 226  11       1         22
#> 227   8       1         22
#> 228  15       1         22
#> 229   6       1         22
#> 230   6       1         22
#> 231   8       1         23
#> 232  11       1         23
#> 233  12       1         23
#> 234  10       1         23
#> 235   6       1         23
#> 236   9       1         23
#> 237  10       1         23
#> 238  10       1         23
#> 239   7       1         23
#> 240   7       1         23
#> 241   4       1         24
#> 242  10       1         24
#> 243   8       1         24
#> 244   6       1         24
#> 245   7       1         24
#> 246  10       1         24
#> 247  14       1         24
#> 248   7       1         24
#> 249  11       1         24
#> 250  13       1         24
#> 251  10       1         25
#> 252  10       1         25
#> 253  12       1         25
#> 254   8       1         25
#> 255   8       1         25
#> 256  10       1         25
#> 257  10       1         25
#> 258  11       1         25
#> 259   7       1         25
#> 260  10       1         25
#> 261  12       1         26
#> 262   9       1         26
#> 263   8       1         26
#> 264  12       1         26
#> 265   7       1         26
#> 266  10       1         26
#> 267   8       1         26
#> 268   9       1         26
#> 269  11       1         26
#> 270   8       1         26
#> 271   8       1         27
#> 272   6       1         27
#> 273   6       1         27
#> 274  14       1         27
#> 275  10       1         27
#> 276   6       1         27
#> 277   4       1         27
#> 278  12       1         27
#> 279   8       1         27
#> 280  12       1         27
#> 281  15       1         28
#> 282  17       1         28
#> 283  11       1         28
#> 284  10       1         28
#> 285  15       1         28
#> 286  13       1         28
#> 287  16       1         28
#> 288   8       1         28
#> 289  10       1         28
#> 290   7       1         28
#> 291  12       1         29
#> 292   8       1         29
#> 293  11       1         29
#> 294  12       1         29
#> 295  11       1         29
#> 296   9       1         29
#> 297  10       1         29
#> 298  10       1         29
#> 299  14       1         29
#> 300  13       1         29
#> 301  10       1         30
#> 302   9       1         30
#> 303  14       1         30
#> 304  11       1         30
#> 305  17       1         30
#> 306  11       1         30
#> 307  18       1         30
#> 308   8       1         30
#> 309  13       1         30
#> 310  10       1         30
#> 311   7       1         31
#> 312  11       1         31
#> 313  15       1         31
#> 314  17       1         31
#> 315   5       1         31
#> 316  12       1         31
#> 317  14       1         31
#> 318   8       1         31
#> 319   9       1         31
#> 320  11       1         31
#> 321  13       1         32
#> 322  14       1         32
#> 323  16       1         32
#> 324  12       1         32
#> 325  15       1         32
#> 326  13       1         32
#> 327  15       1         32
#> 328  13       1         32
#> 329   9       1         32
#> 330  18       1         32
#> 331  15       1         33
#> 332  14       1         33
#> 333  17       1         33
#> 334  12       1         33
#> 335   8       1         33
#> 336  15       1         33
#> 337  14       1         33
#> 338  10       1         33
#> 339  10       1         33
#> 340  10       1         33
#> 341  15       1         34
#> 342  15       1         34
#> 343  16       1         34
#> 344  17       1         34
#> 345  10       1         34
#> 346  12       1         34
#> 347  20       1         34
#> 348  16       1         34
#> 349  18       1         34
#> 350  16       1         34
#> 351  10       1         35
#> 352  20       1         35
#> 353   9       1         35
#> 354  13       1         35
#> 355  11       1         35
#> 356   9       1         35
#> 357  10       1         35
#> 358  16       1         35
#> 359  18       1         35
#> 360  14       1         35
#> 361  15       1         36
#> 362  18       1         36
#> 363  19       1         36
#> 364  16       1         36
#> 365  12       1         36
#> 366  19       1         36
#> 367  13       1         36
#> 368  11       1         36
#> 369  17       1         36
#> 370  17       1         36
#> 371  14       1         37
#> 372  13       1         37
#> 373  13       1         37
#> 374  15       1         37
#> 375  16       1         37
#> 376  15       1         37
#> 377  15       1         37
#> 378  16       1         37
#> 379  17       1         37
#> 380  15       1         37
#> 381  17       1         38
#> 382  10       1         38
#> 383  14       1         38
#> 384  19       1         38
#> 385  13       1         38
#> 386  11       1         38
#> 387  15       1         38
#> 388  13       1         38
#> 389  17       1         38
#> 390  12       1         38
#> 391  15       1         39
#> 392  16       1         39
#> 393  13       1         39
#> 394  14       1         39
#> 395  20       1         39
#> 396  17       1         39
#> 397   9       1         39
#> 398  15       1         39
#> 399  11       1         39
#> 400  14       1         39
#> 401  10       1         40
#> 402  20       1         40
#> 403  10       1         40
#> 404  13       1         40
#> 405  23       1         40
#> 406  14       1         40
#> 407  17       1         40
#> 408  17       1         40
#> 409  17       1         40
#> 410  18       1         40
#> 411  14       1         41
#> 412  20       1         41
#> 413  18       1         41
#> 414  18       1         41
#> 415  13       1         41
#> 416  13       1         41
#> 417  16       1         41
#> 418  19       1         41
#> 419  16       1         41
#> 420  12       1         41
#> 421  14       1         42
#> 422  16       1         42
#> 423  15       1         42
#> 424  21       1         42
#> 425  19       1         42
#> 426  12       1         42
#> 427  16       1         42
#> 428  14       1         42
#> 429  16       1         42
#> 430  18       1         42
#> 431  17       1         43
#> 432  17       1         43
#> 433  20       1         43
#> 434  14       1         43
#> 435  15       1         43
#> 436  19       1         43
#> 437  16       1         43
#> 438  16       1         43
#> 439  15       1         43
#> 440  11       1         43
#> 441  21       1         44
#> 442  13       1         44
#> 443  17       1         44
#> 444  18       1         44
#> 445  19       1         44
#> 446  20       1         44
#> 447  20       1         44
#> 448  19       1         44
#> 449  17       1         44
#> 450  25       2         44
#> 451  22       1         45
#> 452  17       1         45
#> 453  22       1         45
#> 454  20       1         45
#> 455  16       1         45
#> 456  19       1         45
#> 457  20       1         45
#> 458  21       1         45
#> 459  23       1         45
#> 460  15       1         45
#> 461  18       1         46
#> 462  19       1         46
#> 463  14       1         46
#> 464  12       1         46
#> 465  22       1         46
#> 466  19       1         46
#> 467  15       1         46
#> 468  13       1         46
#> 469  21       1         46
#> 470  17       1         46
#> 471  23       1         47
#> 472  15       1         47
#> 473  21       1         47
#> 474  25       2         47
#> 475  19       1         47
#> 476  16       1         47
#> 477  21       1         47
#> 478  16       1         47
#> 479  19       1         47
#> 480  22       1         47
#> 481  16       1         48
#> 482  15       1         48
#> 483  28       2         48
#> 484  23       1         48
#> 485  16       1         48
#> 486  20       1         48
#> 487  17       1         48
#> 488  14       1         48
#> 489  15       1         48
#> 490  21       1         48
#> 491  18       1         49
#> 492  20       1         49
#> 493  16       1         49
#> 494  20       1         49
#> 495  15       1         49
#> 496  14       1         49
#> 497  21       1         49
#> 498  20       1         49
#> 499  16       1         49
#> 500  21       1         49
#> 501  20       1         50
#> 502  16       1         50
#> 503  22       1         50
#> 504  28       2         50
#> 505  18       1         50
#> 506  20       1         50
#> 507  17       1         50
#> 508  22       1         50
#> 509  23       1         50
#> 510  17       1         50
#> 511  20       1         51
#> 512  21       1         51
#> 513  16       1         51
#> 514  24       1         51
#> 515  24       1         51
#> 516  17       1         51
#> 517  15       1         51
#> 518  25       2         51
#> 519  19       1         51
#> 520  23       1         51
#> 521  24       1         52
#> 522  24       1         52
#> 523  21       1         52
#> 524  25       2         52
#> 525  23       1         52
#> 526  14       1         52
#> 527  18       1         52
#> 528  18       1         52
#> 529  14       1         52
#> 530  11       1         52
#> 531  24       1         53
#> 532  20       1         53
#> 533  23       1         53
#> 534  24       1         53
#> 535  20       1         53
#> 536  21       1         53
#> 537  22       1         53
#> 538  19       1         53
#> 539  20       1         53
#> 540  22       1         53
#> 541  28       2         54
#> 542  22       1         54
#> 543  21       1         54
#> 544  24       1         54
#> 545  23       1         54
#> 546  22       1         54
#> 547  24       1         54
#> 548  22       1         54
#> 549  24       1         54
#> 550  16       1         54
#> 551  20       1         55
#> 552  23       1         55
#> 553  27       2         55
#> 554  21       1         55
#> 555  23       1         55
#> 556  21       1         55
#> 557  21       1         55
#> 558  21       1         55
#> 559  26       2         55
#> 560  22       1         55
#> 561  25       2         56
#> 562  16       1         56
#> 563  20       1         56
#> 564  23       1         56
#> 565  23       1         56
#> 566  22       1         56
#> 567  21       1         56
#> 568  20       1         56
#> 569  20       1         56
#> 570  23       1         56
#> 571  23       1         57
#> 572  28       2         57
#> 573  18       1         57
#> 574  27       2         57
#> 575  22       1         57
#> 576  27       2         57
#> 577  31       2         57
#> 578  27       2         57
#> 579  27       2         57
#> 580  19       1         57
#> 581  23       1         58
#> 582  16       1         58
#> 583  22       1         58
#> 584  26       2         58
#> 585  24       1         58
#> 586  24       1         58
#> 587  21       1         58
#> 588  23       1         58
#> 589  25       2         58
#> 590  24       1         58
#> 591  26       2         59
#> 592  24       1         59
#> 593  22       1         59
#> 594  25       2         59
#> 595  25       2         59
#> 596  23       1         59
#> 597  21       1         59
#> 598  27       2         59
#> 599  12       1         59
#> 600  21       1         59
#> 601  23       1         60
#> 602  16       1         60
#> 603  23       1         60
#> 604  22       1         60
#> 605  20       1         60
#> 606  31       2         60
#> 607  24       1         60
#> 608  26       2         60
#> 609  29       2         60
#> 610  21       1         60
#> 611  22       1         61
#> 612  17       1         61
#> 613  23       1         61
#> 614  19       1         61
#> 615  26       2         61
#> 616  25       2         61
#> 617  23       1         61
#> 618  26       2         61
#> 619  26       2         61
#> 620  25       2         61
#> 621  26       2         62
#> 622  28       2         62
#> 623  22       1         62
#> 624  29       2         62
#> 625  25       2         62
#> 626  24       1         62
#> 627  25       2         62
#> 628  28       2         62
#> 629  20       1         62
#> 630  28       2         62
#> 631  24       1         63
#> 632  27       2         63
#> 633  22       1         63
#> 634  22       1         63
#> 635  24       1         63
#> 636  27       2         63
#> 637  21       1         63
#> 638  27       2         63
#> 639  21       1         63
#> 640  21       1         63
#> 641  14       1         64
#> 642  26       2         64
#> 643  21       1         64
#> 644  23       1         64
#> 645  26       2         64
#> 646  26       2         64
#> 647  26       2         64
#> 648  22       1         64
#> 649  24       1         64
#> 650  27       2         64
#> 651  24       1         65
#> 652  31       2         65
#> 653  30       2         65
#> 654  29       2         65
#> 655  25       2         65
#> 656  26       2         65
#> 657  32       2         65
#> 658  26       2         65
#> 659  30       2         65
#> 660  29       2         65
#> 661  27       2         66
#> 662  24       1         66
#> 663  30       2         66
#> 664  28       2         66
#> 665  22       1         66
#> 666  30       2         66
#> 667  22       1         66
#> 668  27       2         66
#> 669  32       2         66
#> 670  20       1         66
#> 671  26       2         67
#> 672  27       2         67
#> 673  22       1         67
#> 674  29       2         67
#> 675  22       1         67
#> 676  26       2         67
#> 677  26       2         67
#> 678  32       2         67
#> 679  21       1         67
#> 680  23       1         67
#> 681  27       2         68
#> 682  26       2         68
#> 683  23       1         68
#> 684  24       1         68
#> 685  27       2         68
#> 686  28       2         68
#> 687  31       2         68
#> 688  30       2         68
#> 689  24       1         68
#> 690  26       2         68
#> 691  29       2         69
#> 692  31       2         69
#> 693  26       2         69
#> 694  23       1         69
#> 695  25       2         69
#> 696  30       2         69
#> 697  28       2         69
#> 698  29       2         69
#> 699  23       1         69
#> 700  25       2         69
#> 701  28       2         70
#> 702  26       2         70
#> 703  30       2         70
#> 704  27       2         70
#> 705  29       2         70
#> 706  25       2         70
#> 707  22       1         70
#> 708  24       1         70
#> 709  28       2         70
#> 710  25       2         70
#> 711  33       2         71
#> 712  36       2         71
#> 713  30       2         71
#> 714  30       2         71
#> 715  25       2         71
#> 716  29       2         71
#> 717  21       1         71
#> 718  31       2         71
#> 719  28       2         71
#> 720  31       2         71
#> 721  26       2         72
#> 722  27       2         72
#> 723  25       2         72
#> 724  31       2         72
#> 725  29       2         72
#> 726  29       2         72
#> 727  28       2         72
#> 728  31       2         72
#> 729  32       2         72
#> 730  30       2         72
#> 731  29       2         73
#> 732  28       2         73
#> 733  28       2         73
#> 734  30       2         73
#> 735  31       2         73
#> 736  27       2         73
#> 737  29       2         73
#> 738  30       2         73
#> 739  31       2         73
#> 740  25       2         73
#> 741  25       2         74
#> 742  30       2         74
#> 743  25       2         74
#> 744  31       2         74
#> 745  29       2         74
#> 746  33       2         74
#> 747  30       2         74
#> 748  32       2         74
#> 749  34       2         74
#> 750  34       2         74
#> 751  30       2         75
#> 752  29       2         75
#> 753  25       2         75
#> 754  29       2         75
#> 755  32       2         75
#> 756  28       2         75
#> 757  33       2         75
#> 758  33       2         75
#> 759  28       2         75
#> 760  26       2         75
#> 761  30       2         76
#> 762  30       2         76
#> 763  32       2         76
#> 764  34       2         76
#> 765  31       2         76
#> 766  32       2         76
#> 767  31       2         76
#> 768  25       2         76
#> 769  33       2         76
#> 770  29       2         76
#> 771  34       2         77
#> 772  29       2         77
#> 773  29       2         77
#> 774  31       2         77
#> 775  29       2         77
#> 776  31       2         77
#> 777  31       2         77
#> 778  30       2         77
#> 779  31       2         77
#> 780  29       2         77
#> 781  28       2         78
#> 782  32       2         78
#> 783  32       2         78
#> 784  31       2         78
#> 785  33       2         78
#> 786  33       2         78
#> 787  32       2         78
#> 788  31       2         78
#> 789  33       2         78
#> 790  31       2         78
#> 791  33       2         79
#> 792  31       2         79
#> 793  32       2         79
#> 794  28       2         79
#> 795  33       2         79
#> 796  30       2         79
#> 797  30       2         79
#> 798  33       2         79
#> 799  34       2         79
#> 800  35       2         79
#> 801  30       2         80
#> 802  32       2         80
#> 803  32       2         80
#> 804  31       2         80
#> 805  34       2         80
#> 806  30       2         80
#> 807  36       2         80
#> 808  31       2         80
#> 809  31       2         80
#> 810  34       2         80
#> 811  33       2         81
#> 812  31       2         81
#> 813  31       2         81
#> 814  34       2         81
#> 815  35       2         81
#> 816  33       2         81
#> 817  34       2         81
#> 818  33       2         81
#> 819  33       2         81
#> 820  37       3         81
#> 821  33       2         82
#> 822  33       2         82
#> 823  28       2         82
#> 824  31       2         82
#> 825  37       3         82
#> 826  26       2         82
#> 827  33       2         82
#> 828  29       2         82
#> 829  32       2         82
#> 830  32       2         82
#> 831  34       2         83
#> 832  31       2         83
#> 833  35       2         83
#> 834  35       2         83
#> 835  28       2         83
#> 836  34       2         83
#> 837  35       2         83
#> 838  36       2         83
#> 839  32       2         83
#> 840  25       2         83
#> 841  34       2         84
#> 842  36       2         84
#> 843  34       2         84
#> 844  35       2         84
#> 845  36       2         84
#> 846  31       2         84
#> 847  32       2         84
#> 848  35       2         84
#> 849  34       2         84
#> 850  35       2         84
#> 851  32       2         85
#> 852  36       2         85
#> 853  36       2         85
#> 854  28       2         85
#> 855  38       3         85
#> 856  31       2         85
#> 857  26       2         85
#> 858  37       3         85
#> 859  37       3         85
#> 860  33       2         85
#> 861  34       2         86
#> 862  33       2         86
#> 863  35       2         86
#> 864  32       2         86
#> 865  35       2         86
#> 866  33       2         86
#> 867  35       2         86
#> 868  35       2         86
#> 869  32       2         86
#> 870  33       2         86
#> 871  38       3         87
#> 872  34       2         87
#> 873  32       2         87
#> 874  36       2         87
#> 875  36       2         87
#> 876  36       2         87
#> 877  37       3         87
#> 878  33       2         87
#> 879  33       2         87
#> 880  34       2         87
#> 881  34       2         88
#> 882  38       3         88
#> 883  37       3         88
#> 884  34       2         88
#> 885  33       2         88
#> 886  35       2         88
#> 887  36       2         88
#> 888  36       2         88
#> 889  36       2         88
#> 890  36       2         88
#> 891  37       3         89
#> 892  35       2         89
#> 893  34       2         89
#> 894  35       2         89
#> 895  35       2         89
#> 896  38       3         89
#> 897  38       3         89
#> 898  36       2         89
#> 899  39       3         89
#> 900  34       2         89
#> 901  39       3         90
#> 902  36       2         90
#> 903  38       3         90
#> 904  39       3         90
#> 905  38       3         90
#> 906  35       2         90
#> 907  35       2         90
#> 908  37       3         90
#> 909  36       2         90
#> 910  36       2         90
#> 911  37       3         91
#> 912  38       3         91
#> 913  36       2         91
#> 914  35       2         91
#> 915  38       3         91
#> 916  38       3         91
#> 917  38       3         91
#> 918  38       3         91
#> 919  36       2         91
#> 920  35       2         91
#> 921  38       3         92
#> 922  39       3         92
#> 923  37       3         92
#> 924  40       3         92
#> 925  38       3         92
#> 926  35       2         92
#> 927  38       3         92
#> 928  38       3         92
#> 929  40       3         92
#> 930  34       2         92
#> 931  39       3         93
#> 932  37       3         93
#> 933  37       3         93
#> 934  34       2         93
#> 935  39       3         93
#> 936  36       2         93
#> 937  38       3         93
#> 938  36       2         93
#> 939  38       3         93
#> 940  39       3         93
#> 941  38       3         94
#> 942  40       3         94
#> 943  38       3         94
#> 944  38       3         94
#> 945  40       3         94
#> 946  38       3         94
#> 947  33       2         94
#> 948  34       2         94
#> 949  36       2         94
#> 950  39       3         94
#> 951  38       3         95
#> 952  39       3         95
#> 953  38       3         95
#> 954  36       2         95
#> 955  36       2         95
#> 956  39       3         95
#> 957  37       3         95
#> 958  35       2         95
#> 959  40       3         95
#> 960  39       3         95
#> 961  40       3         96
#> 962  40       3         96
#> 963  36       2         96
#> 964  39       3         96
#> 965  39       3         96
#> 966  40       3         96
#> 967  39       3         96
#> 968  40       3         96
#> 969  39       3         96
#> 970  38       3         96
#> 971  39       3         97
#> 972  39       3         97
#> 973  38       3         97
#> 974  39       3         97
#> 975  39       3         97
#> 976  40       3         97
#> 977  40       3         97
#> 978  40       3         97
#> 979  38       3         97
#> 980  38       3         97
#> 981  38       3         98
#> 982  39       3         98
#> 983  40       3         98
#> 984  39       3         98
#> 985  40       3         98
#> 986  40       3         98
#> 987  39       3         98
#> 988  40       3         98
#> 989  40       3         98
#> 990  38       3         98
#> 991  40       3         99
#> 992  39       3         99
#> 993  40       3         99
#> 994  40       3         99
#> 995  40       3         99
#> 996  39       3         99
#> 997  39       3         99
#> 998  39       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100