Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simul_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    0       1          1
#> 13    0       1          1
#> 14    1       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    1       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    0       1          2
#> 23    0       1          2
#> 24    1       1          2
#> 25    0       1          2
#> 26    1       1          2
#> 27    0       1          2
#> 28    2       1          2
#> 29    0       1          2
#> 30    0       1          2
#> 31    2       1          3
#> 32    2       1          3
#> 33    0       1          3
#> 34    1       1          3
#> 35    2       1          3
#> 36    0       1          3
#> 37    3       1          3
#> 38    3       1          3
#> 39    1       1          3
#> 40    0       1          3
#> 41    1       1          4
#> 42    1       1          4
#> 43    1       1          4
#> 44    1       1          4
#> 45    0       1          4
#> 46    0       1          4
#> 47    2       1          4
#> 48    1       1          4
#> 49    1       1          4
#> 50    2       1          4
#> 51    2       1          5
#> 52    2       1          5
#> 53    1       1          5
#> 54    3       1          5
#> 55    9       1          5
#> 56    3       1          5
#> 57    2       1          5
#> 58    5       1          5
#> 59    1       1          5
#> 60    2       1          5
#> 61    4       1          6
#> 62    2       1          6
#> 63    3       1          6
#> 64    1       1          6
#> 65    0       1          6
#> 66    2       1          6
#> 67    2       1          6
#> 68    1       1          6
#> 69    4       1          6
#> 70    2       1          6
#> 71    3       1          7
#> 72    4       1          7
#> 73    4       1          7
#> 74    5       1          7
#> 75    2       1          7
#> 76    6       1          7
#> 77    4       1          7
#> 78    2       1          7
#> 79    1       1          7
#> 80    1       1          7
#> 81    5       1          8
#> 82    6       1          8
#> 83    2       1          8
#> 84    2       1          8
#> 85    3       1          8
#> 86    4       1          8
#> 87    3       1          8
#> 88    3       1          8
#> 89    5       1          8
#> 90    1       1          8
#> 91    1       1          9
#> 92    2       1          9
#> 93    2       1          9
#> 94    3       1          9
#> 95    3       1          9
#> 96    6       1          9
#> 97    3       1          9
#> 98    3       1          9
#> 99    5       1          9
#> 100   3       1          9
#> 101   2       1         10
#> 102   1       1         10
#> 103   7       1         10
#> 104   1       1         10
#> 105   5       1         10
#> 106   4       1         10
#> 107   3       1         10
#> 108   7       1         10
#> 109   1       1         10
#> 110   7       1         10
#> 111   4       1         11
#> 112   4       1         11
#> 113   9       1         11
#> 114   4       1         11
#> 115   6       1         11
#> 116   5       1         11
#> 117   5       1         11
#> 118   7       1         11
#> 119   6       1         11
#> 120   5       1         11
#> 121   5       1         12
#> 122   6       1         12
#> 123   9       1         12
#> 124   5       1         12
#> 125   6       1         12
#> 126   6       1         12
#> 127   2       1         12
#> 128   5       1         12
#> 129   5       1         12
#> 130   2       1         12
#> 131   5       1         13
#> 132   3       1         13
#> 133   4       1         13
#> 134   5       1         13
#> 135   5       1         13
#> 136   6       1         13
#> 137   3       1         13
#> 138   4       1         13
#> 139   3       1         13
#> 140   7       1         13
#> 141   5       1         14
#> 142   5       1         14
#> 143   2       1         14
#> 144   2       1         14
#> 145   6       1         14
#> 146   1       1         14
#> 147   7       1         14
#> 148   5       1         14
#> 149   4       1         14
#> 150   4       1         14
#> 151   8       1         15
#> 152  10       1         15
#> 153   5       1         15
#> 154   6       1         15
#> 155   6       1         15
#> 156   9       1         15
#> 157   6       1         15
#> 158   5       1         15
#> 159   6       1         15
#> 160   6       1         15
#> 161   7       1         16
#> 162   6       1         16
#> 163   5       1         16
#> 164   6       1         16
#> 165   6       1         16
#> 166   6       1         16
#> 167   6       1         16
#> 168   7       1         16
#> 169   5       1         16
#> 170   5       1         16
#> 171   9       1         17
#> 172   9       1         17
#> 173   9       1         17
#> 174   4       1         17
#> 175   7       1         17
#> 176   7       1         17
#> 177   4       1         17
#> 178   8       1         17
#> 179   4       1         17
#> 180   9       1         17
#> 181   5       1         18
#> 182   9       1         18
#> 183   7       1         18
#> 184   3       1         18
#> 185   7       1         18
#> 186   7       1         18
#> 187   6       1         18
#> 188   8       1         18
#> 189   5       1         18
#> 190   7       1         18
#> 191  10       1         19
#> 192   9       1         19
#> 193   8       1         19
#> 194  13       1         19
#> 195   9       1         19
#> 196  11       1         19
#> 197   7       1         19
#> 198  10       1         19
#> 199   8       1         19
#> 200   6       1         19
#> 201   8       1         20
#> 202   5       1         20
#> 203  10       1         20
#> 204   4       1         20
#> 205  10       1         20
#> 206  13       1         20
#> 207   8       1         20
#> 208   4       1         20
#> 209   4       1         20
#> 210   8       1         20
#> 211  13       1         21
#> 212   9       1         21
#> 213   8       1         21
#> 214   5       1         21
#> 215   5       1         21
#> 216  11       1         21
#> 217   8       1         21
#> 218  11       1         21
#> 219   6       1         21
#> 220   7       1         21
#> 221   6       1         22
#> 222   7       1         22
#> 223   7       1         22
#> 224   7       1         22
#> 225   7       1         22
#> 226   6       1         22
#> 227   8       1         22
#> 228  10       1         22
#> 229  11       1         22
#> 230  11       1         22
#> 231  11       1         23
#> 232   9       1         23
#> 233   8       1         23
#> 234   8       1         23
#> 235   5       1         23
#> 236  10       1         23
#> 237  13       1         23
#> 238   4       1         23
#> 239   9       1         23
#> 240   8       1         23
#> 241  10       1         24
#> 242   8       1         24
#> 243  11       1         24
#> 244  13       1         24
#> 245  10       1         24
#> 246   3       1         24
#> 247  13       1         24
#> 248   8       1         24
#> 249  12       1         24
#> 250  11       1         24
#> 251  11       1         25
#> 252   7       1         25
#> 253   7       1         25
#> 254  10       1         25
#> 255   8       1         25
#> 256  14       1         25
#> 257   8       1         25
#> 258  11       1         25
#> 259   6       1         25
#> 260   9       1         25
#> 261   9       1         26
#> 262   6       1         26
#> 263  13       1         26
#> 264   8       1         26
#> 265  14       1         26
#> 266   6       1         26
#> 267  11       1         26
#> 268  12       1         26
#> 269  15       1         26
#> 270   8       1         26
#> 271  10       1         27
#> 272  11       1         27
#> 273  12       1         27
#> 274  14       1         27
#> 275  11       1         27
#> 276  11       1         27
#> 277   9       1         27
#> 278  10       1         27
#> 279  12       1         27
#> 280   7       1         27
#> 281   7       1         28
#> 282  12       1         28
#> 283  10       1         28
#> 284  14       1         28
#> 285  12       1         28
#> 286  13       1         28
#> 287  11       1         28
#> 288  11       1         28
#> 289  14       1         28
#> 290  11       1         28
#> 291  13       1         29
#> 292  19       1         29
#> 293  12       1         29
#> 294   8       1         29
#> 295  12       1         29
#> 296  10       1         29
#> 297   9       1         29
#> 298  10       1         29
#> 299   6       1         29
#> 300   7       1         29
#> 301  12       1         30
#> 302   8       1         30
#> 303  11       1         30
#> 304  14       1         30
#> 305  13       1         30
#> 306   8       1         30
#> 307   8       1         30
#> 308  18       1         30
#> 309  13       1         30
#> 310  17       1         30
#> 311  12       1         31
#> 312  11       1         31
#> 313   8       1         31
#> 314  14       1         31
#> 315  12       1         31
#> 316   7       1         31
#> 317   8       1         31
#> 318   8       1         31
#> 319  16       1         31
#> 320  16       1         31
#> 321  14       1         32
#> 322  12       1         32
#> 323  12       1         32
#> 324  12       1         32
#> 325  19       1         32
#> 326   6       1         32
#> 327  16       1         32
#> 328  15       1         32
#> 329  18       1         32
#> 330  13       1         32
#> 331  15       1         33
#> 332  12       1         33
#> 333  11       1         33
#> 334  16       1         33
#> 335  13       1         33
#> 336  14       1         33
#> 337  12       1         33
#> 338  11       1         33
#> 339  14       1         33
#> 340  15       1         33
#> 341  12       1         34
#> 342  12       1         34
#> 343  12       1         34
#> 344  13       1         34
#> 345  17       1         34
#> 346  14       1         34
#> 347  17       1         34
#> 348  16       1         34
#> 349  19       1         34
#> 350  15       1         34
#> 351  16       1         35
#> 352  17       1         35
#> 353  16       1         35
#> 354  13       1         35
#> 355  16       1         35
#> 356  15       1         35
#> 357  15       1         35
#> 358  15       1         35
#> 359  16       1         35
#> 360  15       1         35
#> 361  12       1         36
#> 362  18       1         36
#> 363   8       1         36
#> 364  15       1         36
#> 365  16       1         36
#> 366  14       1         36
#> 367  10       1         36
#> 368  17       1         36
#> 369  21       1         36
#> 370  22       1         36
#> 371  18       1         37
#> 372  18       1         37
#> 373  20       1         37
#> 374  12       1         37
#> 375  12       1         37
#> 376  16       1         37
#> 377  14       1         37
#> 378  17       1         37
#> 379   9       1         37
#> 380  22       1         37
#> 381  14       1         38
#> 382  14       1         38
#> 383  14       1         38
#> 384  16       1         38
#> 385  12       1         38
#> 386   8       1         38
#> 387  16       1         38
#> 388  10       1         38
#> 389  15       1         38
#> 390  19       1         38
#> 391  10       1         39
#> 392  16       1         39
#> 393  15       1         39
#> 394  12       1         39
#> 395  20       1         39
#> 396  16       1         39
#> 397  16       1         39
#> 398  14       1         39
#> 399  17       1         39
#> 400  10       1         39
#> 401  13       1         40
#> 402  10       1         40
#> 403  16       1         40
#> 404  18       1         40
#> 405  13       1         40
#> 406  20       1         40
#> 407  14       1         40
#> 408  15       1         40
#> 409  22       1         40
#> 410  13       1         40
#> 411  12       1         41
#> 412  16       1         41
#> 413  19       1         41
#> 414  11       1         41
#> 415  19       1         41
#> 416  12       1         41
#> 417   9       1         41
#> 418  17       1         41
#> 419  20       1         41
#> 420  14       1         41
#> 421  13       1         42
#> 422  13       1         42
#> 423  12       1         42
#> 424  24       1         42
#> 425  18       1         42
#> 426  18       1         42
#> 427  19       1         42
#> 428  17       1         42
#> 429  12       1         42
#> 430  15       1         42
#> 431  16       1         43
#> 432  17       1         43
#> 433  16       1         43
#> 434  21       1         43
#> 435  14       1         43
#> 436  18       1         43
#> 437  18       1         43
#> 438  23       1         43
#> 439  20       1         43
#> 440  19       1         43
#> 441  15       1         44
#> 442  19       1         44
#> 443  20       1         44
#> 444  22       1         44
#> 445  16       1         44
#> 446  15       1         44
#> 447  14       1         44
#> 448  18       1         44
#> 449  20       1         44
#> 450  16       1         44
#> 451  18       1         45
#> 452  17       1         45
#> 453  18       1         45
#> 454  19       1         45
#> 455  18       1         45
#> 456  16       1         45
#> 457  17       1         45
#> 458  17       1         45
#> 459  21       1         45
#> 460  18       1         45
#> 461  12       1         46
#> 462  17       1         46
#> 463  18       1         46
#> 464  23       1         46
#> 465  19       1         46
#> 466  17       1         46
#> 467  11       1         46
#> 468  14       1         46
#> 469  16       1         46
#> 470  20       1         46
#> 471  18       1         47
#> 472  14       1         47
#> 473  19       1         47
#> 474  24       1         47
#> 475  14       1         47
#> 476  19       1         47
#> 477  14       1         47
#> 478  22       1         47
#> 479  19       1         47
#> 480  17       1         47
#> 481  17       1         48
#> 482  21       1         48
#> 483  19       1         48
#> 484  17       1         48
#> 485  18       1         48
#> 486  19       1         48
#> 487  17       1         48
#> 488  20       1         48
#> 489  19       1         48
#> 490  22       1         48
#> 491  19       1         49
#> 492  21       1         49
#> 493  20       1         49
#> 494  17       1         49
#> 495  18       1         49
#> 496  15       1         49
#> 497  18       1         49
#> 498  16       1         49
#> 499  17       1         49
#> 500  22       1         49
#> 501  20       1         50
#> 502  20       1         50
#> 503  20       1         50
#> 504  18       1         50
#> 505  17       1         50
#> 506  20       1         50
#> 507  20       1         50
#> 508  21       1         50
#> 509  24       1         50
#> 510  13       1         50
#> 511  22       1         51
#> 512  17       1         51
#> 513  15       1         51
#> 514  17       1         51
#> 515  21       1         51
#> 516  15       1         51
#> 517  27       2         51
#> 518  23       1         51
#> 519  23       1         51
#> 520  22       1         51
#> 521  21       1         52
#> 522  20       1         52
#> 523  22       1         52
#> 524  23       1         52
#> 525  18       1         52
#> 526  21       1         52
#> 527  16       1         52
#> 528  20       1         52
#> 529  24       1         52
#> 530  24       1         52
#> 531  21       1         53
#> 532  22       1         53
#> 533  17       1         53
#> 534  21       1         53
#> 535  15       1         53
#> 536  20       1         53
#> 537  24       1         53
#> 538  27       2         53
#> 539  28       2         53
#> 540  21       1         53
#> 541  18       1         54
#> 542  22       1         54
#> 543  17       1         54
#> 544  19       1         54
#> 545  22       1         54
#> 546  20       1         54
#> 547  17       1         54
#> 548  21       1         54
#> 549  24       1         54
#> 550  24       1         54
#> 551  26       2         55
#> 552  16       1         55
#> 553  19       1         55
#> 554  19       1         55
#> 555  23       1         55
#> 556  23       1         55
#> 557  18       1         55
#> 558  19       1         55
#> 559  21       1         55
#> 560  19       1         55
#> 561  23       1         56
#> 562  27       2         56
#> 563  21       1         56
#> 564  23       1         56
#> 565  24       1         56
#> 566  28       2         56
#> 567  21       1         56
#> 568  20       1         56
#> 569  24       1         56
#> 570  26       2         56
#> 571  25       2         57
#> 572  23       1         57
#> 573  21       1         57
#> 574  20       1         57
#> 575  23       1         57
#> 576  20       1         57
#> 577  23       1         57
#> 578  24       1         57
#> 579  22       1         57
#> 580  25       2         57
#> 581  22       1         58
#> 582  28       2         58
#> 583  25       2         58
#> 584  21       1         58
#> 585  25       2         58
#> 586  24       1         58
#> 587  25       2         58
#> 588  22       1         58
#> 589  22       1         58
#> 590  24       1         58
#> 591  25       2         59
#> 592  25       2         59
#> 593  19       1         59
#> 594  18       1         59
#> 595  21       1         59
#> 596  26       2         59
#> 597  20       1         59
#> 598  23       1         59
#> 599  22       1         59
#> 600  26       2         59
#> 601  26       2         60
#> 602  20       1         60
#> 603  26       2         60
#> 604  23       1         60
#> 605  21       1         60
#> 606  24       1         60
#> 607  25       2         60
#> 608  26       2         60
#> 609  22       1         60
#> 610  20       1         60
#> 611  30       2         61
#> 612  26       2         61
#> 613  20       1         61
#> 614  27       2         61
#> 615  24       1         61
#> 616  23       1         61
#> 617  16       1         61
#> 618  22       1         61
#> 619  25       2         61
#> 620  21       1         61
#> 621  27       2         62
#> 622  23       1         62
#> 623  24       1         62
#> 624  21       1         62
#> 625  24       1         62
#> 626  29       2         62
#> 627  27       2         62
#> 628  30       2         62
#> 629  26       2         62
#> 630  21       1         62
#> 631  21       1         63
#> 632  32       2         63
#> 633  22       1         63
#> 634  28       2         63
#> 635  20       1         63
#> 636  27       2         63
#> 637  27       2         63
#> 638  28       2         63
#> 639  25       2         63
#> 640  26       2         63
#> 641  28       2         64
#> 642  22       1         64
#> 643  20       1         64
#> 644  29       2         64
#> 645  25       2         64
#> 646  28       2         64
#> 647  24       1         64
#> 648  18       1         64
#> 649  28       2         64
#> 650  17       1         64
#> 651  30       2         65
#> 652  30       2         65
#> 653  25       2         65
#> 654  22       1         65
#> 655  28       2         65
#> 656  27       2         65
#> 657  26       2         65
#> 658  26       2         65
#> 659  26       2         65
#> 660  25       2         65
#> 661  25       2         66
#> 662  29       2         66
#> 663  29       2         66
#> 664  27       2         66
#> 665  26       2         66
#> 666  23       1         66
#> 667  29       2         66
#> 668  26       2         66
#> 669  30       2         66
#> 670  27       2         66
#> 671  24       1         67
#> 672  27       2         67
#> 673  25       2         67
#> 674  27       2         67
#> 675  28       2         67
#> 676  30       2         67
#> 677  28       2         67
#> 678  23       1         67
#> 679  22       1         67
#> 680  22       1         67
#> 681  27       2         68
#> 682  27       2         68
#> 683  28       2         68
#> 684  25       2         68
#> 685  27       2         68
#> 686  24       1         68
#> 687  26       2         68
#> 688  25       2         68
#> 689  31       2         68
#> 690  27       2         68
#> 691  25       2         69
#> 692  29       2         69
#> 693  35       2         69
#> 694  29       2         69
#> 695  29       2         69
#> 696  20       1         69
#> 697  31       2         69
#> 698  25       2         69
#> 699  25       2         69
#> 700  26       2         69
#> 701  29       2         70
#> 702  29       2         70
#> 703  26       2         70
#> 704  24       1         70
#> 705  26       2         70
#> 706  27       2         70
#> 707  31       2         70
#> 708  25       2         70
#> 709  31       2         70
#> 710  26       2         70
#> 711  31       2         71
#> 712  34       2         71
#> 713  24       1         71
#> 714  28       2         71
#> 715  26       2         71
#> 716  28       2         71
#> 717  26       2         71
#> 718  28       2         71
#> 719  29       2         71
#> 720  32       2         71
#> 721  30       2         72
#> 722  31       2         72
#> 723  31       2         72
#> 724  29       2         72
#> 725  22       1         72
#> 726  27       2         72
#> 727  29       2         72
#> 728  25       2         72
#> 729  27       2         72
#> 730  26       2         72
#> 731  26       2         73
#> 732  26       2         73
#> 733  30       2         73
#> 734  31       2         73
#> 735  32       2         73
#> 736  30       2         73
#> 737  34       2         73
#> 738  26       2         73
#> 739  32       2         73
#> 740  29       2         73
#> 741  28       2         74
#> 742  29       2         74
#> 743  27       2         74
#> 744  26       2         74
#> 745  31       2         74
#> 746  34       2         74
#> 747  27       2         74
#> 748  24       1         74
#> 749  30       2         74
#> 750  32       2         74
#> 751  31       2         75
#> 752  25       2         75
#> 753  24       1         75
#> 754  31       2         75
#> 755  25       2         75
#> 756  30       2         75
#> 757  32       2         75
#> 758  33       2         75
#> 759  26       2         75
#> 760  29       2         75
#> 761  32       2         76
#> 762  34       2         76
#> 763  33       2         76
#> 764  32       2         76
#> 765  31       2         76
#> 766  32       2         76
#> 767  33       2         76
#> 768  31       2         76
#> 769  32       2         76
#> 770  35       2         76
#> 771  32       2         77
#> 772  32       2         77
#> 773  33       2         77
#> 774  33       2         77
#> 775  29       2         77
#> 776  33       2         77
#> 777  29       2         77
#> 778  27       2         77
#> 779  30       2         77
#> 780  29       2         77
#> 781  31       2         78
#> 782  36       2         78
#> 783  34       2         78
#> 784  31       2         78
#> 785  30       2         78
#> 786  30       2         78
#> 787  29       2         78
#> 788  33       2         78
#> 789  29       2         78
#> 790  33       2         78
#> 791  33       2         79
#> 792  30       2         79
#> 793  29       2         79
#> 794  31       2         79
#> 795  31       2         79
#> 796  29       2         79
#> 797  25       2         79
#> 798  30       2         79
#> 799  32       2         79
#> 800  36       2         79
#> 801  34       2         80
#> 802  35       2         80
#> 803  27       2         80
#> 804  32       2         80
#> 805  32       2         80
#> 806  34       2         80
#> 807  33       2         80
#> 808  32       2         80
#> 809  36       2         80
#> 810  31       2         80
#> 811  33       2         81
#> 812  29       2         81
#> 813  31       2         81
#> 814  35       2         81
#> 815  33       2         81
#> 816  35       2         81
#> 817  36       2         81
#> 818  32       2         81
#> 819  36       2         81
#> 820  30       2         81
#> 821  34       2         82
#> 822  33       2         82
#> 823  34       2         82
#> 824  34       2         82
#> 825  38       3         82
#> 826  31       2         82
#> 827  35       2         82
#> 828  31       2         82
#> 829  34       2         82
#> 830  32       2         82
#> 831  32       2         83
#> 832  32       2         83
#> 833  35       2         83
#> 834  33       2         83
#> 835  34       2         83
#> 836  30       2         83
#> 837  27       2         83
#> 838  30       2         83
#> 839  30       2         83
#> 840  33       2         83
#> 841  36       2         84
#> 842  34       2         84
#> 843  34       2         84
#> 844  31       2         84
#> 845  34       2         84
#> 846  33       2         84
#> 847  35       2         84
#> 848  35       2         84
#> 849  31       2         84
#> 850  33       2         84
#> 851  33       2         85
#> 852  35       2         85
#> 853  33       2         85
#> 854  37       3         85
#> 855  33       2         85
#> 856  34       2         85
#> 857  33       2         85
#> 858  35       2         85
#> 859  35       2         85
#> 860  36       2         85
#> 861  32       2         86
#> 862  35       2         86
#> 863  29       2         86
#> 864  35       2         86
#> 865  33       2         86
#> 866  31       2         86
#> 867  38       3         86
#> 868  37       3         86
#> 869  30       2         86
#> 870  34       2         86
#> 871  35       2         87
#> 872  38       3         87
#> 873  35       2         87
#> 874  37       3         87
#> 875  28       2         87
#> 876  36       2         87
#> 877  34       2         87
#> 878  34       2         87
#> 879  34       2         87
#> 880  36       2         87
#> 881  34       2         88
#> 882  34       2         88
#> 883  31       2         88
#> 884  35       2         88
#> 885  34       2         88
#> 886  34       2         88
#> 887  34       2         88
#> 888  38       3         88
#> 889  33       2         88
#> 890  36       2         88
#> 891  37       3         89
#> 892  35       2         89
#> 893  35       2         89
#> 894  35       2         89
#> 895  37       3         89
#> 896  37       3         89
#> 897  37       3         89
#> 898  37       3         89
#> 899  36       2         89
#> 900  35       2         89
#> 901  37       3         90
#> 902  31       2         90
#> 903  38       3         90
#> 904  36       2         90
#> 905  35       2         90
#> 906  36       2         90
#> 907  36       2         90
#> 908  38       3         90
#> 909  39       3         90
#> 910  36       2         90
#> 911  38       3         91
#> 912  36       2         91
#> 913  37       3         91
#> 914  37       3         91
#> 915  36       2         91
#> 916  35       2         91
#> 917  36       2         91
#> 918  35       2         91
#> 919  40       3         91
#> 920  35       2         91
#> 921  35       2         92
#> 922  39       3         92
#> 923  35       2         92
#> 924  39       3         92
#> 925  35       2         92
#> 926  39       3         92
#> 927  38       3         92
#> 928  35       2         92
#> 929  36       2         92
#> 930  37       3         92
#> 931  40       3         93
#> 932  38       3         93
#> 933  35       2         93
#> 934  38       3         93
#> 935  40       3         93
#> 936  36       2         93
#> 937  36       2         93
#> 938  37       3         93
#> 939  32       2         93
#> 940  37       3         93
#> 941  36       2         94
#> 942  39       3         94
#> 943  36       2         94
#> 944  38       3         94
#> 945  38       3         94
#> 946  37       3         94
#> 947  37       3         94
#> 948  40       3         94
#> 949  38       3         94
#> 950  38       3         94
#> 951  39       3         95
#> 952  40       3         95
#> 953  40       3         95
#> 954  36       2         95
#> 955  38       3         95
#> 956  37       3         95
#> 957  38       3         95
#> 958  38       3         95
#> 959  38       3         95
#> 960  37       3         95
#> 961  38       3         96
#> 962  39       3         96
#> 963  36       2         96
#> 964  39       3         96
#> 965  39       3         96
#> 966  38       3         96
#> 967  40       3         96
#> 968  40       3         96
#> 969  38       3         96
#> 970  39       3         96
#> 971  37       3         97
#> 972  38       3         97
#> 973  40       3         97
#> 974  40       3         97
#> 975  38       3         97
#> 976  39       3         97
#> 977  39       3         97
#> 978  39       3         97
#> 979  38       3         97
#> 980  37       3         97
#> 981  39       3         98
#> 982  40       3         98
#> 983  40       3         98
#> 984  40       3         98
#> 985  37       3         98
#> 986  38       3         98
#> 987  38       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  39       3         98
#> 991  40       3         99
#> 992  40       3         99
#> 993  39       3         99
#> 994  40       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  40       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 39       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100