Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simulate_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simulate_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    2       1          1
#> 12    0       1          1
#> 13    0       1          1
#> 14    2       1          1
#> 15    1       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    0       1          2
#> 22    2       1          2
#> 23    0       1          2
#> 24    1       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    1       1          2
#> 28    1       1          2
#> 29    0       1          2
#> 30    2       1          2
#> 31    1       1          3
#> 32    1       1          3
#> 33    1       1          3
#> 34    1       1          3
#> 35    2       1          3
#> 36    2       1          3
#> 37    1       1          3
#> 38    1       1          3
#> 39    3       1          3
#> 40    0       1          3
#> 41    2       1          4
#> 42    3       1          4
#> 43    3       1          4
#> 44    1       1          4
#> 45    2       1          4
#> 46    4       1          4
#> 47    2       1          4
#> 48    2       1          4
#> 49    0       1          4
#> 50    2       1          4
#> 51    1       1          5
#> 52    5       1          5
#> 53    2       1          5
#> 54    4       1          5
#> 55    2       1          5
#> 56    2       1          5
#> 57    3       1          5
#> 58    3       1          5
#> 59    0       1          5
#> 60    4       1          5
#> 61    2       1          6
#> 62    1       1          6
#> 63    3       1          6
#> 64    2       1          6
#> 65    0       1          6
#> 66    4       1          6
#> 67    2       1          6
#> 68    4       1          6
#> 69    1       1          6
#> 70    3       1          6
#> 71    3       1          7
#> 72    1       1          7
#> 73    3       1          7
#> 74    4       1          7
#> 75    1       1          7
#> 76    1       1          7
#> 77    3       1          7
#> 78    4       1          7
#> 79    2       1          7
#> 80    2       1          7
#> 81    3       1          8
#> 82    1       1          8
#> 83    5       1          8
#> 84    3       1          8
#> 85    3       1          8
#> 86    6       1          8
#> 87    7       1          8
#> 88    5       1          8
#> 89    3       1          8
#> 90    1       1          8
#> 91    2       1          9
#> 92    2       1          9
#> 93    2       1          9
#> 94    2       1          9
#> 95    3       1          9
#> 96    4       1          9
#> 97    4       1          9
#> 98    6       1          9
#> 99    6       1          9
#> 100   5       1          9
#> 101   4       1         10
#> 102   5       1         10
#> 103   6       1         10
#> 104   2       1         10
#> 105   4       1         10
#> 106   3       1         10
#> 107   5       1         10
#> 108   3       1         10
#> 109   3       1         10
#> 110   3       1         10
#> 111   4       1         11
#> 112   7       1         11
#> 113   5       1         11
#> 114   3       1         11
#> 115   8       1         11
#> 116   2       1         11
#> 117   6       1         11
#> 118   5       1         11
#> 119   5       1         11
#> 120   4       1         11
#> 121   6       1         12
#> 122   3       1         12
#> 123   8       1         12
#> 124   4       1         12
#> 125   6       1         12
#> 126   3       1         12
#> 127   3       1         12
#> 128   6       1         12
#> 129   6       1         12
#> 130   7       1         12
#> 131   8       1         13
#> 132  10       1         13
#> 133   5       1         13
#> 134   5       1         13
#> 135   4       1         13
#> 136   5       1         13
#> 137   3       1         13
#> 138   4       1         13
#> 139   7       1         13
#> 140   9       1         13
#> 141  11       1         14
#> 142   3       1         14
#> 143   7       1         14
#> 144   5       1         14
#> 145   7       1         14
#> 146   5       1         14
#> 147   7       1         14
#> 148   7       1         14
#> 149  10       1         14
#> 150   8       1         14
#> 151   7       1         15
#> 152   4       1         15
#> 153   6       1         15
#> 154   9       1         15
#> 155   6       1         15
#> 156   3       1         15
#> 157   7       1         15
#> 158   4       1         15
#> 159   4       1         15
#> 160  10       1         15
#> 161   7       1         16
#> 162   5       1         16
#> 163   6       1         16
#> 164   5       1         16
#> 165   8       1         16
#> 166   8       1         16
#> 167   9       1         16
#> 168   5       1         16
#> 169   6       1         16
#> 170   6       1         16
#> 171   5       1         17
#> 172   8       1         17
#> 173   8       1         17
#> 174   7       1         17
#> 175   4       1         17
#> 176   9       1         17
#> 177   6       1         17
#> 178   9       1         17
#> 179   6       1         17
#> 180   8       1         17
#> 181  10       1         18
#> 182  10       1         18
#> 183   9       1         18
#> 184   6       1         18
#> 185   7       1         18
#> 186   5       1         18
#> 187   6       1         18
#> 188   5       1         18
#> 189   8       1         18
#> 190   6       1         18
#> 191   6       1         19
#> 192   9       1         19
#> 193   9       1         19
#> 194   7       1         19
#> 195   9       1         19
#> 196   8       1         19
#> 197   8       1         19
#> 198   8       1         19
#> 199  13       1         19
#> 200   5       1         19
#> 201  11       1         20
#> 202  11       1         20
#> 203  10       1         20
#> 204   8       1         20
#> 205   7       1         20
#> 206   8       1         20
#> 207  12       1         20
#> 208   5       1         20
#> 209  10       1         20
#> 210  14       1         20
#> 211   9       1         21
#> 212   8       1         21
#> 213   6       1         21
#> 214   7       1         21
#> 215  10       1         21
#> 216   9       1         21
#> 217   9       1         21
#> 218   9       1         21
#> 219   6       1         21
#> 220   6       1         21
#> 221   6       1         22
#> 222  12       1         22
#> 223   5       1         22
#> 224  10       1         22
#> 225  10       1         22
#> 226   6       1         22
#> 227   7       1         22
#> 228   5       1         22
#> 229   7       1         22
#> 230   5       1         22
#> 231  12       1         23
#> 232  10       1         23
#> 233   8       1         23
#> 234  14       1         23
#> 235   5       1         23
#> 236   7       1         23
#> 237  11       1         23
#> 238  13       1         23
#> 239  13       1         23
#> 240  12       1         23
#> 241   7       1         24
#> 242   8       1         24
#> 243   6       1         24
#> 244  13       1         24
#> 245   6       1         24
#> 246   9       1         24
#> 247  12       1         24
#> 248  10       1         24
#> 249   9       1         24
#> 250   8       1         24
#> 251   7       1         25
#> 252   7       1         25
#> 253  11       1         25
#> 254  11       1         25
#> 255   9       1         25
#> 256  11       1         25
#> 257  11       1         25
#> 258   7       1         25
#> 259   9       1         25
#> 260   9       1         25
#> 261   6       1         26
#> 262   7       1         26
#> 263  12       1         26
#> 264  10       1         26
#> 265  11       1         26
#> 266  14       1         26
#> 267  13       1         26
#> 268  11       1         26
#> 269  13       1         26
#> 270  10       1         26
#> 271  10       1         27
#> 272   7       1         27
#> 273   9       1         27
#> 274   8       1         27
#> 275   8       1         27
#> 276  14       1         27
#> 277  15       1         27
#> 278  11       1         27
#> 279  12       1         27
#> 280   6       1         27
#> 281  12       1         28
#> 282   9       1         28
#> 283   6       1         28
#> 284  10       1         28
#> 285  10       1         28
#> 286  13       1         28
#> 287   8       1         28
#> 288  13       1         28
#> 289  10       1         28
#> 290  13       1         28
#> 291  14       1         29
#> 292  11       1         29
#> 293   5       1         29
#> 294  13       1         29
#> 295  10       1         29
#> 296  11       1         29
#> 297  12       1         29
#> 298  14       1         29
#> 299  18       1         29
#> 300  13       1         29
#> 301  15       1         30
#> 302  14       1         30
#> 303  13       1         30
#> 304   8       1         30
#> 305   9       1         30
#> 306  12       1         30
#> 307  18       1         30
#> 308  12       1         30
#> 309  13       1         30
#> 310  12       1         30
#> 311  11       1         31
#> 312  13       1         31
#> 313   8       1         31
#> 314  10       1         31
#> 315  12       1         31
#> 316  21       1         31
#> 317  14       1         31
#> 318  13       1         31
#> 319  11       1         31
#> 320  17       1         31
#> 321  11       1         32
#> 322  11       1         32
#> 323  13       1         32
#> 324  10       1         32
#> 325  12       1         32
#> 326  16       1         32
#> 327  23       1         32
#> 328  12       1         32
#> 329  12       1         32
#> 330  13       1         32
#> 331  13       1         33
#> 332  14       1         33
#> 333  13       1         33
#> 334  13       1         33
#> 335  15       1         33
#> 336  13       1         33
#> 337  13       1         33
#> 338  14       1         33
#> 339  19       1         33
#> 340  10       1         33
#> 341  13       1         34
#> 342  16       1         34
#> 343  13       1         34
#> 344  17       1         34
#> 345  11       1         34
#> 346  16       1         34
#> 347  15       1         34
#> 348  14       1         34
#> 349   8       1         34
#> 350  10       1         34
#> 351  13       1         35
#> 352  13       1         35
#> 353  14       1         35
#> 354  16       1         35
#> 355  14       1         35
#> 356  16       1         35
#> 357  16       1         35
#> 358  19       1         35
#> 359   9       1         35
#> 360  18       1         35
#> 361  14       1         36
#> 362  17       1         36
#> 363  11       1         36
#> 364  13       1         36
#> 365  19       1         36
#> 366  13       1         36
#> 367  19       1         36
#> 368  13       1         36
#> 369  15       1         36
#> 370  14       1         36
#> 371  11       1         37
#> 372  19       1         37
#> 373  16       1         37
#> 374  13       1         37
#> 375  16       1         37
#> 376  14       1         37
#> 377  10       1         37
#> 378  22       1         37
#> 379  13       1         37
#> 380  15       1         37
#> 381  17       1         38
#> 382  15       1         38
#> 383  11       1         38
#> 384  23       1         38
#> 385  17       1         38
#> 386  15       1         38
#> 387  16       1         38
#> 388  12       1         38
#> 389  17       1         38
#> 390  11       1         38
#> 391  14       1         39
#> 392  19       1         39
#> 393   9       1         39
#> 394  15       1         39
#> 395  10       1         39
#> 396  17       1         39
#> 397  14       1         39
#> 398  18       1         39
#> 399  20       1         39
#> 400  15       1         39
#> 401  16       1         40
#> 402  14       1         40
#> 403  18       1         40
#> 404  14       1         40
#> 405  14       1         40
#> 406  16       1         40
#> 407  12       1         40
#> 408  17       1         40
#> 409  15       1         40
#> 410  19       1         40
#> 411  11       1         41
#> 412  15       1         41
#> 413  16       1         41
#> 414  20       1         41
#> 415  19       1         41
#> 416  19       1         41
#> 417  13       1         41
#> 418  14       1         41
#> 419  15       1         41
#> 420  21       1         41
#> 421  19       1         42
#> 422  13       1         42
#> 423  18       1         42
#> 424  14       1         42
#> 425  16       1         42
#> 426  19       1         42
#> 427  17       1         42
#> 428  13       1         42
#> 429  16       1         42
#> 430  18       1         42
#> 431  15       1         43
#> 432  18       1         43
#> 433  13       1         43
#> 434  13       1         43
#> 435  23       1         43
#> 436  18       1         43
#> 437  17       1         43
#> 438  19       1         43
#> 439  17       1         43
#> 440  12       1         43
#> 441  10       1         44
#> 442  19       1         44
#> 443  25       2         44
#> 444  17       1         44
#> 445  15       1         44
#> 446  16       1         44
#> 447  19       1         44
#> 448  13       1         44
#> 449  20       1         44
#> 450  17       1         44
#> 451  19       1         45
#> 452  16       1         45
#> 453  17       1         45
#> 454  17       1         45
#> 455  20       1         45
#> 456  11       1         45
#> 457  21       1         45
#> 458  18       1         45
#> 459  20       1         45
#> 460  14       1         45
#> 461  17       1         46
#> 462  16       1         46
#> 463  25       2         46
#> 464  17       1         46
#> 465  11       1         46
#> 466  23       1         46
#> 467  28       2         46
#> 468  24       1         46
#> 469  21       1         46
#> 470  17       1         46
#> 471  17       1         47
#> 472  13       1         47
#> 473  20       1         47
#> 474  18       1         47
#> 475  15       1         47
#> 476  16       1         47
#> 477  21       1         47
#> 478  18       1         47
#> 479  19       1         47
#> 480  21       1         47
#> 481  17       1         48
#> 482  20       1         48
#> 483  19       1         48
#> 484  19       1         48
#> 485  14       1         48
#> 486  25       2         48
#> 487  21       1         48
#> 488  18       1         48
#> 489  17       1         48
#> 490  17       1         48
#> 491  13       1         49
#> 492  13       1         49
#> 493  19       1         49
#> 494  23       1         49
#> 495  20       1         49
#> 496  15       1         49
#> 497  18       1         49
#> 498  20       1         49
#> 499  20       1         49
#> 500  20       1         49
#> 501  19       1         50
#> 502  19       1         50
#> 503  18       1         50
#> 504  21       1         50
#> 505  16       1         50
#> 506  19       1         50
#> 507  20       1         50
#> 508  25       2         50
#> 509  18       1         50
#> 510  23       1         50
#> 511  17       1         51
#> 512  19       1         51
#> 513  21       1         51
#> 514  14       1         51
#> 515  20       1         51
#> 516  22       1         51
#> 517  19       1         51
#> 518  19       1         51
#> 519  25       2         51
#> 520  25       2         51
#> 521  20       1         52
#> 522  21       1         52
#> 523  19       1         52
#> 524  23       1         52
#> 525  17       1         52
#> 526  24       1         52
#> 527  20       1         52
#> 528  22       1         52
#> 529  17       1         52
#> 530  26       2         52
#> 531  24       1         53
#> 532  22       1         53
#> 533  17       1         53
#> 534  20       1         53
#> 535  22       1         53
#> 536  23       1         53
#> 537  17       1         53
#> 538  23       1         53
#> 539  23       1         53
#> 540  21       1         53
#> 541  25       2         54
#> 542  28       2         54
#> 543  26       2         54
#> 544  25       2         54
#> 545  21       1         54
#> 546  22       1         54
#> 547  24       1         54
#> 548  26       2         54
#> 549  19       1         54
#> 550  20       1         54
#> 551  30       2         55
#> 552  23       1         55
#> 553  20       1         55
#> 554  22       1         55
#> 555  26       2         55
#> 556  21       1         55
#> 557  22       1         55
#> 558  18       1         55
#> 559  25       2         55
#> 560  24       1         55
#> 561  22       1         56
#> 562  25       2         56
#> 563  21       1         56
#> 564  17       1         56
#> 565  15       1         56
#> 566  27       2         56
#> 567  22       1         56
#> 568  23       1         56
#> 569  19       1         56
#> 570  22       1         56
#> 571  19       1         57
#> 572  19       1         57
#> 573  28       2         57
#> 574  22       1         57
#> 575  24       1         57
#> 576  28       2         57
#> 577  20       1         57
#> 578  18       1         57
#> 579  22       1         57
#> 580  23       1         57
#> 581  25       2         58
#> 582  19       1         58
#> 583  24       1         58
#> 584  19       1         58
#> 585  19       1         58
#> 586  21       1         58
#> 587  26       2         58
#> 588  21       1         58
#> 589  24       1         58
#> 590  21       1         58
#> 591  22       1         59
#> 592  25       2         59
#> 593  29       2         59
#> 594  21       1         59
#> 595  25       2         59
#> 596  23       1         59
#> 597  19       1         59
#> 598  25       2         59
#> 599  25       2         59
#> 600  25       2         59
#> 601  21       1         60
#> 602  23       1         60
#> 603  26       2         60
#> 604  20       1         60
#> 605  21       1         60
#> 606  31       2         60
#> 607  24       1         60
#> 608  23       1         60
#> 609  21       1         60
#> 610  21       1         60
#> 611  27       2         61
#> 612  29       2         61
#> 613  25       2         61
#> 614  22       1         61
#> 615  23       1         61
#> 616  30       2         61
#> 617  24       1         61
#> 618  26       2         61
#> 619  23       1         61
#> 620  24       1         61
#> 621  25       2         62
#> 622  29       2         62
#> 623  27       2         62
#> 624  20       1         62
#> 625  25       2         62
#> 626  29       2         62
#> 627  28       2         62
#> 628  22       1         62
#> 629  22       1         62
#> 630  23       1         62
#> 631  25       2         63
#> 632  25       2         63
#> 633  28       2         63
#> 634  20       1         63
#> 635  27       2         63
#> 636  21       1         63
#> 637  29       2         63
#> 638  26       2         63
#> 639  25       2         63
#> 640  26       2         63
#> 641  26       2         64
#> 642  22       1         64
#> 643  26       2         64
#> 644  26       2         64
#> 645  20       1         64
#> 646  20       1         64
#> 647  25       2         64
#> 648  22       1         64
#> 649  28       2         64
#> 650  25       2         64
#> 651  28       2         65
#> 652  19       1         65
#> 653  26       2         65
#> 654  23       1         65
#> 655  27       2         65
#> 656  31       2         65
#> 657  26       2         65
#> 658  28       2         65
#> 659  27       2         65
#> 660  24       1         65
#> 661  27       2         66
#> 662  28       2         66
#> 663  26       2         66
#> 664  23       1         66
#> 665  26       2         66
#> 666  31       2         66
#> 667  30       2         66
#> 668  24       1         66
#> 669  26       2         66
#> 670  30       2         66
#> 671  29       2         67
#> 672  31       2         67
#> 673  28       2         67
#> 674  32       2         67
#> 675  25       2         67
#> 676  35       2         67
#> 677  27       2         67
#> 678  25       2         67
#> 679  24       1         67
#> 680  27       2         67
#> 681  31       2         68
#> 682  25       2         68
#> 683  23       1         68
#> 684  29       2         68
#> 685  32       2         68
#> 686  26       2         68
#> 687  30       2         68
#> 688  26       2         68
#> 689  26       2         68
#> 690  29       2         68
#> 691  29       2         69
#> 692  29       2         69
#> 693  25       2         69
#> 694  29       2         69
#> 695  32       2         69
#> 696  33       2         69
#> 697  28       2         69
#> 698  29       2         69
#> 699  31       2         69
#> 700  29       2         69
#> 701  27       2         70
#> 702  32       2         70
#> 703  26       2         70
#> 704  28       2         70
#> 705  28       2         70
#> 706  31       2         70
#> 707  31       2         70
#> 708  25       2         70
#> 709  30       2         70
#> 710  32       2         70
#> 711  30       2         71
#> 712  25       2         71
#> 713  27       2         71
#> 714  36       2         71
#> 715  32       2         71
#> 716  30       2         71
#> 717  30       2         71
#> 718  24       1         71
#> 719  27       2         71
#> 720  32       2         71
#> 721  28       2         72
#> 722  28       2         72
#> 723  29       2         72
#> 724  32       2         72
#> 725  29       2         72
#> 726  29       2         72
#> 727  31       2         72
#> 728  27       2         72
#> 729  31       2         72
#> 730  31       2         72
#> 731  30       2         73
#> 732  32       2         73
#> 733  23       1         73
#> 734  32       2         73
#> 735  31       2         73
#> 736  30       2         73
#> 737  32       2         73
#> 738  25       2         73
#> 739  29       2         73
#> 740  32       2         73
#> 741  25       2         74
#> 742  29       2         74
#> 743  28       2         74
#> 744  35       2         74
#> 745  27       2         74
#> 746  28       2         74
#> 747  34       2         74
#> 748  31       2         74
#> 749  28       2         74
#> 750  23       1         74
#> 751  25       2         75
#> 752  34       2         75
#> 753  32       2         75
#> 754  30       2         75
#> 755  34       2         75
#> 756  32       2         75
#> 757  27       2         75
#> 758  27       2         75
#> 759  29       2         75
#> 760  31       2         75
#> 761  29       2         76
#> 762  32       2         76
#> 763  33       2         76
#> 764  28       2         76
#> 765  32       2         76
#> 766  27       2         76
#> 767  31       2         76
#> 768  29       2         76
#> 769  28       2         76
#> 770  30       2         76
#> 771  33       2         77
#> 772  30       2         77
#> 773  30       2         77
#> 774  32       2         77
#> 775  33       2         77
#> 776  32       2         77
#> 777  32       2         77
#> 778  32       2         77
#> 779  31       2         77
#> 780  26       2         77
#> 781  28       2         78
#> 782  30       2         78
#> 783  31       2         78
#> 784  30       2         78
#> 785  30       2         78
#> 786  29       2         78
#> 787  33       2         78
#> 788  33       2         78
#> 789  31       2         78
#> 790  29       2         78
#> 791  29       2         79
#> 792  31       2         79
#> 793  34       2         79
#> 794  36       2         79
#> 795  32       2         79
#> 796  36       2         79
#> 797  29       2         79
#> 798  28       2         79
#> 799  28       2         79
#> 800  28       2         79
#> 801  35       2         80
#> 802  36       2         80
#> 803  33       2         80
#> 804  31       2         80
#> 805  33       2         80
#> 806  30       2         80
#> 807  32       2         80
#> 808  29       2         80
#> 809  33       2         80
#> 810  34       2         80
#> 811  34       2         81
#> 812  33       2         81
#> 813  35       2         81
#> 814  35       2         81
#> 815  29       2         81
#> 816  33       2         81
#> 817  33       2         81
#> 818  32       2         81
#> 819  36       2         81
#> 820  30       2         81
#> 821  32       2         82
#> 822  34       2         82
#> 823  35       2         82
#> 824  35       2         82
#> 825  32       2         82
#> 826  37       3         82
#> 827  35       2         82
#> 828  32       2         82
#> 829  33       2         82
#> 830  33       2         82
#> 831  32       2         83
#> 832  28       2         83
#> 833  35       2         83
#> 834  32       2         83
#> 835  28       2         83
#> 836  34       2         83
#> 837  34       2         83
#> 838  33       2         83
#> 839  29       2         83
#> 840  32       2         83
#> 841  29       2         84
#> 842  34       2         84
#> 843  34       2         84
#> 844  35       2         84
#> 845  32       2         84
#> 846  35       2         84
#> 847  33       2         84
#> 848  33       2         84
#> 849  35       2         84
#> 850  35       2         84
#> 851  35       2         85
#> 852  34       2         85
#> 853  34       2         85
#> 854  34       2         85
#> 855  37       3         85
#> 856  35       2         85
#> 857  36       2         85
#> 858  33       2         85
#> 859  35       2         85
#> 860  35       2         85
#> 861  35       2         86
#> 862  31       2         86
#> 863  36       2         86
#> 864  35       2         86
#> 865  33       2         86
#> 866  30       2         86
#> 867  35       2         86
#> 868  35       2         86
#> 869  37       3         86
#> 870  36       2         86
#> 871  33       2         87
#> 872  34       2         87
#> 873  33       2         87
#> 874  34       2         87
#> 875  36       2         87
#> 876  33       2         87
#> 877  32       2         87
#> 878  35       2         87
#> 879  35       2         87
#> 880  35       2         87
#> 881  34       2         88
#> 882  38       3         88
#> 883  38       3         88
#> 884  36       2         88
#> 885  35       2         88
#> 886  36       2         88
#> 887  36       2         88
#> 888  35       2         88
#> 889  35       2         88
#> 890  32       2         88
#> 891  29       2         89
#> 892  38       3         89
#> 893  36       2         89
#> 894  33       2         89
#> 895  37       3         89
#> 896  38       3         89
#> 897  33       2         89
#> 898  37       3         89
#> 899  37       3         89
#> 900  33       2         89
#> 901  38       3         90
#> 902  37       3         90
#> 903  34       2         90
#> 904  36       2         90
#> 905  35       2         90
#> 906  38       3         90
#> 907  36       2         90
#> 908  38       3         90
#> 909  38       3         90
#> 910  36       2         90
#> 911  34       2         91
#> 912  37       3         91
#> 913  37       3         91
#> 914  38       3         91
#> 915  36       2         91
#> 916  38       3         91
#> 917  36       2         91
#> 918  36       2         91
#> 919  35       2         91
#> 920  36       2         91
#> 921  38       3         92
#> 922  35       2         92
#> 923  37       3         92
#> 924  38       3         92
#> 925  38       3         92
#> 926  35       2         92
#> 927  38       3         92
#> 928  34       2         92
#> 929  36       2         92
#> 930  36       2         92
#> 931  38       3         93
#> 932  37       3         93
#> 933  39       3         93
#> 934  37       3         93
#> 935  40       3         93
#> 936  39       3         93
#> 937  38       3         93
#> 938  39       3         93
#> 939  36       2         93
#> 940  38       3         93
#> 941  36       2         94
#> 942  39       3         94
#> 943  35       2         94
#> 944  37       3         94
#> 945  39       3         94
#> 946  39       3         94
#> 947  37       3         94
#> 948  39       3         94
#> 949  38       3         94
#> 950  38       3         94
#> 951  39       3         95
#> 952  39       3         95
#> 953  38       3         95
#> 954  38       3         95
#> 955  37       3         95
#> 956  38       3         95
#> 957  40       3         95
#> 958  38       3         95
#> 959  39       3         95
#> 960  39       3         95
#> 961  39       3         96
#> 962  40       3         96
#> 963  40       3         96
#> 964  40       3         96
#> 965  37       3         96
#> 966  37       3         96
#> 967  38       3         96
#> 968  38       3         96
#> 969  39       3         96
#> 970  38       3         96
#> 971  37       3         97
#> 972  39       3         97
#> 973  38       3         97
#> 974  38       3         97
#> 975  40       3         97
#> 976  39       3         97
#> 977  39       3         97
#> 978  40       3         97
#> 979  39       3         97
#> 980  40       3         97
#> 981  40       3         98
#> 982  38       3         98
#> 983  39       3         98
#> 984  40       3         98
#> 985  39       3         98
#> 986  40       3         98
#> 987  40       3         98
#> 988  40       3         98
#> 989  40       3         98
#> 990  40       3         98
#> 991  40       3         99
#> 992  40       3         99
#> 993  39       3         99
#> 994  39       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  40       3         99
#> 998  40       3         99
#> 999  39       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100