Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simulate_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simulate_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    1       1          1
#> 12    0       1          1
#> 13    1       1          1
#> 14    1       1          1
#> 15    1       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    2       1          2
#> 22    0       1          2
#> 23    0       1          2
#> 24    0       1          2
#> 25    1       1          2
#> 26    1       1          2
#> 27    0       1          2
#> 28    0       1          2
#> 29    0       1          2
#> 30    0       1          2
#> 31    0       1          3
#> 32    1       1          3
#> 33    1       1          3
#> 34    0       1          3
#> 35    1       1          3
#> 36    0       1          3
#> 37    0       1          3
#> 38    1       1          3
#> 39    0       1          3
#> 40    0       1          3
#> 41    2       1          4
#> 42    0       1          4
#> 43    1       1          4
#> 44    0       1          4
#> 45    1       1          4
#> 46    2       1          4
#> 47    2       1          4
#> 48    1       1          4
#> 49    0       1          4
#> 50    1       1          4
#> 51    1       1          5
#> 52    1       1          5
#> 53    4       1          5
#> 54    3       1          5
#> 55    1       1          5
#> 56    2       1          5
#> 57    5       1          5
#> 58    2       1          5
#> 59    6       1          5
#> 60    1       1          5
#> 61    1       1          6
#> 62    0       1          6
#> 63    3       1          6
#> 64    3       1          6
#> 65    4       1          6
#> 66    4       1          6
#> 67    1       1          6
#> 68    0       1          6
#> 69    4       1          6
#> 70    2       1          6
#> 71    3       1          7
#> 72    0       1          7
#> 73    2       1          7
#> 74    5       1          7
#> 75    3       1          7
#> 76    5       1          7
#> 77    2       1          7
#> 78    1       1          7
#> 79    4       1          7
#> 80    3       1          7
#> 81    2       1          8
#> 82    3       1          8
#> 83    4       1          8
#> 84    0       1          8
#> 85    3       1          8
#> 86    4       1          8
#> 87    5       1          8
#> 88    5       1          8
#> 89    1       1          8
#> 90    4       1          8
#> 91    4       1          9
#> 92    3       1          9
#> 93    7       1          9
#> 94    5       1          9
#> 95    2       1          9
#> 96    3       1          9
#> 97    5       1          9
#> 98    3       1          9
#> 99    4       1          9
#> 100   2       1          9
#> 101   5       1         10
#> 102   5       1         10
#> 103   6       1         10
#> 104   2       1         10
#> 105   8       1         10
#> 106   3       1         10
#> 107   4       1         10
#> 108   2       1         10
#> 109   5       1         10
#> 110   4       1         10
#> 111   2       1         11
#> 112   2       1         11
#> 113   6       1         11
#> 114   4       1         11
#> 115   5       1         11
#> 116   5       1         11
#> 117   3       1         11
#> 118   2       1         11
#> 119   5       1         11
#> 120   5       1         11
#> 121   8       1         12
#> 122   7       1         12
#> 123   4       1         12
#> 124   7       1         12
#> 125   4       1         12
#> 126   3       1         12
#> 127   5       1         12
#> 128   6       1         12
#> 129   2       1         12
#> 130   5       1         12
#> 131   3       1         13
#> 132   1       1         13
#> 133   6       1         13
#> 134   7       1         13
#> 135   6       1         13
#> 136   2       1         13
#> 137   6       1         13
#> 138  10       1         13
#> 139   7       1         13
#> 140   5       1         13
#> 141   4       1         14
#> 142   6       1         14
#> 143   3       1         14
#> 144   6       1         14
#> 145  12       1         14
#> 146   5       1         14
#> 147   5       1         14
#> 148   5       1         14
#> 149   4       1         14
#> 150  11       1         14
#> 151   4       1         15
#> 152   6       1         15
#> 153   2       1         15
#> 154   9       1         15
#> 155   6       1         15
#> 156   6       1         15
#> 157   9       1         15
#> 158   4       1         15
#> 159   6       1         15
#> 160   4       1         15
#> 161   5       1         16
#> 162  10       1         16
#> 163   6       1         16
#> 164  11       1         16
#> 165   9       1         16
#> 166   8       1         16
#> 167   3       1         16
#> 168   7       1         16
#> 169   7       1         16
#> 170   5       1         16
#> 171   4       1         17
#> 172   3       1         17
#> 173   6       1         17
#> 174   5       1         17
#> 175   8       1         17
#> 176   7       1         17
#> 177   7       1         17
#> 178   5       1         17
#> 179   6       1         17
#> 180   8       1         17
#> 181   8       1         18
#> 182   7       1         18
#> 183   9       1         18
#> 184  10       1         18
#> 185   8       1         18
#> 186   4       1         18
#> 187   8       1         18
#> 188   5       1         18
#> 189  11       1         18
#> 190  13       1         18
#> 191   8       1         19
#> 192  10       1         19
#> 193   8       1         19
#> 194   7       1         19
#> 195   5       1         19
#> 196   9       1         19
#> 197   5       1         19
#> 198   9       1         19
#> 199   9       1         19
#> 200   7       1         19
#> 201  10       1         20
#> 202  10       1         20
#> 203   5       1         20
#> 204   9       1         20
#> 205  13       1         20
#> 206   7       1         20
#> 207   3       1         20
#> 208   5       1         20
#> 209   4       1         20
#> 210   7       1         20
#> 211   5       1         21
#> 212   7       1         21
#> 213   5       1         21
#> 214  10       1         21
#> 215   6       1         21
#> 216   8       1         21
#> 217   6       1         21
#> 218   6       1         21
#> 219  11       1         21
#> 220  12       1         21
#> 221   7       1         22
#> 222   8       1         22
#> 223   7       1         22
#> 224   8       1         22
#> 225  12       1         22
#> 226   7       1         22
#> 227   8       1         22
#> 228  13       1         22
#> 229   5       1         22
#> 230  13       1         22
#> 231   4       1         23
#> 232   6       1         23
#> 233   8       1         23
#> 234   5       1         23
#> 235  11       1         23
#> 236  14       1         23
#> 237   8       1         23
#> 238  15       1         23
#> 239   5       1         23
#> 240  10       1         23
#> 241   6       1         24
#> 242  12       1         24
#> 243  10       1         24
#> 244  12       1         24
#> 245   7       1         24
#> 246  11       1         24
#> 247  10       1         24
#> 248  10       1         24
#> 249  10       1         24
#> 250   9       1         24
#> 251   8       1         25
#> 252   6       1         25
#> 253  11       1         25
#> 254  14       1         25
#> 255  14       1         25
#> 256   9       1         25
#> 257   4       1         25
#> 258   9       1         25
#> 259   9       1         25
#> 260  13       1         25
#> 261   8       1         26
#> 262   9       1         26
#> 263  12       1         26
#> 264  14       1         26
#> 265  11       1         26
#> 266   7       1         26
#> 267   8       1         26
#> 268   9       1         26
#> 269  15       1         26
#> 270  15       1         26
#> 271   6       1         27
#> 272  10       1         27
#> 273  15       1         27
#> 274   9       1         27
#> 275  12       1         27
#> 276  16       1         27
#> 277  10       1         27
#> 278  15       1         27
#> 279   9       1         27
#> 280  14       1         27
#> 281  10       1         28
#> 282  16       1         28
#> 283  13       1         28
#> 284  12       1         28
#> 285  13       1         28
#> 286   9       1         28
#> 287  11       1         28
#> 288   9       1         28
#> 289  10       1         28
#> 290   6       1         28
#> 291  12       1         29
#> 292  13       1         29
#> 293  12       1         29
#> 294   9       1         29
#> 295  12       1         29
#> 296  12       1         29
#> 297  10       1         29
#> 298  13       1         29
#> 299  16       1         29
#> 300  12       1         29
#> 301   9       1         30
#> 302   6       1         30
#> 303  11       1         30
#> 304  10       1         30
#> 305  11       1         30
#> 306  18       1         30
#> 307  10       1         30
#> 308  15       1         30
#> 309  11       1         30
#> 310  11       1         30
#> 311  13       1         31
#> 312  15       1         31
#> 313  13       1         31
#> 314   9       1         31
#> 315  12       1         31
#> 316  11       1         31
#> 317  13       1         31
#> 318  11       1         31
#> 319  10       1         31
#> 320  15       1         31
#> 321  11       1         32
#> 322  10       1         32
#> 323   8       1         32
#> 324  15       1         32
#> 325  12       1         32
#> 326  13       1         32
#> 327  14       1         32
#> 328  11       1         32
#> 329  13       1         32
#> 330  18       1         32
#> 331  10       1         33
#> 332  14       1         33
#> 333  15       1         33
#> 334  12       1         33
#> 335  13       1         33
#> 336  10       1         33
#> 337  11       1         33
#> 338  10       1         33
#> 339  14       1         33
#> 340   9       1         33
#> 341  16       1         34
#> 342  13       1         34
#> 343  20       1         34
#> 344  18       1         34
#> 345  16       1         34
#> 346  12       1         34
#> 347  11       1         34
#> 348  14       1         34
#> 349  14       1         34
#> 350  12       1         34
#> 351   9       1         35
#> 352   8       1         35
#> 353  17       1         35
#> 354  10       1         35
#> 355  17       1         35
#> 356  14       1         35
#> 357  16       1         35
#> 358  12       1         35
#> 359  12       1         35
#> 360  14       1         35
#> 361  10       1         36
#> 362  15       1         36
#> 363  14       1         36
#> 364  13       1         36
#> 365  17       1         36
#> 366  14       1         36
#> 367  16       1         36
#> 368  14       1         36
#> 369  13       1         36
#> 370  19       1         36
#> 371  22       1         37
#> 372  16       1         37
#> 373  16       1         37
#> 374  18       1         37
#> 375  17       1         37
#> 376  14       1         37
#> 377  13       1         37
#> 378  10       1         37
#> 379  24       1         37
#> 380  14       1         37
#> 381  18       1         38
#> 382  13       1         38
#> 383  15       1         38
#> 384  14       1         38
#> 385  11       1         38
#> 386  14       1         38
#> 387  15       1         38
#> 388  12       1         38
#> 389  17       1         38
#> 390  17       1         38
#> 391  17       1         39
#> 392  14       1         39
#> 393  13       1         39
#> 394  15       1         39
#> 395  11       1         39
#> 396  15       1         39
#> 397  17       1         39
#> 398  11       1         39
#> 399  21       1         39
#> 400  16       1         39
#> 401  14       1         40
#> 402  14       1         40
#> 403  18       1         40
#> 404  14       1         40
#> 405  15       1         40
#> 406  19       1         40
#> 407  17       1         40
#> 408  14       1         40
#> 409  15       1         40
#> 410  11       1         40
#> 411  19       1         41
#> 412  15       1         41
#> 413  17       1         41
#> 414  16       1         41
#> 415  16       1         41
#> 416  15       1         41
#> 417  22       1         41
#> 418  17       1         41
#> 419  19       1         41
#> 420  15       1         41
#> 421  19       1         42
#> 422  15       1         42
#> 423  15       1         42
#> 424  15       1         42
#> 425  14       1         42
#> 426  14       1         42
#> 427  15       1         42
#> 428  17       1         42
#> 429  15       1         42
#> 430  18       1         42
#> 431  19       1         43
#> 432  16       1         43
#> 433  18       1         43
#> 434  15       1         43
#> 435  29       2         43
#> 436  17       1         43
#> 437  17       1         43
#> 438  18       1         43
#> 439  19       1         43
#> 440  16       1         43
#> 441  22       1         44
#> 442  22       1         44
#> 443  17       1         44
#> 444  21       1         44
#> 445  16       1         44
#> 446  15       1         44
#> 447  20       1         44
#> 448  16       1         44
#> 449  17       1         44
#> 450  18       1         44
#> 451  19       1         45
#> 452  13       1         45
#> 453  18       1         45
#> 454  18       1         45
#> 455  18       1         45
#> 456  15       1         45
#> 457  13       1         45
#> 458  15       1         45
#> 459  23       1         45
#> 460  23       1         45
#> 461  17       1         46
#> 462  20       1         46
#> 463  15       1         46
#> 464  18       1         46
#> 465  19       1         46
#> 466  18       1         46
#> 467  16       1         46
#> 468  14       1         46
#> 469  16       1         46
#> 470  21       1         46
#> 471  21       1         47
#> 472  17       1         47
#> 473  20       1         47
#> 474  20       1         47
#> 475  16       1         47
#> 476  21       1         47
#> 477  17       1         47
#> 478  18       1         47
#> 479  14       1         47
#> 480  18       1         47
#> 481  11       1         48
#> 482  16       1         48
#> 483  17       1         48
#> 484  27       2         48
#> 485  18       1         48
#> 486  19       1         48
#> 487  13       1         48
#> 488  23       1         48
#> 489  17       1         48
#> 490  25       2         48
#> 491  20       1         49
#> 492  21       1         49
#> 493  23       1         49
#> 494  16       1         49
#> 495  27       2         49
#> 496  21       1         49
#> 497  18       1         49
#> 498  18       1         49
#> 499  27       2         49
#> 500  16       1         49
#> 501  17       1         50
#> 502  26       2         50
#> 503  18       1         50
#> 504  20       1         50
#> 505  20       1         50
#> 506  19       1         50
#> 507  24       1         50
#> 508  19       1         50
#> 509  23       1         50
#> 510  26       2         50
#> 511  23       1         51
#> 512  19       1         51
#> 513  24       1         51
#> 514  24       1         51
#> 515  17       1         51
#> 516  21       1         51
#> 517  22       1         51
#> 518  25       2         51
#> 519  20       1         51
#> 520  23       1         51
#> 521  14       1         52
#> 522  19       1         52
#> 523  17       1         52
#> 524  17       1         52
#> 525  24       1         52
#> 526  24       1         52
#> 527  22       1         52
#> 528  20       1         52
#> 529  27       2         52
#> 530  23       1         52
#> 531  19       1         53
#> 532  23       1         53
#> 533  19       1         53
#> 534  22       1         53
#> 535  23       1         53
#> 536  17       1         53
#> 537  22       1         53
#> 538  16       1         53
#> 539  22       1         53
#> 540  20       1         53
#> 541  23       1         54
#> 542  19       1         54
#> 543  19       1         54
#> 544  17       1         54
#> 545  19       1         54
#> 546  21       1         54
#> 547  23       1         54
#> 548  19       1         54
#> 549  22       1         54
#> 550  23       1         54
#> 551  29       2         55
#> 552  21       1         55
#> 553  23       1         55
#> 554  24       1         55
#> 555  30       2         55
#> 556  24       1         55
#> 557  17       1         55
#> 558  16       1         55
#> 559  27       2         55
#> 560  27       2         55
#> 561  20       1         56
#> 562  30       2         56
#> 563  24       1         56
#> 564  19       1         56
#> 565  28       2         56
#> 566  23       1         56
#> 567  13       1         56
#> 568  25       2         56
#> 569  25       2         56
#> 570  27       2         56
#> 571  20       1         57
#> 572  21       1         57
#> 573  21       1         57
#> 574  21       1         57
#> 575  22       1         57
#> 576  27       2         57
#> 577  24       1         57
#> 578  22       1         57
#> 579  18       1         57
#> 580  22       1         57
#> 581  21       1         58
#> 582  25       2         58
#> 583  22       1         58
#> 584  29       2         58
#> 585  23       1         58
#> 586  23       1         58
#> 587  20       1         58
#> 588  21       1         58
#> 589  30       2         58
#> 590  20       1         58
#> 591  19       1         59
#> 592  19       1         59
#> 593  23       1         59
#> 594  22       1         59
#> 595  26       2         59
#> 596  31       2         59
#> 597  22       1         59
#> 598  24       1         59
#> 599  21       1         59
#> 600  24       1         59
#> 601  20       1         60
#> 602  28       2         60
#> 603  23       1         60
#> 604  23       1         60
#> 605  27       2         60
#> 606  21       1         60
#> 607  23       1         60
#> 608  23       1         60
#> 609  24       1         60
#> 610  23       1         60
#> 611  22       1         61
#> 612  29       2         61
#> 613  25       2         61
#> 614  24       1         61
#> 615  24       1         61
#> 616  25       2         61
#> 617  27       2         61
#> 618  22       1         61
#> 619  25       2         61
#> 620  23       1         61
#> 621  29       2         62
#> 622  17       1         62
#> 623  27       2         62
#> 624  26       2         62
#> 625  24       1         62
#> 626  27       2         62
#> 627  28       2         62
#> 628  26       2         62
#> 629  21       1         62
#> 630  28       2         62
#> 631  24       1         63
#> 632  29       2         63
#> 633  20       1         63
#> 634  20       1         63
#> 635  27       2         63
#> 636  19       1         63
#> 637  22       1         63
#> 638  23       1         63
#> 639  27       2         63
#> 640  28       2         63
#> 641  26       2         64
#> 642  29       2         64
#> 643  28       2         64
#> 644  29       2         64
#> 645  26       2         64
#> 646  26       2         64
#> 647  24       1         64
#> 648  26       2         64
#> 649  23       1         64
#> 650  24       1         64
#> 651  25       2         65
#> 652  25       2         65
#> 653  24       1         65
#> 654  30       2         65
#> 655  25       2         65
#> 656  25       2         65
#> 657  27       2         65
#> 658  29       2         65
#> 659  27       2         65
#> 660  28       2         65
#> 661  25       2         66
#> 662  22       1         66
#> 663  25       2         66
#> 664  26       2         66
#> 665  28       2         66
#> 666  23       1         66
#> 667  31       2         66
#> 668  25       2         66
#> 669  28       2         66
#> 670  29       2         66
#> 671  23       1         67
#> 672  26       2         67
#> 673  21       1         67
#> 674  32       2         67
#> 675  25       2         67
#> 676  24       1         67
#> 677  22       1         67
#> 678  19       1         67
#> 679  30       2         67
#> 680  30       2         67
#> 681  26       2         68
#> 682  29       2         68
#> 683  23       1         68
#> 684  22       1         68
#> 685  30       2         68
#> 686  31       2         68
#> 687  32       2         68
#> 688  25       2         68
#> 689  27       2         68
#> 690  28       2         68
#> 691  25       2         69
#> 692  25       2         69
#> 693  29       2         69
#> 694  27       2         69
#> 695  28       2         69
#> 696  26       2         69
#> 697  24       1         69
#> 698  30       2         69
#> 699  28       2         69
#> 700  31       2         69
#> 701  28       2         70
#> 702  25       2         70
#> 703  29       2         70
#> 704  26       2         70
#> 705  28       2         70
#> 706  28       2         70
#> 707  28       2         70
#> 708  30       2         70
#> 709  31       2         70
#> 710  28       2         70
#> 711  25       2         71
#> 712  31       2         71
#> 713  29       2         71
#> 714  28       2         71
#> 715  30       2         71
#> 716  27       2         71
#> 717  27       2         71
#> 718  26       2         71
#> 719  31       2         71
#> 720  28       2         71
#> 721  34       2         72
#> 722  32       2         72
#> 723  26       2         72
#> 724  30       2         72
#> 725  26       2         72
#> 726  20       1         72
#> 727  26       2         72
#> 728  25       2         72
#> 729  28       2         72
#> 730  29       2         72
#> 731  33       2         73
#> 732  27       2         73
#> 733  30       2         73
#> 734  29       2         73
#> 735  30       2         73
#> 736  30       2         73
#> 737  30       2         73
#> 738  30       2         73
#> 739  32       2         73
#> 740  29       2         73
#> 741  27       2         74
#> 742  30       2         74
#> 743  31       2         74
#> 744  31       2         74
#> 745  30       2         74
#> 746  32       2         74
#> 747  26       2         74
#> 748  33       2         74
#> 749  27       2         74
#> 750  27       2         74
#> 751  29       2         75
#> 752  34       2         75
#> 753  26       2         75
#> 754  32       2         75
#> 755  30       2         75
#> 756  26       2         75
#> 757  30       2         75
#> 758  34       2         75
#> 759  29       2         75
#> 760  32       2         75
#> 761  30       2         76
#> 762  28       2         76
#> 763  27       2         76
#> 764  29       2         76
#> 765  34       2         76
#> 766  27       2         76
#> 767  32       2         76
#> 768  30       2         76
#> 769  31       2         76
#> 770  28       2         76
#> 771  29       2         77
#> 772  37       3         77
#> 773  31       2         77
#> 774  27       2         77
#> 775  34       2         77
#> 776  30       2         77
#> 777  28       2         77
#> 778  31       2         77
#> 779  34       2         77
#> 780  38       3         77
#> 781  32       2         78
#> 782  33       2         78
#> 783  25       2         78
#> 784  31       2         78
#> 785  30       2         78
#> 786  35       2         78
#> 787  29       2         78
#> 788  33       2         78
#> 789  25       2         78
#> 790  31       2         78
#> 791  28       2         79
#> 792  35       2         79
#> 793  32       2         79
#> 794  29       2         79
#> 795  34       2         79
#> 796  31       2         79
#> 797  31       2         79
#> 798  29       2         79
#> 799  32       2         79
#> 800  34       2         79
#> 801  32       2         80
#> 802  36       2         80
#> 803  33       2         80
#> 804  33       2         80
#> 805  33       2         80
#> 806  35       2         80
#> 807  31       2         80
#> 808  32       2         80
#> 809  34       2         80
#> 810  33       2         80
#> 811  30       2         81
#> 812  33       2         81
#> 813  33       2         81
#> 814  33       2         81
#> 815  26       2         81
#> 816  33       2         81
#> 817  34       2         81
#> 818  34       2         81
#> 819  32       2         81
#> 820  36       2         81
#> 821  34       2         82
#> 822  33       2         82
#> 823  32       2         82
#> 824  34       2         82
#> 825  28       2         82
#> 826  31       2         82
#> 827  32       2         82
#> 828  36       2         82
#> 829  35       2         82
#> 830  28       2         82
#> 831  34       2         83
#> 832  35       2         83
#> 833  35       2         83
#> 834  34       2         83
#> 835  28       2         83
#> 836  36       2         83
#> 837  31       2         83
#> 838  33       2         83
#> 839  35       2         83
#> 840  31       2         83
#> 841  32       2         84
#> 842  33       2         84
#> 843  33       2         84
#> 844  32       2         84
#> 845  37       3         84
#> 846  31       2         84
#> 847  35       2         84
#> 848  31       2         84
#> 849  35       2         84
#> 850  35       2         84
#> 851  31       2         85
#> 852  39       3         85
#> 853  34       2         85
#> 854  26       2         85
#> 855  31       2         85
#> 856  34       2         85
#> 857  36       2         85
#> 858  33       2         85
#> 859  35       2         85
#> 860  36       2         85
#> 861  38       3         86
#> 862  35       2         86
#> 863  36       2         86
#> 864  35       2         86
#> 865  35       2         86
#> 866  37       3         86
#> 867  32       2         86
#> 868  32       2         86
#> 869  31       2         86
#> 870  34       2         86
#> 871  35       2         87
#> 872  38       3         87
#> 873  37       3         87
#> 874  35       2         87
#> 875  33       2         87
#> 876  36       2         87
#> 877  34       2         87
#> 878  35       2         87
#> 879  32       2         87
#> 880  37       3         87
#> 881  37       3         88
#> 882  32       2         88
#> 883  33       2         88
#> 884  33       2         88
#> 885  35       2         88
#> 886  35       2         88
#> 887  38       3         88
#> 888  36       2         88
#> 889  37       3         88
#> 890  37       3         88
#> 891  35       2         89
#> 892  31       2         89
#> 893  38       3         89
#> 894  31       2         89
#> 895  38       3         89
#> 896  32       2         89
#> 897  32       2         89
#> 898  34       2         89
#> 899  34       2         89
#> 900  35       2         89
#> 901  37       3         90
#> 902  33       2         90
#> 903  40       3         90
#> 904  35       2         90
#> 905  36       2         90
#> 906  36       2         90
#> 907  36       2         90
#> 908  35       2         90
#> 909  35       2         90
#> 910  36       2         90
#> 911  32       2         91
#> 912  39       3         91
#> 913  40       3         91
#> 914  35       2         91
#> 915  33       2         91
#> 916  37       3         91
#> 917  36       2         91
#> 918  37       3         91
#> 919  38       3         91
#> 920  38       3         91
#> 921  37       3         92
#> 922  33       2         92
#> 923  38       3         92
#> 924  38       3         92
#> 925  35       2         92
#> 926  33       2         92
#> 927  39       3         92
#> 928  38       3         92
#> 929  37       3         92
#> 930  36       2         92
#> 931  36       2         93
#> 932  36       2         93
#> 933  40       3         93
#> 934  34       2         93
#> 935  40       3         93
#> 936  37       3         93
#> 937  38       3         93
#> 938  38       3         93
#> 939  36       2         93
#> 940  38       3         93
#> 941  39       3         94
#> 942  38       3         94
#> 943  39       3         94
#> 944  37       3         94
#> 945  36       2         94
#> 946  38       3         94
#> 947  36       2         94
#> 948  39       3         94
#> 949  38       3         94
#> 950  39       3         94
#> 951  38       3         95
#> 952  38       3         95
#> 953  37       3         95
#> 954  38       3         95
#> 955  37       3         95
#> 956  40       3         95
#> 957  39       3         95
#> 958  40       3         95
#> 959  37       3         95
#> 960  38       3         95
#> 961  36       2         96
#> 962  36       2         96
#> 963  39       3         96
#> 964  39       3         96
#> 965  40       3         96
#> 966  39       3         96
#> 967  38       3         96
#> 968  39       3         96
#> 969  39       3         96
#> 970  38       3         96
#> 971  39       3         97
#> 972  39       3         97
#> 973  39       3         97
#> 974  40       3         97
#> 975  39       3         97
#> 976  40       3         97
#> 977  40       3         97
#> 978  40       3         97
#> 979  40       3         97
#> 980  38       3         97
#> 981  39       3         98
#> 982  40       3         98
#> 983  40       3         98
#> 984  39       3         98
#> 985  40       3         98
#> 986  39       3         98
#> 987  39       3         98
#> 988  40       3         98
#> 989  39       3         98
#> 990  40       3         98
#> 991  39       3         99
#> 992  40       3         99
#> 993  40       3         99
#> 994  39       3         99
#> 995  39       3         99
#> 996  40       3         99
#> 997  40       3         99
#> 998  38       3         99
#> 999  39       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100