Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

Usage

simulate_lqas(
  runs = 50,
  pop = NULL,
  n,
  d.lower,
  d.upper,
  p.lower = 0,
  p.upper = 100,
  fine = 1,
  progress = TRUE
)

Arguments

runs

Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop

Population size from which simulated coverage survey data is to be taken from

n

Sample size of actual or test coverage data

d.lower

A numeric value for the lower classification threshold

d.upper

A numeric value for the upper classification threshold

p.lower

Starting proportion for simulations. Default is 0

p.upper

Ending proportion for simulations. Default is 100

fine

Granularity of simulated proportions; Default to 1

progress

Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simulate_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)
#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%
#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    0       1          1
#> 13    1       1          1
#> 14    2       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    2       1          1
#> 19    1       1          1
#> 20    1       1          1
#> 21    2       1          2
#> 22    0       1          2
#> 23    0       1          2
#> 24    1       1          2
#> 25    1       1          2
#> 26    1       1          2
#> 27    0       1          2
#> 28    0       1          2
#> 29    2       1          2
#> 30    0       1          2
#> 31    5       1          3
#> 32    0       1          3
#> 33    0       1          3
#> 34    0       1          3
#> 35    0       1          3
#> 36    0       1          3
#> 37    0       1          3
#> 38    1       1          3
#> 39    1       1          3
#> 40    3       1          3
#> 41    3       1          4
#> 42    0       1          4
#> 43    2       1          4
#> 44    3       1          4
#> 45    1       1          4
#> 46    0       1          4
#> 47    1       1          4
#> 48    1       1          4
#> 49    2       1          4
#> 50    1       1          4
#> 51    1       1          5
#> 52    2       1          5
#> 53    3       1          5
#> 54    0       1          5
#> 55    3       1          5
#> 56    0       1          5
#> 57    1       1          5
#> 58    3       1          5
#> 59    3       1          5
#> 60    0       1          5
#> 61    1       1          6
#> 62    2       1          6
#> 63    1       1          6
#> 64    2       1          6
#> 65    3       1          6
#> 66    2       1          6
#> 67    3       1          6
#> 68    0       1          6
#> 69    1       1          6
#> 70    2       1          6
#> 71    1       1          7
#> 72    2       1          7
#> 73    2       1          7
#> 74    3       1          7
#> 75    3       1          7
#> 76    0       1          7
#> 77    1       1          7
#> 78    1       1          7
#> 79    2       1          7
#> 80   10       1          7
#> 81    4       1          8
#> 82    5       1          8
#> 83    0       1          8
#> 84    0       1          8
#> 85    3       1          8
#> 86    5       1          8
#> 87    0       1          8
#> 88    3       1          8
#> 89    5       1          8
#> 90    2       1          8
#> 91    2       1          9
#> 92    4       1          9
#> 93    6       1          9
#> 94    5       1          9
#> 95    3       1          9
#> 96    3       1          9
#> 97    7       1          9
#> 98    3       1          9
#> 99    4       1          9
#> 100   8       1          9
#> 101   2       1         10
#> 102   5       1         10
#> 103   7       1         10
#> 104   3       1         10
#> 105   4       1         10
#> 106   5       1         10
#> 107   2       1         10
#> 108   7       1         10
#> 109   1       1         10
#> 110   4       1         10
#> 111   2       1         11
#> 112   3       1         11
#> 113   3       1         11
#> 114   8       1         11
#> 115   3       1         11
#> 116   8       1         11
#> 117   3       1         11
#> 118   8       1         11
#> 119   5       1         11
#> 120   6       1         11
#> 121   9       1         12
#> 122   1       1         12
#> 123   6       1         12
#> 124   3       1         12
#> 125   4       1         12
#> 126   5       1         12
#> 127   3       1         12
#> 128   7       1         12
#> 129   5       1         12
#> 130   9       1         12
#> 131   8       1         13
#> 132   5       1         13
#> 133   3       1         13
#> 134   2       1         13
#> 135   6       1         13
#> 136   8       1         13
#> 137   7       1         13
#> 138   5       1         13
#> 139  10       1         13
#> 140   6       1         13
#> 141   1       1         14
#> 142   8       1         14
#> 143   7       1         14
#> 144   8       1         14
#> 145   9       1         14
#> 146   5       1         14
#> 147   3       1         14
#> 148   4       1         14
#> 149   4       1         14
#> 150   5       1         14
#> 151   6       1         15
#> 152   5       1         15
#> 153   6       1         15
#> 154   7       1         15
#> 155   2       1         15
#> 156   5       1         15
#> 157   9       1         15
#> 158   2       1         15
#> 159   4       1         15
#> 160   4       1         15
#> 161   1       1         16
#> 162   8       1         16
#> 163  11       1         16
#> 164   6       1         16
#> 165  10       1         16
#> 166  11       1         16
#> 167   5       1         16
#> 168   4       1         16
#> 169   5       1         16
#> 170   5       1         16
#> 171   4       1         17
#> 172   7       1         17
#> 173   6       1         17
#> 174   1       1         17
#> 175   7       1         17
#> 176   3       1         17
#> 177   8       1         17
#> 178  10       1         17
#> 179   9       1         17
#> 180   6       1         17
#> 181   6       1         18
#> 182   6       1         18
#> 183   4       1         18
#> 184  13       1         18
#> 185   9       1         18
#> 186   6       1         18
#> 187  10       1         18
#> 188   9       1         18
#> 189   8       1         18
#> 190   6       1         18
#> 191   6       1         19
#> 192   8       1         19
#> 193  11       1         19
#> 194   9       1         19
#> 195   8       1         19
#> 196   4       1         19
#> 197   8       1         19
#> 198   5       1         19
#> 199   3       1         19
#> 200   6       1         19
#> 201   8       1         20
#> 202   7       1         20
#> 203   8       1         20
#> 204   4       1         20
#> 205   9       1         20
#> 206   5       1         20
#> 207   5       1         20
#> 208   8       1         20
#> 209   7       1         20
#> 210  11       1         20
#> 211  12       1         21
#> 212   5       1         21
#> 213   8       1         21
#> 214   8       1         21
#> 215   6       1         21
#> 216   7       1         21
#> 217   8       1         21
#> 218  10       1         21
#> 219   9       1         21
#> 220  14       1         21
#> 221   6       1         22
#> 222  10       1         22
#> 223   6       1         22
#> 224   9       1         22
#> 225  10       1         22
#> 226  10       1         22
#> 227  12       1         22
#> 228   5       1         22
#> 229   7       1         22
#> 230  11       1         22
#> 231   9       1         23
#> 232  11       1         23
#> 233   9       1         23
#> 234  11       1         23
#> 235  11       1         23
#> 236   7       1         23
#> 237  12       1         23
#> 238  10       1         23
#> 239  11       1         23
#> 240  12       1         23
#> 241   9       1         24
#> 242  12       1         24
#> 243  17       1         24
#> 244  12       1         24
#> 245  11       1         24
#> 246  13       1         24
#> 247  11       1         24
#> 248   7       1         24
#> 249  11       1         24
#> 250  12       1         24
#> 251   7       1         25
#> 252   7       1         25
#> 253  12       1         25
#> 254   9       1         25
#> 255  11       1         25
#> 256  11       1         25
#> 257   7       1         25
#> 258  10       1         25
#> 259   9       1         25
#> 260   8       1         25
#> 261   8       1         26
#> 262  12       1         26
#> 263  11       1         26
#> 264   9       1         26
#> 265   6       1         26
#> 266  10       1         26
#> 267  14       1         26
#> 268   8       1         26
#> 269   6       1         26
#> 270  13       1         26
#> 271   7       1         27
#> 272  13       1         27
#> 273  11       1         27
#> 274  12       1         27
#> 275  11       1         27
#> 276  14       1         27
#> 277  13       1         27
#> 278  18       1         27
#> 279  11       1         27
#> 280  15       1         27
#> 281  11       1         28
#> 282   7       1         28
#> 283  10       1         28
#> 284  13       1         28
#> 285   8       1         28
#> 286  15       1         28
#> 287  10       1         28
#> 288  10       1         28
#> 289  13       1         28
#> 290  11       1         28
#> 291   8       1         29
#> 292   9       1         29
#> 293  12       1         29
#> 294  15       1         29
#> 295  13       1         29
#> 296  12       1         29
#> 297   7       1         29
#> 298  12       1         29
#> 299  12       1         29
#> 300  12       1         29
#> 301  16       1         30
#> 302   8       1         30
#> 303  10       1         30
#> 304  14       1         30
#> 305  11       1         30
#> 306  13       1         30
#> 307  15       1         30
#> 308  11       1         30
#> 309  11       1         30
#> 310  11       1         30
#> 311   9       1         31
#> 312  19       1         31
#> 313  10       1         31
#> 314  14       1         31
#> 315  14       1         31
#> 316  11       1         31
#> 317  17       1         31
#> 318  10       1         31
#> 319   8       1         31
#> 320  16       1         31
#> 321  19       1         32
#> 322  10       1         32
#> 323  12       1         32
#> 324  14       1         32
#> 325  13       1         32
#> 326  12       1         32
#> 327   9       1         32
#> 328  10       1         32
#> 329  12       1         32
#> 330   9       1         32
#> 331  13       1         33
#> 332  19       1         33
#> 333   6       1         33
#> 334  16       1         33
#> 335  10       1         33
#> 336  15       1         33
#> 337  16       1         33
#> 338  12       1         33
#> 339  13       1         33
#> 340  12       1         33
#> 341  13       1         34
#> 342  15       1         34
#> 343  16       1         34
#> 344  15       1         34
#> 345  14       1         34
#> 346   8       1         34
#> 347  10       1         34
#> 348  17       1         34
#> 349  16       1         34
#> 350  16       1         34
#> 351  18       1         35
#> 352  15       1         35
#> 353  12       1         35
#> 354  14       1         35
#> 355  13       1         35
#> 356   6       1         35
#> 357  20       1         35
#> 358  18       1         35
#> 359  16       1         35
#> 360  15       1         35
#> 361  14       1         36
#> 362  14       1         36
#> 363  17       1         36
#> 364  16       1         36
#> 365  17       1         36
#> 366  15       1         36
#> 367  15       1         36
#> 368  11       1         36
#> 369  16       1         36
#> 370  14       1         36
#> 371  11       1         37
#> 372  14       1         37
#> 373  12       1         37
#> 374  18       1         37
#> 375  18       1         37
#> 376  16       1         37
#> 377  11       1         37
#> 378  14       1         37
#> 379   8       1         37
#> 380  16       1         37
#> 381  16       1         38
#> 382  16       1         38
#> 383   9       1         38
#> 384  16       1         38
#> 385  14       1         38
#> 386  19       1         38
#> 387  15       1         38
#> 388  17       1         38
#> 389  14       1         38
#> 390  20       1         38
#> 391   9       1         39
#> 392  16       1         39
#> 393  16       1         39
#> 394  17       1         39
#> 395  17       1         39
#> 396  13       1         39
#> 397  18       1         39
#> 398  17       1         39
#> 399  13       1         39
#> 400  13       1         39
#> 401  17       1         40
#> 402  19       1         40
#> 403  22       1         40
#> 404  13       1         40
#> 405  10       1         40
#> 406  13       1         40
#> 407  18       1         40
#> 408  12       1         40
#> 409  20       1         40
#> 410   9       1         40
#> 411  13       1         41
#> 412  16       1         41
#> 413  19       1         41
#> 414  10       1         41
#> 415  17       1         41
#> 416  22       1         41
#> 417  15       1         41
#> 418  14       1         41
#> 419  19       1         41
#> 420  17       1         41
#> 421  14       1         42
#> 422  12       1         42
#> 423  20       1         42
#> 424  15       1         42
#> 425  24       1         42
#> 426  14       1         42
#> 427  14       1         42
#> 428  18       1         42
#> 429  17       1         42
#> 430  17       1         42
#> 431  18       1         43
#> 432  14       1         43
#> 433  13       1         43
#> 434  16       1         43
#> 435  21       1         43
#> 436  20       1         43
#> 437  16       1         43
#> 438  16       1         43
#> 439  18       1         43
#> 440  20       1         43
#> 441  18       1         44
#> 442  18       1         44
#> 443  20       1         44
#> 444  21       1         44
#> 445  20       1         44
#> 446  20       1         44
#> 447  16       1         44
#> 448  16       1         44
#> 449  15       1         44
#> 450  16       1         44
#> 451  17       1         45
#> 452  15       1         45
#> 453  16       1         45
#> 454  16       1         45
#> 455  20       1         45
#> 456  17       1         45
#> 457  21       1         45
#> 458  16       1         45
#> 459  16       1         45
#> 460  11       1         45
#> 461  17       1         46
#> 462  20       1         46
#> 463  13       1         46
#> 464  19       1         46
#> 465  23       1         46
#> 466  11       1         46
#> 467  22       1         46
#> 468  14       1         46
#> 469  24       1         46
#> 470  16       1         46
#> 471  18       1         47
#> 472  19       1         47
#> 473  21       1         47
#> 474  21       1         47
#> 475  23       1         47
#> 476  14       1         47
#> 477  21       1         47
#> 478  17       1         47
#> 479  21       1         47
#> 480  21       1         47
#> 481  22       1         48
#> 482  21       1         48
#> 483  18       1         48
#> 484  19       1         48
#> 485  19       1         48
#> 486  24       1         48
#> 487  25       2         48
#> 488  22       1         48
#> 489  11       1         48
#> 490  22       1         48
#> 491  18       1         49
#> 492  20       1         49
#> 493  15       1         49
#> 494  22       1         49
#> 495  17       1         49
#> 496  21       1         49
#> 497  18       1         49
#> 498  24       1         49
#> 499  19       1         49
#> 500  16       1         49
#> 501  23       1         50
#> 502  18       1         50
#> 503  16       1         50
#> 504  17       1         50
#> 505  21       1         50
#> 506  26       2         50
#> 507  21       1         50
#> 508  18       1         50
#> 509  20       1         50
#> 510  19       1         50
#> 511  18       1         51
#> 512  23       1         51
#> 513  24       1         51
#> 514  23       1         51
#> 515  26       2         51
#> 516  18       1         51
#> 517  19       1         51
#> 518  22       1         51
#> 519  16       1         51
#> 520  26       2         51
#> 521  20       1         52
#> 522  15       1         52
#> 523  21       1         52
#> 524  19       1         52
#> 525  20       1         52
#> 526  16       1         52
#> 527  17       1         52
#> 528  17       1         52
#> 529  23       1         52
#> 530  23       1         52
#> 531  19       1         53
#> 532  23       1         53
#> 533  20       1         53
#> 534  24       1         53
#> 535  22       1         53
#> 536  19       1         53
#> 537  18       1         53
#> 538  24       1         53
#> 539  24       1         53
#> 540  22       1         53
#> 541  22       1         54
#> 542  23       1         54
#> 543  20       1         54
#> 544  24       1         54
#> 545  18       1         54
#> 546  26       2         54
#> 547  25       2         54
#> 548  20       1         54
#> 549  22       1         54
#> 550  20       1         54
#> 551  23       1         55
#> 552  23       1         55
#> 553  22       1         55
#> 554  24       1         55
#> 555  22       1         55
#> 556  21       1         55
#> 557  25       2         55
#> 558  19       1         55
#> 559  22       1         55
#> 560  26       2         55
#> 561  20       1         56
#> 562  23       1         56
#> 563  22       1         56
#> 564  24       1         56
#> 565  22       1         56
#> 566  22       1         56
#> 567  27       2         56
#> 568  24       1         56
#> 569  22       1         56
#> 570  18       1         56
#> 571  15       1         57
#> 572  15       1         57
#> 573  23       1         57
#> 574  23       1         57
#> 575  24       1         57
#> 576  27       2         57
#> 577  29       2         57
#> 578  23       1         57
#> 579  25       2         57
#> 580  24       1         57
#> 581  26       2         58
#> 582  31       2         58
#> 583  22       1         58
#> 584  22       1         58
#> 585  26       2         58
#> 586  24       1         58
#> 587  17       1         58
#> 588  23       1         58
#> 589  26       2         58
#> 590  20       1         58
#> 591  22       1         59
#> 592  26       2         59
#> 593  25       2         59
#> 594  23       1         59
#> 595  22       1         59
#> 596  28       2         59
#> 597  25       2         59
#> 598  24       1         59
#> 599  23       1         59
#> 600  25       2         59
#> 601  22       1         60
#> 602  22       1         60
#> 603  23       1         60
#> 604  21       1         60
#> 605  22       1         60
#> 606  26       2         60
#> 607  21       1         60
#> 608  16       1         60
#> 609  25       2         60
#> 610  22       1         60
#> 611  24       1         61
#> 612  22       1         61
#> 613  31       2         61
#> 614  26       2         61
#> 615  23       1         61
#> 616  27       2         61
#> 617  22       1         61
#> 618  30       2         61
#> 619  25       2         61
#> 620  28       2         61
#> 621  25       2         62
#> 622  28       2         62
#> 623  24       1         62
#> 624  22       1         62
#> 625  25       2         62
#> 626  20       1         62
#> 627  24       1         62
#> 628  23       1         62
#> 629  26       2         62
#> 630  26       2         62
#> 631  31       2         63
#> 632  27       2         63
#> 633  23       1         63
#> 634  27       2         63
#> 635  19       1         63
#> 636  25       2         63
#> 637  28       2         63
#> 638  25       2         63
#> 639  24       1         63
#> 640  25       2         63
#> 641  23       1         64
#> 642  26       2         64
#> 643  25       2         64
#> 644  26       2         64
#> 645  22       1         64
#> 646  19       1         64
#> 647  25       2         64
#> 648  31       2         64
#> 649  23       1         64
#> 650  22       1         64
#> 651  25       2         65
#> 652  24       1         65
#> 653  25       2         65
#> 654  26       2         65
#> 655  28       2         65
#> 656  17       1         65
#> 657  27       2         65
#> 658  27       2         65
#> 659  30       2         65
#> 660  25       2         65
#> 661  24       1         66
#> 662  27       2         66
#> 663  26       2         66
#> 664  25       2         66
#> 665  26       2         66
#> 666  29       2         66
#> 667  26       2         66
#> 668  25       2         66
#> 669  32       2         66
#> 670  22       1         66
#> 671  27       2         67
#> 672  29       2         67
#> 673  29       2         67
#> 674  28       2         67
#> 675  26       2         67
#> 676  24       1         67
#> 677  27       2         67
#> 678  22       1         67
#> 679  26       2         67
#> 680  25       2         67
#> 681  24       1         68
#> 682  25       2         68
#> 683  29       2         68
#> 684  24       1         68
#> 685  23       1         68
#> 686  27       2         68
#> 687  28       2         68
#> 688  29       2         68
#> 689  29       2         68
#> 690  22       1         68
#> 691  26       2         69
#> 692  25       2         69
#> 693  23       1         69
#> 694  24       1         69
#> 695  31       2         69
#> 696  25       2         69
#> 697  28       2         69
#> 698  29       2         69
#> 699  23       1         69
#> 700  32       2         69
#> 701  26       2         70
#> 702  28       2         70
#> 703  26       2         70
#> 704  29       2         70
#> 705  30       2         70
#> 706  31       2         70
#> 707  25       2         70
#> 708  26       2         70
#> 709  26       2         70
#> 710  31       2         70
#> 711  32       2         71
#> 712  27       2         71
#> 713  30       2         71
#> 714  29       2         71
#> 715  22       1         71
#> 716  28       2         71
#> 717  27       2         71
#> 718  30       2         71
#> 719  32       2         71
#> 720  31       2         71
#> 721  31       2         72
#> 722  30       2         72
#> 723  28       2         72
#> 724  29       2         72
#> 725  28       2         72
#> 726  26       2         72
#> 727  36       2         72
#> 728  33       2         72
#> 729  34       2         72
#> 730  29       2         72
#> 731  29       2         73
#> 732  33       2         73
#> 733  27       2         73
#> 734  31       2         73
#> 735  33       2         73
#> 736  28       2         73
#> 737  29       2         73
#> 738  31       2         73
#> 739  37       3         73
#> 740  32       2         73
#> 741  30       2         74
#> 742  30       2         74
#> 743  27       2         74
#> 744  31       2         74
#> 745  27       2         74
#> 746  28       2         74
#> 747  35       2         74
#> 748  32       2         74
#> 749  37       3         74
#> 750  30       2         74
#> 751  32       2         75
#> 752  32       2         75
#> 753  27       2         75
#> 754  29       2         75
#> 755  27       2         75
#> 756  31       2         75
#> 757  30       2         75
#> 758  31       2         75
#> 759  36       2         75
#> 760  29       2         75
#> 761  34       2         76
#> 762  30       2         76
#> 763  33       2         76
#> 764  33       2         76
#> 765  31       2         76
#> 766  29       2         76
#> 767  29       2         76
#> 768  30       2         76
#> 769  26       2         76
#> 770  31       2         76
#> 771  34       2         77
#> 772  31       2         77
#> 773  29       2         77
#> 774  30       2         77
#> 775  33       2         77
#> 776  29       2         77
#> 777  29       2         77
#> 778  27       2         77
#> 779  30       2         77
#> 780  32       2         77
#> 781  27       2         78
#> 782  30       2         78
#> 783  35       2         78
#> 784  32       2         78
#> 785  32       2         78
#> 786  27       2         78
#> 787  30       2         78
#> 788  32       2         78
#> 789  30       2         78
#> 790  33       2         78
#> 791  35       2         79
#> 792  32       2         79
#> 793  32       2         79
#> 794  29       2         79
#> 795  33       2         79
#> 796  34       2         79
#> 797  30       2         79
#> 798  30       2         79
#> 799  32       2         79
#> 800  35       2         79
#> 801  32       2         80
#> 802  26       2         80
#> 803  29       2         80
#> 804  32       2         80
#> 805  34       2         80
#> 806  26       2         80
#> 807  32       2         80
#> 808  28       2         80
#> 809  29       2         80
#> 810  28       2         80
#> 811  32       2         81
#> 812  34       2         81
#> 813  34       2         81
#> 814  33       2         81
#> 815  34       2         81
#> 816  35       2         81
#> 817  31       2         81
#> 818  26       2         81
#> 819  31       2         81
#> 820  33       2         81
#> 821  33       2         82
#> 822  33       2         82
#> 823  32       2         82
#> 824  34       2         82
#> 825  35       2         82
#> 826  30       2         82
#> 827  35       2         82
#> 828  33       2         82
#> 829  32       2         82
#> 830  33       2         82
#> 831  31       2         83
#> 832  32       2         83
#> 833  33       2         83
#> 834  32       2         83
#> 835  36       2         83
#> 836  33       2         83
#> 837  30       2         83
#> 838  32       2         83
#> 839  37       3         83
#> 840  32       2         83
#> 841  39       3         84
#> 842  36       2         84
#> 843  31       2         84
#> 844  32       2         84
#> 845  33       2         84
#> 846  34       2         84
#> 847  30       2         84
#> 848  35       2         84
#> 849  29       2         84
#> 850  31       2         84
#> 851  37       3         85
#> 852  32       2         85
#> 853  38       3         85
#> 854  35       2         85
#> 855  36       2         85
#> 856  31       2         85
#> 857  34       2         85
#> 858  32       2         85
#> 859  35       2         85
#> 860  34       2         85
#> 861  32       2         86
#> 862  37       3         86
#> 863  37       3         86
#> 864  36       2         86
#> 865  35       2         86
#> 866  38       3         86
#> 867  29       2         86
#> 868  37       3         86
#> 869  33       2         86
#> 870  36       2         86
#> 871  37       3         87
#> 872  37       3         87
#> 873  37       3         87
#> 874  36       2         87
#> 875  36       2         87
#> 876  38       3         87
#> 877  35       2         87
#> 878  36       2         87
#> 879  36       2         87
#> 880  34       2         87
#> 881  34       2         88
#> 882  36       2         88
#> 883  36       2         88
#> 884  34       2         88
#> 885  37       3         88
#> 886  35       2         88
#> 887  38       3         88
#> 888  35       2         88
#> 889  33       2         88
#> 890  35       2         88
#> 891  33       2         89
#> 892  30       2         89
#> 893  34       2         89
#> 894  37       3         89
#> 895  36       2         89
#> 896  38       3         89
#> 897  34       2         89
#> 898  36       2         89
#> 899  35       2         89
#> 900  36       2         89
#> 901  31       2         90
#> 902  35       2         90
#> 903  38       3         90
#> 904  36       2         90
#> 905  34       2         90
#> 906  38       3         90
#> 907  34       2         90
#> 908  37       3         90
#> 909  35       2         90
#> 910  38       3         90
#> 911  37       3         91
#> 912  36       2         91
#> 913  35       2         91
#> 914  34       2         91
#> 915  35       2         91
#> 916  36       2         91
#> 917  40       3         91
#> 918  39       3         91
#> 919  35       2         91
#> 920  37       3         91
#> 921  38       3         92
#> 922  39       3         92
#> 923  37       3         92
#> 924  38       3         92
#> 925  36       2         92
#> 926  37       3         92
#> 927  37       3         92
#> 928  35       2         92
#> 929  38       3         92
#> 930  36       2         92
#> 931  36       2         93
#> 932  36       2         93
#> 933  36       2         93
#> 934  39       3         93
#> 935  37       3         93
#> 936  38       3         93
#> 937  36       2         93
#> 938  37       3         93
#> 939  37       3         93
#> 940  39       3         93
#> 941  37       3         94
#> 942  38       3         94
#> 943  37       3         94
#> 944  40       3         94
#> 945  37       3         94
#> 946  37       3         94
#> 947  35       2         94
#> 948  39       3         94
#> 949  38       3         94
#> 950  36       2         94
#> 951  39       3         95
#> 952  37       3         95
#> 953  38       3         95
#> 954  36       2         95
#> 955  38       3         95
#> 956  35       2         95
#> 957  39       3         95
#> 958  39       3         95
#> 959  40       3         95
#> 960  37       3         95
#> 961  37       3         96
#> 962  39       3         96
#> 963  40       3         96
#> 964  37       3         96
#> 965  37       3         96
#> 966  39       3         96
#> 967  39       3         96
#> 968  39       3         96
#> 969  39       3         96
#> 970  39       3         96
#> 971  36       2         97
#> 972  38       3         97
#> 973  39       3         97
#> 974  38       3         97
#> 975  38       3         97
#> 976  39       3         97
#> 977  40       3         97
#> 978  40       3         97
#> 979  39       3         97
#> 980  40       3         97
#> 981  39       3         98
#> 982  38       3         98
#> 983  39       3         98
#> 984  39       3         98
#> 985  40       3         98
#> 986  39       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  39       3         98
#> 991  39       3         99
#> 992  40       3         99
#> 993  40       3         99
#> 994  40       3         99
#> 995  40       3         99
#> 996  38       3         99
#> 997  40       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100